### 2023-02-20 #5176 @editorialbot
[PRE REVIEW]: ERF: Energy Research and Forecasting
**Submitting author:** @asalmgren (<a href="http://orcid.org/0000-0003-2103-312X">Ann Almgren</a>)
**Repository:** https://github.com/erf-model/ERF
**Branch with paper.md** (empty if default branch): development
**Version:** v1.0.0
