### 2023-02-20 a92735b
update bib file (#970)
