+++ paper/paper.bib
@@ -6,7 +6,7 @@
          volume = {4}, 
          number = {37}, 
          pages = {1370}, 
-         author = {Weiqun Zhang and Ann Almgren and Vince Beckner and John Bell and Johannes Blaschke and Cy Chan and Marcus Day and Brian Friesen and Kevin Gott and Daniel Graves and Max P. Katz and Andrew Myers and Tan Nguyen and Andrew Nonaka and Michele Rosso and Samuel Williams and Michael Zingale}, 
+         author = {Zhang, Weiqun and Almgren, Ann and Beckner, Vince and Bell, John and Blaschke, Johannes and Chan, Cy and Day, Marcus and Friesen, Brian and Gott, Kevin and Graves, Daniel and Katz, Max P. and Myers, Andrew and Nguyen, Tan and Nonaka, Andrew and Rosso, Michele and Williams, Samuel and Zingale, Michael}, 
          title = {AMReX: a framework for block-structured adaptive mesh refinement}, 
          journal = {Journal of Open Source Software} 
 } 
@@ -17,7 +17,7 @@
          volume = {35}, 
          number = {6}, 
          pages = {508-526}, 
-         author = {Zhang W., Myers A., Gott K., Almgren A., Bell J., et al.},
+         author = {Zhang, Weqiun and Myers, Andrew and Gott, Kevin and Almgren, Ann and Bell, John},
          title = {AMReX: Block-structured adaptive mesh refinement for multiphysics applications},
          journal = {The International Journal of High Performance Computing Applications}
 }
@@ -28,7 +28,7 @@
          year = {1982}, 
          volume = {20}, 
          pages = {851-875}, 
-         author = {Mellor G.L. and Yamada T.},
+         author = {Mellor, G.L. and Yamada, T.},
          title = {Development of a turbulence closure model for geophysical fluid problems},
          journal = {Rev. Geophys.}
 }
@@ -47,7 +47,7 @@
 @article{LES:Lilly, 
          year = {1967}, 
          pages = {195-210}, 
-         author = {Lilly D.k.},
+         author = {Lilly, D.K.},
          title = {The Representation of Small-Scale Turbulence in Numerical Simulation Experiments},
          journal = {IBM Form}
 }
@@ -58,7 +58,7 @@
          year = {1963}, 
          volume = {91}, 
          pages = {99-164}, 
-         author = {Nakanishi M. and Niino H.},
+         author = {Nakanishi, M. and Niino, H.},
          title = {General circulation experiments with the primitive equations},
          journal = {Monthly Weather Review}
 }
@@ -68,7 +68,7 @@
          year = {1980}, 
          volume = {18}, 
          pages = {495-527}, 
-         author = {Deardprff J.W.},
+         author = {Deardprff, J.W.},
          title = {Stratocumulus-capped mixed layers derived from a three-dimensional model},
          journal = {Boundary-Layer Meteorol}
 }
@@ -85,12 +85,12 @@
 }
 
 @article{SAMXX:marat, 
-         doi = {10.1175/1520-0469},
+         doi = {10.1175/1520-0469(2003)060<0607:CRMOTA>2.0.CO;2},
          url = {https://doi.org/10.1175/1520-0469(2003)060<0607:CRMOTA>2.0.CO;2},
          year = {2003}, 
          volume = {60}, 
          pages = {607-625}, 
-         author = {Marat F. Khairoutdinov, David A. Randall},
+         author = {Khairoutdinov, Marat F. and Randall, David A.},
          title = {Cloud Resolving Modeling of the ARM Summer 1997 IOP: Model Formulation, Results, Uncertainties, and Sensitivities},
          journal = {Journal of the Atmospheric Sciences}
 }
