### 2023-02-20 4cf638f
fix refs in paper (#971)
