+++ paper/paper.md
@@ -79,7 +79,7 @@ computational framework for the exploration and investigation of different physi
 and numerical strategies, and to provide a characterization of the flow field that impacts the 
 ability of wind turbines to extract wind energy.  The ERF development is part of a broader effort 
 led by the US Department of Energy's Wind Energy Technologies Office (WETO).
-ERF is built on AMReX [@AMReX:JOSS, @AMReX:IJHPCA],
+ERF is built on AMReX [@AMReX:JOSS; @AMReX:IJHPCA],
 a block-structured adaptive mesh refinement software framework that
 provides the underlying performance-portable software infrastructure for block-structured mesh operations. 
 
@@ -118,7 +118,7 @@ conditions can be specified as periodic, inflow/outflow, or time-varying
 values read in from external files in netcdf format generated by the WRF
 Preprocessing System (WPS). The surface boundary condition may either be
 specified as a simple wall or by using Monin-Obukhov similarity theory (MOST)
-[@MOST:monin, @MOST:van] to model the surface layer. The initial data can
+[@MOST:monin; @MOST:van] to model the surface layer. The initial data can
 be read from WPS-generated files, reconstructed from 1-d input sounding
 data, or specified by the user.
 
