+++ paper/paper.bib
@@ -39,7 +39,7 @@
          year = {2009}, 
          volume = {87}, 
          pages = {895-912}, 
-         author = {Nakanishi M. and Niino H.},
+         author = {Nakanishi, M. and Niino, H.},
          title = {Development of an improved turbulence closure model for the atmospheric boundary layer},
          journal = {J. Meteorol. Soc. Jpn. Ser II.}
 }
@@ -68,7 +68,7 @@
          year = {1980}, 
          volume = {18}, 
          pages = {495-527}, 
-         author = {Deardprff, J.W.},
+         author = {Deardorff, J.W.},
          title = {Stratocumulus-capped mixed layers derived from a three-dimensional model},
          journal = {Boundary-Layer Meteorol}
 }
@@ -79,7 +79,7 @@
          year = {2007}, 
          volume = {135}, 
          pages = {2897-2913}, 
-         author = {Klemp J., Skamarock W. and Dudhia J.},
+         author = {Klemp, J. and Skamarock, W. and Dudhia, J.},
          title = {Conservative split-explicit time integration methods for the compressible nonhydrostatic equations},
          journal = {Monthly Weather Review}
 }
@@ -109,7 +109,7 @@
          doi = {10.1002/we.2017},
          url = {https://doi.org/10.1002/we.2017},
          title={A new k-epsilon model consistent with Monin--Obukhov similarity theory},
-         author={van der Laan, M Paul and Kelly, Mark C and S{\o}rensen, Niels N},
+         author={van~der~Laan, M Paul and Kelly, Mark C and S{\o}rensen, Niels N},
          journal={Wind Energy},
          volume={20},
          number={3},
