+++ paper/paper.bib
@@ -6,7 +6,7 @@
          volume = {4}, 
          number = {37}, 
          pages = {1370}, 
-         author = {Zhang, Weiqun and Almgren, Ann and Beckner, Vince and Bell, John and Blaschke, Johannes and Chan, Cy and Day, Marcus and Friesen, Brian and Gott, Kevin and Graves, Daniel and Katz, Max P. and Myers, Andrew and Nguyen, Tan and Nonaka, Andrew and Rosso, Michele and Williams, Samuel and Zingale, Michael}, 
+         author = {Zhang, W. and Almgren, A. and Beckner, V. and Bell, J. and Blaschke, J. and Chan, C. and Day, M. and Friesen, B. and Gott, K. and Graves, D. and Katz, M.P. and Myers, A. and Nguyen, T. and Nonaka, A. and Rosso, M. and Williams, S. and Zingale, M.}, 
          title = {AMReX: a framework for block-structured adaptive mesh refinement}, 
          journal = {Journal of Open Source Software} 
 } 
@@ -17,7 +17,7 @@
          volume = {35}, 
          number = {6}, 
          pages = {508-526}, 
-         author = {Zhang, Weqiun and Myers, Andrew and Gott, Kevin and Almgren, Ann and Bell, John},
+         author = {Zhang, W. and Myers, A. and Gott, K. and Almgren, A. and Bell, J.},
          title = {AMReX: Block-structured adaptive mesh refinement for multiphysics applications},
          journal = {The International Journal of High Performance Computing Applications}
 }
@@ -108,8 +108,8 @@
 @article{MOST:van,
          doi = {10.1002/we.2017},
          url = {https://doi.org/10.1002/we.2017},
-         title={A new k-epsilon model consistent with Monin--Obukhov similarity theory},
-         author={van~der~Laan, M Paul and Kelly, Mark C and S{\o}rensen, Niels N},
+         title={A new k-epsilon model consistent with {M}onin--{O}bukhov similarity theory},
+         author={{van der Laan}, M Paul and Kelly, Mark C and S{\o}rensen, Niels N},
          journal={Wind Energy},
          volume={20},
          number={3},
