### 2023-02-22 c4aed19
update call to WFromOmega to not use u,v
