### 2023-02-22 440bf7e
Fix w from omega (#973)
* spatial_order --> horiz_spatial_order and vert_spatial_order

* update call to WFromOmega to not use u,v

* update regtest files
