### 2023-02-23 4f8e59d
Update line sampling output
