+++ Source/IO/ERF_Write1DProfiles.cpp
@@ -1,6 +1,7 @@
 #include <iomanip>
 
 #include "ERF.H"
+#include "EOS.H"
 
 using namespace amrex;
 
@@ -23,11 +24,15 @@ ERF::write_1D_profiles(Real time)
         Gpu::HostVector<Real> h_avg_uth, h_avg_vth, h_avg_wth, h_avg_thth;
         Gpu::HostVector<Real> h_avg_uu, h_avg_uv, h_avg_uw, h_avg_vv, h_avg_vw, h_avg_ww;
         Gpu::HostVector<Real> h_avg_tau11, h_avg_tau12, h_avg_tau13, h_avg_tau22, h_avg_tau23, h_avg_tau33;
+        Gpu::HostVector<Real> h_avg_k, h_avg_ku, h_avg_kv, h_avg_kw;
+        Gpu::HostVector<Real> h_avg_p, h_avg_pu, h_avg_pv, h_avg_pw;
 
         if (NumDataLogs() > 1) {
             derive_diag_profiles(h_avg_u, h_avg_v, h_avg_w, h_avg_th,
                                  h_avg_uu, h_avg_uv, h_avg_uw, h_avg_vv, h_avg_vw, h_avg_ww,
-                                 h_avg_uth, h_avg_vth, h_avg_wth, h_avg_thth);
+                                 h_avg_uth, h_avg_vth, h_avg_wth, h_avg_thth,
+                                 h_avg_k, h_avg_ku, h_avg_kv, h_avg_kw,
+                                 h_avg_p, h_avg_pu, h_avg_pv, h_avg_pw);
         }
 
         if (NumDataLogs() > 3) {
@@ -59,16 +64,22 @@ ERF::write_1D_profiles(Real time)
                       Real z = (k + 0.5)* dx[2];
                       data_log2 << std::setw(datwidth) << std::setprecision(timeprecision) << time << " "
                                 << std::setw(datwidth) << std::setprecision(datprecision) << z << " "
-                                << h_avg_uu[k]  - h_avg_u[k]*h_avg_u[k]  << " "
-                                << h_avg_uv[k]  - h_avg_u[k]*h_avg_v[k]  << " "
-                                << h_avg_uw[k]  - h_avg_u[k]*h_avg_w[k]  << " "
-                                << h_avg_vv[k]  - h_avg_v[k]*h_avg_v[k]  << " "
-                                << h_avg_vw[k]  - h_avg_v[k]*h_avg_w[k]  << " "
-                                << h_avg_ww[k]  - h_avg_w[k]*h_avg_w[k]  << " "
-                                << h_avg_uth[k] - h_avg_u[k]*h_avg_th[k] << " "
-                                << h_avg_vth[k] - h_avg_v[k]*h_avg_th[k] << " "
-                                << h_avg_wth[k] - h_avg_w[k]*h_avg_th[k] << " "
-                                << h_avg_thth[k]-h_avg_th[k]*h_avg_th[k]
+                                << h_avg_uu[k]   - h_avg_u[k]*h_avg_u[k]  << " "
+                                << h_avg_uv[k]   - h_avg_u[k]*h_avg_v[k]  << " "
+                                << h_avg_uw[k]   - h_avg_u[k]*h_avg_w[k]  << " "
+                                << h_avg_vv[k]   - h_avg_v[k]*h_avg_v[k]  << " "
+                                << h_avg_vw[k]   - h_avg_v[k]*h_avg_w[k]  << " "
[…]
