+++ Source/IO/ReadFromWRFBdy.cpp
@@ -18,6 +18,15 @@ using namespace amrex;
 
 #ifdef ERF_USE_NETCDF
 
+namespace WRFBdyTypes {
+    enum {
+        x_lo,
+        x_hi,
+        y_lo,
+        y_hi
+    };
+}
+
 // Converts UTC time string to a time_t value.
 std::time_t getEpochTime(const std::string& dateTime, const std::string& dateTimeFormat)
 {
@@ -46,7 +55,8 @@ read_from_wrfbdy(std::string nc_bdy_file, const Box& domain,
                  Vector<Vector<FArrayBox>>& bdy_data_xlo,
                  Vector<Vector<FArrayBox>>& bdy_data_xhi,
                  Vector<Vector<FArrayBox>>& bdy_data_ylo,
-                 Vector<Vector<FArrayBox>>& bdy_data_yhi)
+                 Vector<Vector<FArrayBox>>& bdy_data_yhi,
+                 int& width)
 {
     amrex::Print() << "Loading boundary data from NetCDF file " << std::endl;
 
@@ -127,7 +137,6 @@ read_from_wrfbdy(std::string nc_bdy_file, const Box& domain,
     using RARRAY = NDArray<float>;
     amrex::Vector<RARRAY> arrays(nc_var_names.size());
 
-    int ng;
     if (ParallelDescriptor::IOProcessor())
     {
         ReadNetCDFFile(nc_bdy_file, nc_var_names, arrays);
@@ -136,10 +145,12 @@ read_from_wrfbdy(std::string nc_bdy_file, const Box& domain,
         int itimes = static_cast<int>(arrays[0].get_vshape()[0]);
         AMREX_ALWAYS_ASSERT(itimes == ntimes);
 
-        // Width of the boundary region (1 <= ng <= 5)
-        ng = arrays[0].get_vshape()[1];
+        // Width of the boundary region
+        width = arrays[0].get_vshape()[1];
+
+        AMREX_ALWAYS_ASSERT(1 <= width && width <= 5);
     }
-    ParallelDescriptor::Bcast(&ng,1,ioproc);
+    ParallelDescriptor::Bcast(&width,1,ioproc);
 
     // This loops over every variable on every face, so nvars should be 4 * number of "ivartype" below
     for (int iv = 0; iv < nvars; iv++)
@@ -190,15 +201,15 @@ read_from_wrfbdy(std::string nc_bdy_file, const Box& domain,
                 // *******************************************************************************
                 // xlo bdy
                 // *******************************************************************************
-                plo[0] = lo[0]     ; plo[1] = lo[1]; plo[2] = lo[2];
-                phi[0] = lo[0]+ng-1; phi[1] = hi[1]; phi[2] = hi[2];
+                plo[0] = lo[0]        ; plo[1] = lo[1]; plo[2] = lo[2];
+                phi[0] = lo[0]+width-1; phi[1] = hi[1]; phi[2] = hi[2];
                 const Box pbx_xlo(plo, phi);
 
                 Box xlo_plane_no_stag(pbx_xlo);
                 Box xlo_plane_x_stag = pbx_xlo; xlo_plane_x_stag.shiftHalf(0,-1);
                 Box xlo_plane_y_stag = convert(pbx_xlo, {0, 1, 0});
 
-                Box xlo_line(IntVect(lo[0], lo[1], 0), IntVect(lo[0]+ng-1, hi[1], 0));
+                Box xlo_line(IntVect(lo[0], lo[1], 0), IntVect(lo[0]+width-1, hi[1], 0));
 
                 if        (bdyVarType == WRFBdyVars::U) {
                     for (int nt(0); nt < ntimes; ++nt) {
[…]

[…]
