### 2023-02-23 a44381d
Add profile_int output interval control
