### 2023-02-24 0b51435
More fast fixes (#977)
* more fixes for fast integrator

* fix white space
