### 2023-02-24 0cb8504
fixes for gbxo (#979)
