### 2023-02-24 3d86288
fix for moving terrain
