### 2023-02-24 6fa9d5b
fix for GPU compilation
