### 2023-03-01 49f904f
fix for wrfbdy cases (#985)
