### 2023-03-08 68afe3d
Passed basic tests with OMP and diffusion.
