+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -108,7 +108,258 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
         dflux_x = new MultiFab(convert(ba,IntVect(1,0,0)), dm, nvars, 0);
         dflux_y = new MultiFab(convert(ba,IntVect(0,1,0)), dm, nvars, 0);
         dflux_z = new MultiFab(convert(ba,IntVect(0,0,1)), dm, nvars, 0);
-    }
+
+#ifdef _OPENMP
+#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
+#endif
+        for ( MFIter mfi(S_data[IntVar::cons],TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        {
+            // Construct intersection of current tilebox and valid region for updating
+            const Box& valid_bx = grids_to_evolve[mfi.index()];
+            Box bx = mfi.tilebox() & valid_bx;
+
+            // Velocities
+            const Array4<const Real> & u = xvel.array(mfi);
+            const Array4<const Real> & v = yvel.array(mfi);
+            const Array4<const Real> & w = zvel.array(mfi);
+
+            // Contravariant velocity
+            const Array4<Real>& omega_arr = Omega.array(mfi);
+
+            // Map factors
+            const Array4<const Real>& mf_m   = mapfac_m->const_array(mfi);
+            const Array4<const Real>& mf_u   = mapfac_u->const_array(mfi);
+            const Array4<const Real>& mf_v   = mapfac_v->const_array(mfi);
+
+            // Eddy viscosity
+            const Array4<Real const>& mu_turb = l_use_turb ? eddyDiffs->const_array(mfi) : Array4<const Real>{};
+
+            // Terrain metrics
+            const Array4<const Real>& z_nd   = l_use_terrain ? z_phys_nd->const_array(mfi) : Array4<const Real>{};
+            const Array4<const Real>& detJ   = l_use_terrain ?        dJ->const_array(mfi) : Array4<const Real>{};
+
+            //-----------------------------------------
+            // Expansion rate
+            //-----------------------------------------
+            Array4<Real> er_arr;
+            if (expr)
+                er_arr = expr->array(mfi);
+            else
+                er_arr = Array4<Real>{};
+            {
+             BL_PROFILE("slow_rhs_making_er");
+             Box gbx2 = mfi.growntilebox(IntVect(1,1,0));
+
+             if (l_use_terrain) {
+                 // First create Omega using velocity (not momentum)
+                 Box gbxo = surroundingNodes(bx,2); gbxo.grow(IntVect(1,1,0));
+                 amrex::ParallelFor(gbxo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
+                    omega_arr(i,j,k) = (k == 0) ? 0. : OmegaFromW(i,j,k,w(i,j,k),u,v,z_nd,dxInv);
+                });
+
+                 amrex::ParallelFor(gbx2, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
+
+                    Real met_u_h_zeta_hi = Compute_h_zeta_AtIface(i+1, j  , k, dxInv, z_nd);
+                    Real met_u_h_zeta_lo = Compute_h_zeta_AtIface(i  , j  , k, dxInv, z_nd);
+
+                    Real met_v_h_zeta_hi = Compute_h_zeta_AtJface(i  , j+1, k, dxInv, z_nd);
+                    Real met_v_h_zeta_lo = Compute_h_zeta_AtJface(i  , j  , k, dxInv, z_nd);
+
+                    Real Omega_hi = omega_arr(i,j,k+1);
[…]
