### 2023-03-09 9aa3581
Fix most average with multiple procs.
