### 2023-03-09 2a3fd9c
optimization of plane avg.
