### 2023-03-09 987e5a1
Fix double-counting issue on shared faces when averaging
