### 2023-03-09 d6bfa43
Only perform averaging for scalars if data_log is specified
