+++ Source/IO/ERF_WriteScalarProfiles.cpp
@@ -28,7 +28,7 @@ ERF::sum_integrated_quantities(Real time)
         Gpu::HostVector<Real> h_avg_ustar; h_avg_ustar.resize(1);
         Gpu::HostVector<Real> h_avg_tstar; h_avg_tstar.resize(1);
         Gpu::HostVector<Real> h_avg_olen; h_avg_olen.resize(1);
-        if (m_most != nullptr) {
+        if ((m_most != nullptr) && (NumDataLogs() > 0)) {
             Box domain = geom[0].Domain();
             int zdir = 2;
             h_avg_ustar = sumToLine(*m_most->get_u_star(0),0,1,domain,zdir);
