### 2023-03-09 a1676cc
Fix MOST average (#988)
* Fix most average with multiple procs.

* Wrong location. Should apply to plane avg.

* optimization of plane avg.
