+++ Source/BoundaryConditions/MOSTAverage.cpp
@@ -507,6 +507,7 @@ MOSTAverage::compute_plane_averages(int lev)
     //----------------------------------------------------------
     for (int imf(0); imf < m_nvar; ++imf) {
         const amrex::Real denom = 1.0 / (amrex::Real)ncell_plane[imf];
+        amrex::Real d_val_old   = plane_average[imf]*d_fact_old;
 
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
@@ -514,9 +515,16 @@ MOSTAverage::compute_plane_averages(int lev)
         for (amrex::MFIter mfi(*fields[imf], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
             amrex::Box pbx = mfi.tilebox(); pbx.setSmall(2,0); pbx.setBig(2,0);
 
-            auto mf_arr = fields[imf]->const_array(mfi);
+            // Avoid double counting nodal data
+            amrex::Box domain = geom.Domain();
+            amrex::IntVect dom_hi(domain.hiVect());
+            amrex::IndexType ixt = averages[imf]->boxArray().ixType();
+            for (int idim(0); idim < AMREX_SPACEDIM-1; ++idim) {
+                if ( ixt.nodeCentered(idim) && (pbx.bigEnd(idim) < dom_hi[idim]) )
+                    pbx.growHi(idim,-1);
+            }
 
-            amrex::Real d_val_old = plane_average[imf]*d_fact_old;
+            auto mf_arr = fields[imf]->const_array(mfi);
 
             if (m_interp) {
                 const auto plo   = geom.ProbLoArray();
@@ -530,8 +538,8 @@ MOSTAverage::compute_plane_averages(int lev)
                 {
                     amrex::Real interp{0};
                     trilinear_interp_T(x_pos_arr(i,j,k), y_pos_arr(i,j,k), z_pos_arr(i,j,k),
-                                            &interp, mf_arr, z_phys_arr, plo, dxInv, 1);
-                    amrex::Real val = denom * ( interp*d_fact_new + d_val_old );
+                                       &interp, mf_arr, z_phys_arr, plo, dxInv, 1);
+                    amrex::Real val = interp;
                     amrex::Gpu::deviceReduceSum(&plane_avg[imf], val, handler);
                 });
             } else {
@@ -544,11 +552,15 @@ MOSTAverage::compute_plane_averages(int lev)
                     int mk = k_arr(i,j,k);
                     int mj = j_arr ? j_arr(i,j,k) : j;
                     int mi = i_arr ? i_arr(i,j,k) : i;
-                    amrex::Real val = denom * ( mf_arr(mi,mj,mk)*d_fact_new + d_val_old );
+                    amrex::Real val = mf_arr(mi,mj,mk);
                     amrex::Gpu::deviceReduceSum(&plane_avg[imf], val, handler);
                 });
             }
         }
+
+        // Normalize and time average
+        plane_avg[imf] *= denom*d_fact_new;
+        plane_avg[imf] += d_val_old;
     }
 
     // Averages for the tangential velocity magnitude
@@ -557,6 +569,7 @@ MOSTAverage::compute_plane_averages(int lev)
         int imf  = 0;
         int iavg = m_navg - 1;
         const amrex::Real denom = 1.0 / (amrex::Real)ncell_plane[iavg];
+        amrex::Real d_val_old   = plane_average[iavg]*d_fact_old;
 
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
@@ -564,10 +577,10 @@ MOSTAverage::compute_plane_averages(int lev)
[…]
