### 2023-03-10 19d1953
FIX: Calculate correct mixing length for eddy viscosity; remove hard-coded C_k
