+++ Source/DataStruct.H
@@ -167,9 +167,8 @@ struct SolverChoice {
         pp.query("Ce" , Ce);
         pp.query("sigma_k" , sigma_k);
 
-        // Enforce that we input a value for Ck so it isn't the default (0)
-        if  (les_type == LESType::Deardorff) {
-            pp.get("Ck" , Ck);
+        if (les_type == LESType::Deardorff) {
+            pp.query("Ck" , Ck);
         }
 
         // Compute relevant forms of diffusion parameters
