+++ Source/Diffusion/ComputeTurbulentViscosity.cpp
@@ -69,7 +69,8 @@ void ComputeTurbulentViscosityLES (const amrex::MultiFab& Tau11, const amrex::Mu
     //***********************************************************************************
     else if (solverChoice.les_type == LESType::Deardorff)
     {
-      amrex::Real l_C_k = solverChoice.Ck;
+      const amrex::Real l_C_k  = solverChoice.Ck;
+      const amrex::Real abs_g  = solverChoice.gravity;
 
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
@@ -89,8 +90,22 @@ void ComputeTurbulentViscosityLES (const amrex::MultiFab& Tau11, const amrex::Mu
           Real cellVolMsf = 1.0 / (dxInv[0] * mf_u(i,j,0) * dxInv[1] * mf_v(i,j,0) * dxInv[2]);
           Real DeltaMsf   = std::pow(cellVolMsf,1.0/3.0);
 
-          // K = rho * C_k * Delta * KE^(1/2) = C_k * Delta * (rho * RhoKE)^1/2
-          mu_turb(i,j,k,EddyDiff::Mom_h) = l_C_k * DeltaMsf *
+          // calculate stratification-dependent mixing length (Deardorff 1980)
+          Real eps       = std::numeric_limits<Real>::epsilon();
+          Real theta     = cell_data(i,j,k,RhoTheta_comp) / cell_data(i,j,k,Rho_comp);
+          Real dtheta_dz = 0.5*(  cell_data(i,j,k+1,RhoTheta_comp)/cell_data(i,j,k+1,Rho_comp)
+                                - cell_data(i,j,k-1,RhoTheta_comp)/cell_data(i,j,k-1,Rho_comp))*dxInv[2];
+          Real E         = cell_data(i,j,k,RhoKE_comp) / cell_data(i,j,k,Rho_comp);
+          Real strat     = abs_g * dtheta_dz / theta; // stratification
+          Real length;
+          if (strat <= eps) {
+              length = DeltaMsf;
+          } else {
+            length = 0.76 * std::sqrt(E / strat);
+          }
+
+          // K = rho * C_k * l * KE^(1/2) = C_k * Delta * (rho * RhoKE)^1/2
+          mu_turb(i,j,k,EddyDiff::Mom_h) = l_C_k * length *
             std::sqrt(cell_data(i,j,k,RhoKE_comp) * cell_data(i,j,k,Rho_comp));
           mu_turb(i, j, k, EddyDiff::Mom_v) = mu_turb(i, j, k, EddyDiff::Mom_h);
         });
