+++ Source/Diffusion/DiffusionSrcForState_N.cpp
@@ -347,7 +347,6 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain, int n_st
         amrex::Real l_C_k = solverChoice.Ck;
         amrex::ParallelFor(bx,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
         {
-            // Add Buoyancy Source
             Real eps       = std::numeric_limits<Real>::epsilon();
             Real dtheta_dz = 0.5*(cell_prim(i,j,k+1,PrimTheta_comp)-cell_prim(i,j,k-1,PrimTheta_comp))*dz_inv;
             Real E         = cell_prim(i,j,k,PrimKE_comp);
@@ -358,8 +357,16 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain, int n_st
             } else {
               length = 0.76 * std::sqrt(E / strat);
             }
-            Real KH = l_C_k * (1.+2.*length/l_Delta) * std::sqrt(E);
-            cell_rhs(i,j,k,qty_index) += cell_data(i,j,k,Rho_comp) * grav_gpu[2] * KH * dtheta_dz;
+
+            // From eddy viscosity mu_turb = rho * C_k * l * KE^(1/2), the
+            // eddy diffusivity for heat, KH = (1 + 2*l/delta) * mu_turb
+            Real KH = cell_data(i,j,k,Rho_comp) * l_C_k * (1.+2.*length/l_Delta) * std::sqrt(E);
+
+            // Add Buoyancy Source
+            // where the SGS buoyancy flux tau_{theta,i} = -KH * dtheta/dx_i,
+            // such that for dtheta/dz < 0, there is a positive (upward) heat flux;
+            // the buoyant production is then g/theta_0 * tau_{theta,w}
+            cell_rhs(i,j,k,qty_index) += grav_gpu[2] * l_inv_theta0 * -(KH * dtheta_dz);
 
             // TKE production
             cell_rhs(i,j,k,qty_index) += 2.0*mu_turb(i,j,k,EddyDiff::Mom_h) * SmnSmn_a(i,j,k);
