### 2023-03-10 e7e6b5e
Set default sigma_k to 0.5 to match downgradient diffusion assumption
from Moeng 1984:
resolved u_i' (e' + p'/rho_0) = 2 K_M de'/dx_i
noting that her diffusion source term is negative

our diffusivity of kinetic energy is K_M (diffusivity of momentum, or eddy
diffusivity) divided by sigma_k, i.e., K_k = K_M/0.5 = 2 K_M
