### 2023-03-10 de9bcf7
Make DiffusionSrcForState match with and without terrain
