+++ Source/IO/ERF_Write1DProfiles.cpp
@@ -23,9 +23,10 @@ ERF::write_1D_profiles(Real time)
         Gpu::HostVector<Real> h_avg_u, h_avg_v, h_avg_w, h_avg_th;
         Gpu::HostVector<Real> h_avg_uth, h_avg_vth, h_avg_wth, h_avg_thth;
         Gpu::HostVector<Real> h_avg_uu, h_avg_uv, h_avg_uw, h_avg_vv, h_avg_vw, h_avg_ww;
-        Gpu::HostVector<Real> h_avg_tau11, h_avg_tau12, h_avg_tau13, h_avg_tau22, h_avg_tau23, h_avg_tau33;
         Gpu::HostVector<Real> h_avg_k, h_avg_ku, h_avg_kv, h_avg_kw;
         Gpu::HostVector<Real> h_avg_p, h_avg_pu, h_avg_pv, h_avg_pw;
+        Gpu::HostVector<Real> h_avg_tau11, h_avg_tau12, h_avg_tau13, h_avg_tau22, h_avg_tau23, h_avg_tau33;
+        Gpu::HostVector<Real> h_avg_sgshfx, h_avg_sgsdiss; // only output tau_{theta,w} and epsilon for now
 
         if (NumDataLogs() > 1) {
             derive_diag_profiles(h_avg_u, h_avg_v, h_avg_w, h_avg_th,
@@ -37,7 +38,9 @@ ERF::write_1D_profiles(Real time)
 
         if (NumDataLogs() > 3) {
             derive_stress_profiles(h_avg_tau11, h_avg_tau12, h_avg_tau13,
-                                   h_avg_tau22, h_avg_tau23, h_avg_tau33);
+                                   h_avg_tau22, h_avg_tau23, h_avg_tau33,
+                                   h_avg_sgshfx,
+                                   h_avg_sgsdiss);
         }
 
         int hu_size =  h_avg_u.size();
@@ -96,7 +99,8 @@ ERF::write_1D_profiles(Real time)
                       data_log3 << std::setw(datwidth) << std::setprecision(timeprecision) << time << " "
                                 << std::setw(datwidth) << std::setprecision(datprecision) << z << " "
                                 << h_avg_tau11[k] << " " << h_avg_tau12[k] << " " << h_avg_tau13[k] << " "
-                                << h_avg_tau22[k] << " " << h_avg_tau23[k] << " " << h_avg_tau33[k]
+                                << h_avg_tau22[k] << " " << h_avg_tau23[k] << " " << h_avg_tau33[k] << " "
+                                << h_avg_sgshfx[k] << " " << h_avg_sgsdiss[k]
                                 << std::endl;
                   } // loop over z
                 } // if good
@@ -232,12 +236,13 @@ ERF::derive_diag_profiles(Gpu::HostVector<Real>& h_avg_u   , Gpu::HostVector<Rea
 void
 ERF::derive_stress_profiles(Gpu::HostVector<Real>& h_avg_tau11, Gpu::HostVector<Real>& h_avg_tau12,
                             Gpu::HostVector<Real>& h_avg_tau13, Gpu::HostVector<Real>& h_avg_tau22,
-                            Gpu::HostVector<Real>& h_avg_tau23, Gpu::HostVector<Real>& h_avg_tau33)
+                            Gpu::HostVector<Real>& h_avg_tau23, Gpu::HostVector<Real>& h_avg_tau33,
[…]
