### 2023-03-11 8a79865
Add additional scalar profile output for completeness
