+++ Source/IO/ERF_Write1DProfiles.cpp
@@ -20,7 +20,8 @@ ERF::write_1D_profiles(Real time)
     if (verbose > 0 && NumDataLogs() > 1)
     {
         // Define the 1d arrays we will need
-        Gpu::HostVector<Real> h_avg_u, h_avg_v, h_avg_w, h_avg_th;
+        Gpu::HostVector<Real> h_avg_u, h_avg_v, h_avg_w;
+        Gpu::HostVector<Real> h_avg_rho, h_avg_th, h_avg_ksgs;
         Gpu::HostVector<Real> h_avg_uth, h_avg_vth, h_avg_wth, h_avg_thth;
         Gpu::HostVector<Real> h_avg_uu, h_avg_uv, h_avg_uw, h_avg_vv, h_avg_vw, h_avg_ww;
         Gpu::HostVector<Real> h_avg_k, h_avg_ku, h_avg_kv, h_avg_kw;
@@ -29,7 +30,8 @@ ERF::write_1D_profiles(Real time)
         Gpu::HostVector<Real> h_avg_sgshfx, h_avg_sgsdiss; // only output tau_{theta,w} and epsilon for now
 
         if (NumDataLogs() > 1) {
-            derive_diag_profiles(h_avg_u, h_avg_v, h_avg_w, h_avg_th,
+            derive_diag_profiles(h_avg_u, h_avg_v, h_avg_w,
+                                 h_avg_rho, h_avg_th, h_avg_ksgs,
                                  h_avg_uu, h_avg_uv, h_avg_uw, h_avg_vv, h_avg_vw, h_avg_ww,
                                  h_avg_uth, h_avg_vth, h_avg_wth, h_avg_thth,
                                  h_avg_k, h_avg_ku, h_avg_kv, h_avg_kw,
@@ -55,7 +57,8 @@ ERF::write_1D_profiles(Real time)
                       Real z = (k + 0.5)* dx[2];
                       data_log1 << std::setw(datwidth) << std::setprecision(timeprecision) << time << " "
                                 << std::setw(datwidth) << std::setprecision(datprecision) << z << " "
-                                << h_avg_u[k]  << " " << h_avg_v[k] << " " << h_avg_w[k] << " " << h_avg_th[k]
+                                << h_avg_u[k]  << " " << h_avg_v[k] << " " << h_avg_w[k] << " "
+                                << h_avg_rho[k] << " " << h_avg_th[k] << " " << h_avg_ksgs[k]
                                 << std::endl;
                   } // loop over z
                 } // if good
@@ -110,8 +113,8 @@ ERF::write_1D_profiles(Real time)
 }
 
 void
-ERF::derive_diag_profiles(Gpu::HostVector<Real>& h_avg_u   , Gpu::HostVector<Real>& h_avg_v  ,
-                          Gpu::HostVector<Real>& h_avg_w   , Gpu::HostVector<Real>& h_avg_th ,
+ERF::derive_diag_profiles(Gpu::HostVector<Real>& h_avg_u   , Gpu::HostVector<Real>& h_avg_v  , Gpu::HostVector<Real>& h_avg_w,
+                          Gpu::HostVector<Real>& h_avg_rho , Gpu::HostVector<Real>& h_avg_th , Gpu::HostVector<Real>& h_avg_ksgs,
[…]
