### 2023-03-12 f24a63a
Fix sign issue with const g (abs val) vs vector g
