### 2023-03-12 73ce1f7
FIX: Don't recalc eddy viscosity part of KH
