### 2023-03-12 704d9da
Fix diffusion w/ terrain (#990)
* Order of operations must be changed to avoid an error with tau_{33} where linear comb terms are multiplied by h_{zeta} and mu twice.

* Remove debug print statement.

---------

Co-authored-by: asalmgren <asalmgren@lbl.gov>
