### 2023-03-12 635ae2d
fixes crash but not regtest
