### 2023-03-13 094c36c
more fixes for OpenMP (#995)
