+++ Source/TimeIntegration/ERF_make_fast_coeffs.cpp
@@ -72,7 +72,7 @@ void make_fast_coeffs (int /*level*/,
         const Array4<const Real>& r0_ca       = r0->const_array(mfi);
         const Array4<const Real>& pi0_ca      = pi0->const_array(mfi); const Array4<const Real>& pi_stage_ca = pi_stage.const_array(mfi);
 
-        gam_fab.resize(bx);
+        gam_fab.resize(surroundingNodes(bx,2),1);
         Elixir gEli = gam_fab.elixir();
 
         auto const& coeffA_a  = coeff_A_mf.array(mfi);
