### 2023-03-13 f369f8e
Update mu_turb at every RK stage, make mixing length calc consistent btwn mu_turb and KH
