### 2023-03-13 10ebd22
add the option to fix the random seed
