### 2023-03-14 8797e3c
Only add ghost cells if using num diff.
