+++ Source/Diffusion/NumericalDiffusion.cpp
@@ -7,11 +7,37 @@ NumericalDiffusion (const Box& bx,
                     const int  n_start,
                     const int  n_end,
                     const Real dt,
+                    const SolverChoice& solverChoice,
                     const Array4<const Real>& data,
                     const Array4<      Real>& rhs,
-                    const SolverChoice& solverChoice)
+                    const Array4<const Real>& mf_x,
+                    const Array4<const Real>& mf_y,
+                    const bool avg_mf_x_y,
+                    const bool avg_mf_y_x)
 {
     BL_PROFILE_VAR("NumericalDiffusion()",NumericalDiffusion);
+
+    // Average map factors to correct locations
+    Box planebx(bx); planebx.setSmall(2,0); planebx.setBig(2,0);
+    FArrayBox mf_x_bar; mf_x_bar.resize(planebx,1);
+    FArrayBox mf_y_bar; mf_y_bar.resize(planebx,1);
+    Elixir mfx_eli = mf_x_bar.elixir();
+    Elixir mfy_eli = mf_y_bar.elixir();
+    const Array4<Real>& mfx_arr  = mf_x_bar.array();
+    const Array4<Real>& mfy_arr  = mf_y_bar.array();
+    amrex::ParallelFor(planebx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+    {
+        if (avg_mf_x_y) {
+            mfx_arr(i,j,k) = 0.5 * ( mf_x(i,j-1,k) + mf_x(i,j,k) );
+        } else {
+            mfx_arr(i,j,k) = mf_x(i,j,k);
+        }
+        if (avg_mf_y_x) {
+            mfy_arr(i,j,k) = 0.5 * ( mf_y(i-1,j,k) + mf_y(i,j,k) );
+        } else {
+            mfy_arr(i,j,k) = mf_y(i,j,k);
+        }
+    });
 
     // Set indices for components
     const int ncomp = n_end - n_start + 1;
@@ -39,7 +65,7 @@ NumericalDiffusion (const Box& bx,
                        - 5. * (data(i,j+2,k,n) - data(i,j-1,k,n))
                             + (data(i,j+3,k,n) - data(i,j-2,k,n));
         if ( (yflux_hi * (data(i,j+1,k,n) - data(i,j,k,n)) ) > 0.) yflux_hi = 0.;
-        rhs(i,j,k,n) += coeff6 * (xflux_hi - xflux_lo
-                                + yflux_hi - yflux_lo);
+        rhs(i,j,k,n) += coeff6 * ( (xflux_hi - xflux_lo) * mfx_arr(i,j,0)
+                                 + (yflux_hi - yflux_lo) * mfy_arr(i,j,0) );
     });
 }
