### 2023-03-14 73b4149
fix for omp (#1002)
