+++ Exec/CouetteFlow/prob.cpp
@@ -44,28 +44,31 @@ erf_init_rayleigh(amrex::Vector<Real>& /*tau*/,
 
 void
 init_custom_prob(
-  const amrex::Box& bx,
-  Array4<Real> const& state,
-  Array4<Real> const& x_vel,
-  Array4<Real> const& y_vel,
-  Array4<Real> const& z_vel,
-  Array4<Real> const&,
-  Array4<Real> const&,
-  Array4<Real const> const&,
-  Array4<Real const> const&,
+    const Box& bx,
+    const Box& xbx,
+    const Box& ybx,
+    const Box& zbx,
+    Array4<Real> const& state,
+    Array4<Real> const& x_vel,
+    Array4<Real> const& y_vel,
+    Array4<Real> const& z_vel,
+    Array4<Real> const&,
+    Array4<Real> const&,
+    Array4<Real const> const&,
+    Array4<Real const> const&,
 #if defined(ERF_USE_MOISTURE)
-  Array4<Real      > const&,
-  Array4<Real      > const&,
-  Array4<Real      > const&,
+    Array4<Real      > const&,
+    Array4<Real      > const&,
+    Array4<Real      > const&,
 #elif defined(ERF_USE_WARM_NO_PRECIP)
-  Array4<Real      > const&,
-  Array4<Real      > const&,
+    Array4<Real      > const&,
+    Array4<Real      > const&,
 #endif
-  amrex::GeometryData const& geomdata,
-  Array4<Real const> const& /*mf_m*/,
-  Array4<Real const> const& /*mf_u*/,
-  Array4<Real const> const& /*mf_v*/,
-  const SolverChoice&)
+    amrex::GeometryData const& geomdata,
+    Array4<Real const> const& /*mf_m*/,
+    Array4<Real const> const& /*mf_u*/,
+    Array4<Real const> const& /*mf_v*/,
+    const SolverChoice&)
 {
     amrex::ParallelFor(bx, [=, parms=parms] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
 
@@ -88,7 +91,6 @@ init_custom_prob(
 
   });
 
-  const amrex::Box& xbx = amrex::surroundingNodes(bx,0);
   amrex::ParallelFor(xbx, [=, parms=parms] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
     const auto prob_hi  = geomdata.ProbHi();
     const auto dx       = geomdata.CellSize();
@@ -96,7 +98,6 @@ init_custom_prob(
     x_vel(i, j, k) = parms.u_0 * z / prob_hi[2];
   });
 
-  const amrex::Box& ybx = amrex::surroundingNodes(bx,1);
   amrex::ParallelFor(ybx, [=, parms=parms] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
     const auto prob_hi  = geomdata.ProbHi();
     const auto dx       = geomdata.CellSize();
@@ -104,7 +105,6 @@ init_custom_prob(
     y_vel(i, j, k) = parms.v_0 * z / prob_hi[2];
   });
 
-  const amrex::Box& zbx = amrex::surroundingNodes(bx,2);
   amrex::ParallelFor(zbx, [=, parms=parms] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
     z_vel(i, j, k) = parms.w_0;
   });
