+++ Source/Diffusion/DiffusionSrcForState_N.cpp
@@ -369,7 +369,7 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain, int n_st
 
             // From eddy viscosity (or eddy diffusivity for momentum)
             //   mu_turb = rho * C_k * l * KE^(1/2)
-            // and the eddy diffusivity for heat is
+            // and the eddy diffusivity for heat [kg m^-1 s^-1] is
             //   KH = (1 + 2*l/delta) * mu_turb
             // Note: mu_turb is fixed for all 3 RK stages, so recomputing the
             // eddy viscosity part of KH here would be inconsistent (since
@@ -382,7 +382,7 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain, int n_st
             // the TKE buoyancy production is then B = g/theta_0 * tau_{theta,w}
             hfx_x(i,j,k) = 0.0;
             hfx_y(i,j,k) = 0.0;
-            hfx_z(i,j,k) = -KH * dtheta_dz; // w'theta'  
+            hfx_z(i,j,k) = -KH * dtheta_dz; // (rho*w)'theta' [kg m^-2 s^-1 K]
             cell_rhs(i,j,k,qty_index) += l_abs_g * l_inv_theta0 * hfx_z(i,j,k);
 
             // TKE shear production
