### 2023-03-15 ef0f262
Bound KE to ensure it is positive (>= eps)
- AMR-Wind uses a clip value of 1e-8 (in equation_systems/tke/tke_ops.H)
- SOWFA uses a lower bound of "small" (i.e., eps, from SOWFA LESModel property
  kMin; small is defined as floatScalarSmall in OpenFOAM/primitives/Scalar/...)
