+++ Source/ERF.cpp
@@ -402,8 +402,10 @@ ERF::InitData ()
         for (int lev = 0; lev <= finest_level; lev++)
             init_only(lev, time);
 
-        // For now we initialize rho_KE to 0
-        Real RhoKE_0 = 0.0;
+        // We initialize rho_KE to be a small positive number by default;
+        // epsilon is also the lower bound used when calculating the Deardorff
+        // mixing length
+        Real RhoKE_0 = std::numeric_limits<Real>::epsilon();
         ParmParse pp(pp_prefix);
         pp.query("RhoKE_0", RhoKE_0);
         int lb = std::max(finest_level-1,0);
