### 2023-03-15 6bb4acc
Make bounding on SFS KE consistent across mu_turb and diffusion src updates
