### 2023-03-15 2a7eff5
Update default value of RhoKE_0
If using a SFS model that calculates mu_turb ~ sqrt(k) and the SFS KE is
too small, the MOST boundary condition will run into issues because the
extrapolated boundary value is inversely proportional to the eddy
diffusivity and eddy diffusivity is proportional to mu_turb. The boundary
values blowing up will eventually lead to an erronenous arithmetic error.

For reference, SOWFA users typically define k=0.1; AMR-Wind has a hard-coded
default value of 1.0.
