### 2023-03-15 6888b85
Make diffusion src for state consistent w/w/o terrain
