### 2023-03-16 0e24594
removed statement causing buffer overflow
