### 2023-03-16 9bf49a3
Use bounded e instead of cell_data(RhoKE) in Deardorff mu_turb calc
