### 2023-03-17 6015bd8
Clip entire RhoKE field; revert to once/timestep mu_turb update
- before: calculate mixing length based on clipped local e, but the e field
  remains untouched -- this could create issues for the TKE transport/diffusion
- confirmed that WRF updates the "exchange coefficients" KM, KH only in the
  first RK stage
