+++ Source/Diffusion/DiffusionSrcForState_T.cpp
@@ -535,7 +535,7 @@ DiffusionSrcForState_T (const amrex::Box& bx, const amrex::Box& domain, int n_st
             Real eps       = std::numeric_limits<Real>::epsilon();
             Real dtheta_dz = 0.5*(cell_prim(i,j,k+1,PrimTheta_comp)-cell_prim(i,j,k-1,PrimTheta_comp))*dz_inv;
             dtheta_dz      /= met_h_zeta;
-            Real E         = amrex::max(cell_prim(i,j,k,PrimKE_comp), eps);
+            Real E         = cell_prim(i,j,k,PrimKE_comp);
             Real strat     = l_abs_g * dtheta_dz * l_inv_theta0; // ==N^2 under stable conditions
             Real length;
             if (strat <= eps) {
@@ -583,11 +583,8 @@ DiffusionSrcForState_T (const amrex::Box& bx, const amrex::Box& domain, int n_st
                 Ce = l_C_e_wall;
             else
                 Ce = 1.9*l_C_k + Ce_lcoeff*length / DeltaMsf;
-            if (std::abs(E) > 0.) {
-                diss(i,j,k) = cell_data(i,j,k,Rho_comp) * Ce *
-                    std::pow(E,1.5) / length;
-                cell_rhs(i,j,k,qty_index) -= diss(i,j,k);
-            }
+            diss(i,j,k) = cell_data(i,j,k,Rho_comp) * Ce * std::pow(E,1.5) / length;
+            cell_rhs(i,j,k,qty_index) -= diss(i,j,k);
         });
     }
 
