### 2023-03-17 5cb37b7
Fix copy of s_ij to tau_ij without terrain.
