### 2023-03-18 d7cc041
Flip sign of deviatoric stress tensor tau_ij...
To be consistent with the SGS Reynolds stress definition (e.g., Moeng 1984):
  tau_ij = R_ij - R_kk delta_ij/3
where
  R_ij = (u_i' u_j') + (u_i' (u_j)_r) + ((u_i)_r u_j'),
the RHS includes filtered quantities, (.)_r is the resolved-scale field, and
delta_ij is the kronecker delta.

The RHS diffusion src terms have been updated accordingly

Verified that the following solutions are identical (fcompare gives all 0):
* Deardorff w/ no-slip lower wall
* Deardorff w/ MOST lower BC
