+++ Source/Diffusion/ComputeStress_N.cpp
@@ -13,21 +13,21 @@ ComputeStressConsVisc_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz, Real mu_e
     // Cell centered strains
     amrex::ParallelFor(bxcc, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
     {
-        tau11(i,j,k) = mu_eff * ( tau11(i,j,k) - OneThird*er_arr(i,j,k) );
-        tau22(i,j,k) = mu_eff * ( tau22(i,j,k) - OneThird*er_arr(i,j,k) );
-        tau33(i,j,k) = mu_eff * ( tau33(i,j,k) - OneThird*er_arr(i,j,k) );
+        tau11(i,j,k) = -mu_eff * ( tau11(i,j,k) - OneThird*er_arr(i,j,k) );
+        tau22(i,j,k) = -mu_eff * ( tau22(i,j,k) - OneThird*er_arr(i,j,k) );
+        tau33(i,j,k) = -mu_eff * ( tau33(i,j,k) - OneThird*er_arr(i,j,k) );
     });
 
     // Off-diagonal strains
     amrex::ParallelFor(tbxxy,tbxxz,tbxyz,
     [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
-        tau12(i,j,k) *= mu_eff;
+        tau12(i,j,k) *= -mu_eff;
     },
     [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
-        tau13(i,j,k) *= mu_eff;
+        tau13(i,j,k) *= -mu_eff;
     },
     [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
-        tau23(i,j,k) *= mu_eff;
+        tau23(i,j,k) *= -mu_eff;
     });
 
 }
@@ -47,9 +47,9 @@ ComputeStressVarVisc_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz, Real mu_ef
         Real mu_11 = mu_eff + 2.0 * mu_turb(i, j, k, EddyDiff::Mom_h);
         Real mu_22 = mu_11;
         Real mu_33 = mu_eff + 2.0 * mu_turb(i, j, k, EddyDiff::Mom_v);
-        tau11(i,j,k) = mu_11 * ( tau11(i,j,k) - OneThird*er_arr(i,j,k) );
-        tau22(i,j,k) = mu_22 * ( tau22(i,j,k) - OneThird*er_arr(i,j,k) );
-        tau33(i,j,k) = mu_33 * ( tau33(i,j,k) - OneThird*er_arr(i,j,k) );
+        tau11(i,j,k) = -mu_11 * ( tau11(i,j,k) - OneThird*er_arr(i,j,k) );
+        tau22(i,j,k) = -mu_22 * ( tau22(i,j,k) - OneThird*er_arr(i,j,k) );
+        tau33(i,j,k) = -mu_33 * ( tau33(i,j,k) - OneThird*er_arr(i,j,k) );
     });
 
     // Off-diagonal strains
@@ -58,18 +58,18 @@ ComputeStressVarVisc_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz, Real mu_ef
         Real mu_bar = 0.25*( mu_turb(i-1, j  , k, EddyDiff::Mom_h) + mu_turb(i, j  , k, EddyDiff::Mom_h)
                            + mu_turb(i-1, j-1, k, EddyDiff::Mom_h) + mu_turb(i, j-1, k, EddyDiff::Mom_h) );
         Real mu_12  = mu_eff + 2.0*mu_bar;
[…]
