+++ Source/Diffusion/DiffusionSrcForMom_T.cpp
@@ -36,7 +36,7 @@ DiffusionSrcForMom_T (const Box& bxx, const Box& bxy , const Box& bxz,
                                 + (tau13(i  , j  , k+1) - tau13(i  , j  ,k  )) * dzinv );     // Contribution to x-mom eqn from diffusive flux in z-dir;
             diffContrib      /= 0.5*(detJ(i,j,k) + detJ(i-1,j,k));
             diffContrib      *= 0.5 * (cons(i,j,k,Rho_comp) + cons(i-1,j,k,Rho_comp))  / solverChoice.rho0_trans;
-            rho_u_rhs(i,j,k) += diffContrib;
+            rho_u_rhs(i,j,k) -= diffContrib;
         },
         [=] AMREX_GPU_DEVICE (int i, int j, int k)
         {
@@ -47,7 +47,7 @@ DiffusionSrcForMom_T (const Box& bxx, const Box& bxy , const Box& bxz,
                                 + (tau23(i  , j  , k+1) - tau23(i  , j  , k  )) * dzinv );     // Contribution to y-mom eqn from diffusive flux in z-dir;
             diffContrib      /= 0.5*(detJ(i,j,k) + detJ(i,j-1,k));
             diffContrib      *= 0.5 * (cons(i,j,k,Rho_comp) + cons(i,j-1,k,Rho_comp))  / solverChoice.rho0_trans;
-            rho_v_rhs(i,j,k) += diffContrib;
+            rho_v_rhs(i,j,k) -= diffContrib;
         },
         [=] AMREX_GPU_DEVICE (int i, int j, int k)
         {
@@ -58,7 +58,7 @@ DiffusionSrcForMom_T (const Box& bxx, const Box& bxy , const Box& bxz,
                                 + (tau33(i  , j  , k  ) - tau33(i  , j  , k-1)) * dzinv );     // Contribution to z-mom eqn from diffusive flux in z-dir;
             diffContrib      /= 0.5*(detJ(i,j,k) + detJ(i,j,k-1));
             diffContrib      *= 0.5 * (cons(i,j,k,Rho_comp) + cons(i,j,k-1,Rho_comp))  / solverChoice.rho0_trans;
-            rho_w_rhs(i,j,k) += diffContrib;
+            rho_w_rhs(i,j,k) -= diffContrib;
         });
 
     } else {
@@ -71,7 +71,7 @@ DiffusionSrcForMom_T (const Box& bxx, const Box& bxy , const Box& bxz,
                                 + (tau12(i  , j+1, k  ) - tau12(i  , j  ,k  )) * dyinv * mf   // Contribution to x-mom eqn from diffusive flux in y-dir
                                 + (tau13(i  , j  , k+1) - tau13(i  , j  ,k  )) * dzinv );     // Contribution to x-mom eqn from diffusive flux in z-dir;
             diffContrib      /= 0.5*(detJ(i,j,k) + detJ(i-1,j,k));
-            rho_u_rhs(i,j,k) += diffContrib;
+            rho_u_rhs(i,j,k) -= diffContrib;
         },
         [=] AMREX_GPU_DEVICE (int i, int j, int k)
         {
@@ -81,7 +81,7 @@ DiffusionSrcForMom_T (const Box& bxx, const Box& bxy , const Box& bxz,
                                 + (tau22(i  , j  , k  ) - tau22(i  , j-1, k  )) * dyinv * mf   // Contribution to y-mom eqn from diffusive flux in y-dir
                                 + (tau23(i  , j  , k+1) - tau23(i  , j  , k  )) * dzinv );     // Contribution to y-mom eqn from diffusive flux in z-dir;
[…]

[…]
