### 2023-03-18 210547b
Update docs to reflect change in sign convention for tau (deviatoric sfs stress tensor)
