+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -137,19 +137,45 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
             const Array4<const Real>& z_nd   = l_use_terrain ? z_phys_nd->const_array(mfi) : Array4<const Real>{};
             const Array4<const Real>& detJ   = l_use_terrain ?        dJ->const_array(mfi) : Array4<const Real>{};
 
+            //-------------------------------------------------------------------------------
+            // NOTE: Tile boxes with terrain are not intuitive. The linear combination of
+            //       stress terms requires care. Create a tile box that intersects the
+            //       valid box, then grow the box in x/y. Compute the strain on the local
+            //       FAB over this grown tile box. Compute the stress over the tile box,
+            //       except tau_ii which still needs the halo cells. Finally, write from
+            //       the local FAB to the Tau MF but only on the tile box.
+            //-------------------------------------------------------------------------------
+
+            //-------------------------------------------------------------------------------
+            // TODO: Avoid recomputing strain on the first RK stage. One could populate
+            //       the FABs with tau_ij, compute stress, and then write to tau_ij. The
+            //       problem with this approach is you will over-write the needed halo layer
+            //       needed by subsequent tile boxes (particularly S_ii becomes Tau_ii).
+            //-------------------------------------------------------------------------------
+
+            // Strain/Stress tile boxes
+            Box vbx   = valid_bx;
+            Box bxcc  = mfi.tilebox() & valid_bx;
+            Box tbxxy = mfi.tilebox(IntVect(1,1,0)) & vbx.convert(IntVect(1,1,0));
+            Box tbxxz = mfi.tilebox(IntVect(1,0,1)) & vbx.convert(IntVect(1,0,1));
+            Box tbxyz = mfi.tilebox(IntVect(0,1,1)) & vbx.convert(IntVect(0,1,1));
+            // We need a halo cells for terrain
+             bxcc.grow(IntVect(1,1,0));
+            tbxxy.grow(IntVect(1,1,0));
+            tbxxz.grow(IntVect(1,1,0));
+            tbxyz.grow(IntVect(1,1,0));
+
             //-----------------------------------------
             // Expansion rate
             //-----------------------------------------
             Array4<Real> er_arr = expr->array(mfi);
-
             {
              BL_PROFILE("slow_rhs_making_er");
-             Box gvbx = valid_bx; gvbx.grow(IntVect(1,1,0));
-             Box gbx2 = mfi.growntilebox(IntVect(1,1,0)) & gvbx;
+             Box gbx2 = bxcc;
 
              if (l_use_terrain) {
                  // First create Omega using velocity (not momentum)
-                 Box gbxo = surroundingNodes(bx & valid_bx,2); gbxo.grow(IntVect(1,1,0));
+                 Box gbxo = surroundingNodes(gbx2,2);
                  amrex::ParallelFor(gbxo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
                     omega_arr(i,j,k) = (k == 0) ? 0. : OmegaFromW(i,j,k,w(i,j,k),u,v,z_nd,dxInv);
                  });
@@ -187,112 +213,55 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
             //-----------------------------------------
             // Strain tensor
             //-----------------------------------------
+            // Temporary storage for tiling/OMP
+            FArrayBox S11,S22,S33;
[…]
