### 2023-03-19 erf-model/ERF#1010 @pratikvn
JOSS reviewer 1 feedback.
I am reviewer for your [JOSS paper](https://github.com/openjournals/joss-reviews/issues/5202) on ERF. The repository and its aims are very interesting and I look forward to the review.

To make the review easier for me and for you, I am creating this main issue to keep track of my review regarding general suggestions and issues. To keep clutter to a minimum, I will try to keep all my questions in here as well.

If I face any compilation/running issues, I will create new issues for those and link them back here so that you can also keep track of the progress.

Please let me know if you would prefer another way instead.

## Repository related suggestions:

+ [x] I see that you already have a `license.txt` file, which is great but due to the way github works, I think it would be useful to move that to a file named LICENSE or LICENSE.md, which allows the github webpage to more promimently display your license. [See their documentation for more information](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository)

## Paper suggestions

+ [x] The paper itself is well-written. One suggestion I would have is to split up the large Summary section into a smaller section with a high level overview and a Features section that explains the different features, models and options that the software provides.

## Some general questions:

+ [x] Could you please explain the rationale for the author list  ? I see that some authors who are in the [github contributors are not part of the paper](https://github.com/erf-model/ERF/graphs/contributors).

## Documentation suggestions

+ [x] I see that ERF depends on AMREX, RRTMGP and on YAKL. While AMREX has been mentioned as a dependency, the other two have not. I think this should be clarified in the installation section in the documentation.
+ [x] The Cdash link for nightly test seems to not exist now ? https://my.cdash.org/index.php?project=ERF
+ [x] I really like the `Build/` directory with sample cmake commands. But I did not find this mentioned in the documentation. I think this would be useful to have in the docs.
+ [x] API documentation
+ [ ] Some basic examples and how to run them

## Compilation issues

+ [x] Resolve #1015

## Run and test issues

+ [x] Resolve #1014
