### 2023-03-19 erf-model/ERF#1010 @asalmgren
Thank you for your questions/comments/suggestions!  Here I address two of the comments and we will respond to the others directly once we have addressed them

1.  We have moved license.txt to LICENSE.md
2. Re the author list -- the github contributor history is longer than it should be due to how this code was initially created on github -- it was started by copying, then gutting, an existing code, which had itself started that way.   The github contributors who are not on the author list did not contribute to the ERF code itself.
