### 2023-03-20 6815460
Fix the GPU test by moving elixir outside of an if use terrain statement. To avoid allocating more temporary memory without terrain, reformulated the precomputation.
