+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -165,22 +165,48 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
             tbxxz.grow(IntVect(1,1,0));
             tbxyz.grow(IntVect(1,1,0));
 
-            //-----------------------------------------
             // Expansion rate
-            //-----------------------------------------
             Array4<Real> er_arr = expr->array(mfi);
-            {
-             BL_PROFILE("slow_rhs_making_er");
-             Box gbx2 = bxcc;
 
-             if (l_use_terrain) {
-                 // First create Omega using velocity (not momentum)
-                 Box gbxo = surroundingNodes(gbx2,2);
-                 amrex::ParallelFor(gbxo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
+            // Temporary storage for tiling/OMP
+            FArrayBox S11,S22,S33;
+            FArrayBox S12,S13,S23;
+            S11.resize(bxcc,1);  S22.resize(bxcc,1);  S33.resize(bxcc,1);
+            S12.resize(tbxxy,1); S13.resize(tbxxz,1); S23.resize(tbxyz,1);
+            Elixir S11_eli   = S11.elixir(); Elixir S22_eli   = S22.elixir(); Elixir S33_eli   = S33.elixir();
+            Elixir S12_eli   = S12.elixir(); Elixir S13_eli   = S13.elixir(); Elixir S23_eli   = S23.elixir();
+            Array4<Real> s11 = S11.array();  Array4<Real> s22 = S22.array();  Array4<Real> s33 = S33.array();
+            Array4<Real> s12 = S12.array();  Array4<Real> s13 = S13.array();  Array4<Real> s23 = S23.array();
+
+            // Symmetric strain/stresses
+            Array4<Real> tau11 = Tau11->array(mfi); Array4<Real> tau22 = Tau22->array(mfi); Array4<Real> tau33 = Tau33->array(mfi);
+            Array4<Real> tau12 = Tau12->array(mfi); Array4<Real> tau13 = Tau13->array(mfi); Array4<Real> tau23 = Tau23->array(mfi);
+
+            // Strain magnitude
+            Array4<Real> SmnSmn_a;
+
+            if (l_use_terrain) {
+                // Terrain non-symmetric terms
+                FArrayBox S21,S31,S32;
+                S21.resize(tbxxy,1); S31.resize(tbxxz,1); S32.resize(tbxyz,1);
+                Elixir S21_eli     = S21.elixir();      Elixir S31_eli     = S31.elixir();      Elixir S32_eli     = S32.elixir();
+                Array4<Real> s21   = S21.array();       Array4<Real> s31   = S31.array();       Array4<Real> s32   = S32.array();
+                Array4<Real> tau21 = Tau21->array(mfi); Array4<Real> tau31 = Tau31->array(mfi); Array4<Real> tau32 = Tau32->array(mfi);
+
+                //-----------------------------------------
+                // Expansion rate compute terrain
+                //-----------------------------------------
+                {
+                BL_PROFILE("slow_rhs_making_er_T");
+                // First create Omega using velocity (not momentum)
+                Box gbxo = surroundingNodes(bxcc,2);
+                amrex::ParallelFor(gbxo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+                {
                     omega_arr(i,j,k) = (k == 0) ? 0. : OmegaFromW(i,j,k,w(i,j,k),u,v,z_nd,dxInv);
-                 });
+                });
 
-                 amrex::ParallelFor(gbx2, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
+                amrex::ParallelFor(bxcc, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+                {
 
[…]
