+++ Source/ERF.H
@@ -709,12 +709,26 @@ private:
     AMREX_FORCE_INLINE
     int
     ComputeGhostCells(const int& horiz_spatial_order, const int& vert_spatial_order,
-                      const int& spatial_order_WENO,  const bool& use_WENO) {
+                      const int& spatial_order_WENO,  const bool& use_WENO,
+                      const bool& use_numerical_diff) {
       int nGhostCells;
 
       int spatial_order = std::max(horiz_spatial_order, vert_spatial_order);
 
-      if(!use_WENO) {
+      if (use_numerical_diff) {
+          nGhostCells = 3;
+      } else if(use_WENO) {
+          switch (spatial_order_WENO) {
+          case 3:
+              nGhostCells = 2;
+              break;
+          case 5:
+              nGhostCells = 3;
+              break;
+          default:
+              amrex::Error("Must specify WENO spatial order to be 3 or 5");
+          }
+      } else {
           switch (spatial_order) {
           case 2:
               nGhostCells = 2; // We need this many to compute the eddy viscosity in the ghost cells
@@ -733,17 +747,6 @@ private:
               break;
           default:
               amrex::Error("Must specify spatial order to be 2,3,4,5 or 6");
-          }
-      } else {
-          switch (spatial_order_WENO) {
-          case 3:
-              nGhostCells = 2;
-              break;
-          case 5:
-              nGhostCells = 3;
-              break;
-          default:
-              amrex::Error("Must specify WENO spatial order to be 3 or 5");
           }
       }
 
