### 2023-03-20 4f2b46f
Clarify that Build provides example cmake scripts and delete deprecated cdash website.
