### 2023-03-20 #5202 @adam-m-jcbs
@kthyng in the interests of integrity and transparency, I'd like to declare what could be perceived as a COI and also that I nonetheless feel able and sufficiently independent to carry out "an impartial scientific evaluation" for the purposes of this JOSS review.

The submitting author, Ann (@asalmgren), and Berkeley Lab's [CCSE](https://ccse.lbl.gov) were collaborators on / partial funders of my PhD thesis work.  However, I earned my PhD in 2016 and am now based at NVIDIA.  The authors of this submission and I are not active collaborators currently (much to my dismay!) and I'm independent funding-wise, so I feel very able to carry out an impartial review.  I just wanted to be careful and transparent with this disclosure.

I'll proceed with the review, Kristen, but please do let me know if this raises issues or if you'd like to discuss more.
