### 2023-03-20 27fc827
Change default to initialize to constant KE rather than RhoKE
KE_0 param to override default of 0.1 (to match SOWFA, AMR-Wind)
RhoKE_0 param retained for backwards compatibility
