### 2023-03-21 71dec55
Add lower limit to the mixing length, following WRF
