+++ Source/Diffusion/ComputeStrain_N.cpp
@@ -54,6 +54,7 @@ ComputeStrain_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz,
         });
     }
     if (xh_v_dir) {
+        // note: tilebox xy should be nodal, so i|i-1|i-2 at the bigEnd is analogous to i-1|i|i+1 at the smallEnd
         Box planexy = tbxxy; planexy.setSmall(0, planexy.bigEnd(0) );
         tbxxy.growHi(0,-1);
         amrex::ParallelFor(planexy,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
@@ -71,6 +72,7 @@ ComputeStrain_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz,
         });
     }
     if (xh_w_dir) {
+        // note: tilebox xz should be nodal, so i|i-1|i-2 at the bigEnd is analogous to i-1|i|i+1 at the smallEnd
         Box planexz = tbxxz; planexz.setSmall(0, planexz.bigEnd(0) );
         tbxxz.growHi(0,-1);
         amrex::ParallelFor(planexz,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
@@ -90,6 +92,7 @@ ComputeStrain_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz,
         });
     }
     if (yh_u_dir) {
+        // note: tilebox xy should be nodal, so j|j-1|j-2 at the bigEnd is analogous to j-1|j|j+1 at the smallEnd
         Box planexy = tbxxy; planexy.setSmall(1, planexy.bigEnd(1) );
         tbxxy.growHi(1,-1);
         amrex::ParallelFor(planexy,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
@@ -107,6 +110,7 @@ ComputeStrain_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz,
         });
     }
     if (yh_w_dir) {
+        // note: tilebox yz should be nodal, so j|j-1|j-2 at the bigEnd is analogous to j-1|j|j+1 at the smallEnd
         Box planeyz = tbxyz; planeyz.setSmall(1, planeyz.bigEnd(1) );
         tbxyz.growHi(1,-1);
         amrex::ParallelFor(planeyz,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
@@ -123,9 +127,17 @@ ComputeStrain_N(Box& bxcc, Box& tbxxy, Box& tbxxz, Box& tbxyz,
         amrex::ParallelFor(planexz,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
             tau13(i,j,k) = 0.5 * ( (-(8./3.) * u(i,j,k-1) + 3. * u(i,j,k) - (1./3.) * u(i,j,k+1))*dxInv[2] +
                                    (w(i, j, k) - w(i-1, j, k))*dxInv[0]*mf_u(i,j,0) );
+            if((i==0)&&(j==0)&&(k==0))
+                amrex::AllPrint() << "ComputeStrain13(0,0,0):"
+                    << " " << u(i,j,k-1)
+                    << " " << u(i,j,k)
+                    << " " << u(i,j,k+1)
+                    << " " << tau13(i,j,k)
+                    << std::endl;
         });
     }
     if (zh_u_dir) {
+        // note: tilebox xz should be nodal, so k|k-1|k-2 at the bigEnd is analogous to k-1|k|k+1 at the smallEnd
[…]
