### 2023-03-21 0ca4b68
Update TKE RHS only once per timestep...
* use ComputeTurbulentViscosity to update the SFS heat flux and dissipation
  prior to the first RK stage
* use precomputed hfx, diss in DiffusionSrcForState_*
* only compute SmnSmn once (to keep production term frozen)
