+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -430,11 +430,14 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
 
         // Populate SmnSmn if using Deardorff (used as diff src in post)
         if (solverChoice.les_type == LESType::Deardorff) {
-            SmnSmn_a = SmnSmn->array(mfi);
-            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-            {
-                SmnSmn_a(i,j,k) = ComputeSmnSmn(i,j,k,tau11,tau22,tau33,tau12,tau13,tau23);
-            });
+            // keep the TKE tendencies constant for nrk>0
+            if (nrk==0) {
+                SmnSmn_a = SmnSmn->array(mfi);
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+                {
+                    SmnSmn_a(i,j,k) = ComputeSmnSmn(i,j,k,tau11,tau22,tau33,tau12,tau13,tau23);
+                });
+            }
         } else {
             SmnSmn_a = Array4<Real>{};
         }
