### 2023-03-21 3cddb42
Fix calls to ComputeSmnSmn
