### 2023-03-21 8f8dddf
Remove debugging print
