### 2023-03-21 50bb6e9
Clarify stability-corrected mixing length calc
