### 2023-03-21 f4f0e11
Remove unused mixing length calc
