+++ Source/Diffusion/DiffusionSrcForState_T.cpp
@@ -527,20 +527,6 @@ DiffusionSrcForState_T (const amrex::Box& bx, const amrex::Box& domain, int n_st
             Real DeltaMsf   = std::pow(cellVolMsf,1.0/3.0);
             Real met_h_zeta = Compute_h_zeta_AtCellCenter(i,j,k,dxInv,z_nd);
 
-            // Calculate stratification-dependent mixing length (Deardorff 1980)
-            Real eps       = std::numeric_limits<Real>::epsilon();
-            Real dtheta_dz = 0.5*(cell_prim(i,j,k+1,PrimTheta_comp)-cell_prim(i,j,k-1,PrimTheta_comp))*dz_inv;
-            dtheta_dz      /= met_h_zeta;
-            Real E         = cell_prim(i,j,k,PrimKE_comp);
-            Real strat     = l_abs_g * dtheta_dz * l_inv_theta0; // ==N^2 under stable conditions
-            Real length;
-            if (strat <= eps) {
-                length = DeltaMsf;
-            } else {
-                length = amrex::min(DeltaMsf, 0.76 * std::sqrt(E / strat));
-                length = amrex::max(length, 0.001 * DeltaMsf);
-            }
-
             // Add Buoyancy Source
             // where the SGS buoyancy flux tau_{theta,i} = -KH * dtheta/dx_i,
             // such that for dtheta/dz < 0, there is a positive (upward) heat
