### 2023-03-21 #5202 @kthyng
@adam-m-jcbs Given that none of the listed people were your direct supervisors, the collaborations were 6 years ago, and you feel able to carry out an impartial review, we are able to waive this conflict. Thank you for stating it clearing for the review issue.
