### 2023-03-22 b7bd99d
Cmake file for compiling with MPI + CUDA.
