### 2023-03-22 0e27951
Numerous adjustments to Deardorff model, change sign convention for tau (#1019)
* Add additional scalar profile output for completeness

* Fix sign issue with const g (abs val) vs vector g

* Add limit on mixing length for Deardorff

* Bound entire SFS KE field to ensure it is positive (>= eps)

Notes:
- AMR-Wind uses a clip value of 1e-8 (in equation_systems/tke/tke_ops.H)
- SOWFA uses a lower bound of "small" (i.e., eps, from SOWFA LESModel property
  kMin; small is defined as floatScalarSmall in OpenFOAM/primitives/Scalar/...)

* Update default value of RhoKE_0

If using a SFS model that ca
[…]
