+++ Source/Diffusion/DiffusionSrcForState_N.cpp
@@ -40,9 +40,8 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain, int n_st
 
     bool l_use_QKE       = solverChoice.use_QKE && solverChoice.advect_QKE;
     bool l_use_deardorff = (solverChoice.les_type == LESType::Deardorff);
-    Real l_Delta         = std::pow(dx_inv * dy_inv * dz_inv,-1./3.);
-    Real l_C_e           = solverChoice.Ce;
     Real l_inv_theta0    = 1.0 / solverChoice.theta_ref;
+    Real l_abs_g         = std::abs(grav_gpu[2]);
 
     bool l_consA  = (solverChoice.molec_diff_type == MolecDiffType::ConstantAlpha);
     bool l_turb   = ( (solverChoice.les_type == LESType::Smagorinsky) ||
@@ -348,43 +347,29 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain, int n_st
     // Using Deardorff
     if (l_use_deardorff && n_end >= RhoKE_comp) {
         int qty_index = RhoKE_comp;
-        amrex::Real l_C_k = solverChoice.Ck;
         amrex::ParallelFor(bx,[=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
         {
-            Real eps       = std::numeric_limits<Real>::epsilon();
-            Real dtheta_dz = 0.5*(cell_prim(i,j,k+1,PrimTheta_comp)-cell_prim(i,j,k-1,PrimTheta_comp))*dz_inv;
-            Real E         = cell_prim(i,j,k,PrimKE_comp);
-            Real strat     = std::abs(grav_gpu[2]) * dtheta_dz * l_inv_theta0; // stratification
-            Real length;
-            if (strat <= eps) {
-                length = l_Delta;
-            } else {
-              length = 0.76 * std::sqrt(E / strat);
-            }
-
-            // From eddy viscosity mu_turb = rho * C_k * l * KE^(1/2), the
-            // eddy diffusivity for heat, KH = (1 + 2*l/delta) * mu_turb
-            Real KH = cell_data(i,j,k,Rho_comp) * l_C_k * length * (1.+2.*length/l_Delta) * std::sqrt(E);
-
             // Add Buoyancy Source
             // where the SGS buoyancy flux tau_{theta,i} = -KH * dtheta/dx_i,
-            // such that for dtheta/dz < 0, there is a positive (upward) heat flux;
-            // the TKE buoyancy production is then g/theta_0 * tau_{theta,w}
-            hfx_x(i,j,k) = 0.0;
-            hfx_y(i,j,k) = 0.0;
-            hfx_z(i,j,k) = -KH * dtheta_dz;
-            cell_rhs(i,j,k,qty_index) += grav_gpu[2] * l_inv_theta0 * hfx_z(i,j,k);
+            // such that for dtheta/dz < 0, there is a positive (upward) heat
+            // flux; the TKE buoyancy production is then
+            //   B = g/theta_0 * tau_{theta,w}
+            // for a dry atmosphere (see, e.g., Sullivan et al 1994). To
+            // account for moisture, the Brunt-Vaisala frequency,
+            //   N^2 = g[1/theta * dtheta/dz + ...]
+            // **should** be a function of the water vapor and total water
+            // mixing ratios, depending on whether conditions are saturated or
+            // not (see the WRF model description, Skamarock et al 2019).
+            cell_rhs(i,j,k,qty_index) += l_abs_g * l_inv_theta0 * hfx_z(i,j,k);
 
             // TKE shear production
[…]

[…]
