### 2023-03-24 15ba1b1
Fix diffusion in slow rhs post.
