### 2023-03-27 erf-model/ERF#1010 @dwillcox
@pratikvn I've addressed the point you raised about the paper in #1022 -- the paper is now organized into a high level overview and a set of feature sub-sections. Can you confirm this addresses your concerns?
