+++ Source/ERF.cpp
@@ -10,6 +10,7 @@
 
 #include <Utils.H>
 #include <TerrainMetrics.H>
+#include <memory>
 
 #ifdef ERF_USE_MULTIBLOCK
 #include <MultiBlockContainer.H>
@@ -53,13 +54,13 @@ amrex::Real ERF::sum_per       = -1.0;
 std::string ERF::plotfile_type    = "amrex";
 
 // init_type:  "ideal", "real", "input_sounding", "metgrid" or ""
-std::string ERF::init_type        = "";
+std::string ERF::init_type;
 
 // NetCDF wrfinput (initialization) file(s)
 amrex::Vector<amrex::Vector<std::string>> ERF::nc_init_file = {{""}}; // Must provide via input
 
 // NetCDF wrfbdy (lateral boundary) file
-std::string ERF::nc_bdy_file = ""; // Must provide via input
+std::string ERF::nc_bdy_file; // Must provide via input
 
 // Text input_sounding file
 std::string ERF::input_sounding_file = "input_sounding";
@@ -185,8 +186,7 @@ ERF::ERF ()
 }
 
 ERF::~ERF ()
-{
-}
+= default;
 
 // advance solution to final time
 void
@@ -364,7 +364,7 @@ ERF::InitData ()
     last_plot_file_step_2 = -1;
     last_check_file_step = -1;
 
-    if (restart_chkfile == "") {
+    if (restart_chkfile.empty()) {
         // start simulation from the beginning
 
         const Real time = 0.0;
@@ -532,7 +532,7 @@ ERF::InitData ()
         m_most->update_fluxes(lev);
     }
 
-    if (restart_chkfile == "" && check_int > 0)
+    if (restart_chkfile.empty() && check_int > 0)
     {
 #ifdef ERF_USE_NETCDF
         if (check_type == "netcdf") {
@@ -545,8 +545,8 @@ ERF::InitData ()
         last_check_file_step = 0;
     }
 
-    if ( (restart_chkfile == "") ||
-         (restart_chkfile != "" && plot_file_on_restart) )
+    if ( (restart_chkfile.empty()) ||
+         (!restart_chkfile.empty() && plot_file_on_restart) )
     {
         if (plot_int_1 > 0)
         {
@@ -566,7 +566,7 @@ ERF::InitData ()
         if (init_type == "input_sounding")
         {
             // overwrite Ubar, Tbar, and thetabar with input profiles
-            bool restarting = (restart_chkfile != "");
+            bool restarting = (!restart_chkfile.empty());
             setRayleighRefFromSounding(restarting);
         }
 
@@ -956,25 +956,25 @@ void ERF::MakeNewLevelFromScratch (int lev, Real /*time*/, const BoxArray& ba,
     SmnSmn_lev.resize(lev+1);
 
     if (l_use_diff) {
-        Tau11_lev[lev].reset( new MultiFab(ba  , dm, 1, IntVect(1,1,0)) );
-        Tau22_lev[lev].reset( new MultiFab(ba  , dm, 1, IntVect(1,1,0)) );
-        Tau33_lev[lev].reset( new MultiFab(ba  , dm, 1, IntVect(1,1,0)) );
-        Tau12_lev[lev].reset( new MultiFab(ba12, dm, 1, IntVect(1,1,0)) );
-        Tau13_lev[lev].reset( new MultiFab(ba13, dm, 1, IntVect(1,1,0)) );
[…]
