+++ Source/Utils/DirectionSelector.H
@@ -9,25 +9,25 @@
 template <int Index>
 struct DirectionSelector
 {
-   int getIndx(int i, int j, int k) const;
+   [[nodiscard]] int getIndx(int i, int j, int k) const;
 };
 
 template <>
 struct DirectionSelector<0>
 {
-    AMREX_GPU_HOST_DEVICE int getIndx(int i, int, int) const { return i; }
+    AMREX_GPU_HOST_DEVICE [[nodiscard]] static int getIndx(int i, int, int) { return i; }
 };
 
 template <>
 struct DirectionSelector<1>
 {
-    AMREX_GPU_HOST_DEVICE int getIndx(int, int j, int) const { return j; }
+    AMREX_GPU_HOST_DEVICE [[nodiscard]] static int getIndx(int, int j, int) { return j; }
 };
 
 template <>
 struct DirectionSelector<2>
 {
-    AMREX_GPU_HOST_DEVICE int getIndx(int, int, int k) const { return k; }
+    AMREX_GPU_HOST_DEVICE [[nodiscard]] static int getIndx(int, int, int k) { return k; }
 };
 
 using XDir = DirectionSelector<0>;
