+++ Source/BoundaryConditions/ABLMost.cpp
@@ -9,8 +9,8 @@ void ABLMost::update_fluxes(int lev, int max_iters)
     m_ma.compute_averages(lev);
 
     // Pointers to the computed averages
-    const auto tm_ptr  = m_ma.get_average(lev,2);
-    const auto umm_ptr = m_ma.get_average(lev,3);
+    const auto *const tm_ptr  = m_ma.get_average(lev,2);
+    const auto *const umm_ptr = m_ma.get_average(lev,3);
 
     // GPU device captures
     amrex::Real d_kappa = kappa;
@@ -155,10 +155,10 @@ ABLMost::impose_most_bcs(const int lev,
         const auto  eta_arr = eddyDiffs->array(mfi);
 
         // Get average arrays
-        const auto u_mean     = m_ma.get_average(lev,0);
-        const auto v_mean     = m_ma.get_average(lev,1);
-        const auto t_mean     = m_ma.get_average(lev,2);
-        const auto u_mag_mean = m_ma.get_average(lev,3);
+        const auto *const u_mean     = m_ma.get_average(lev,0);
+        const auto *const v_mean     = m_ma.get_average(lev,1);
+        const auto *const t_mean     = m_ma.get_average(lev,2);
+        const auto *const u_mag_mean = m_ma.get_average(lev,3);
 
         const auto um_arr  = u_mean->array(mfi);
         const auto vm_arr  = v_mean->array(mfi);
