### 2023-03-28 16470c4
clang-tidy suggested these fixes; update amrex submodule (#1023)
* clang-tidy suggested these fixes; update amrex submodule

* fix for HIP
