+++ Source/IO/ERF_WriteBndryPlanes.H
@@ -17,7 +17,7 @@ public:
     explicit WriteBndryPlanes(amrex::Vector<amrex::BoxArray>& grids,
                               amrex::Vector<amrex::Geometry>& geom);
 
-    void write_planes(int istep, amrex::Real time,
+    void write_planes(int t_step, amrex::Real time,
                       amrex::Vector<amrex::Vector<amrex::MultiFab>>& vars_new);
 
 private:
