+++ Source/Initialization/ERF_init_custom.cpp
@@ -60,7 +60,7 @@ ERF::init_custom(int lev)
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
 #endif
-    for (MFIter mfi(lev_new[Vars::cons], TilingIfNotGPU()); mfi.isValid(); ++mfi)
+    for (MFIter mfi(lev_new[Vars::cons], false); mfi.isValid(); ++mfi)
     {
         const Box &bx  = mfi.tilebox();
         const Box &xbx = mfi.tilebox(IntVect(1,0,0));
