+++ Source/IndexDefines.H
@@ -28,8 +28,8 @@
 #define PrimQKE_comp     (RhoQKE_comp   -1)
 #define PrimScalar_comp  (RhoScalar_comp-1)
 #if defined(ERF_USE_MOISTURE)
-  #define PrimQt_comp    RhoQt_comp-1
-  #define PrimQp_comp    RhoQp_comp-1
+  #define PrimQt_comp    (RhoQt_comp-1)
+  #define PrimQp_comp    (RhoQp_comp-1)
 #elif defined(ERF_USE_WARM_NO_PRECIP)
   #define PrimQv_comp    RhoQv_comp-1
   #define PrimQc_comp    RhoQc_comp-1
