### 2023-03-28 c88486e
More clang fixes (#1024)
* more clang-tidy fixes

* update CUDA CI
