+++ Source/IO/ERF_ReadBndryPlanes.H
@@ -41,8 +41,8 @@ public:
     [[nodiscard]] int ingested_density()  const {return is_density_read;}
     [[nodiscard]] int ingested_scalar()   const {return is_scalar_read;}
 #if defined(ERF_USE_MOISTURE)
-    int ingested_qt()       const {return is_qt_read;}
-    int ingested_qp()       const {return is_qp_read;}
+    [[nodiscard]] int ingested_qt()       const {return is_qt_read;}
+    [[nodiscard]] int ingested_qp()       const {return is_qp_read;}
 #elif defined(ERF_USE_WARM_NO_PRECIP)
     int ingested_qv()       const {return is_qv_read;}
     int ingested_qc()       const {return is_qc_read;}
