### 2023-03-29 f5fd7b7
function TileNoZ() to return tile size that prevents tiling across z
