+++ Source/EOS.H
@@ -3,6 +3,8 @@
 
 #include <ERF_Constants.H>
 #include <AMReX.H>
+#include <AMReX_IntVect.H>
+#include <AMReX_MFIter.H>
 #include <cmath>
 
 AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
@@ -70,5 +72,33 @@ amrex::Real getRhoThetagivenP(const amrex::Real p)
     // see https://erf.readthedocs.io/en/latest/theory/NavierStokesEquations.html
     return std::pow(p*std::pow(p_0, Gamma-1), iGamma) * iR_d;
 }
+
+// this function returns explicit tile size that prevents tiling in z
+AMREX_GPU_HOST_DEVICE AMREX_FORCE_INLINE
+amrex::IntVect TileNoZ()
+{
+    if (amrex::TilingIfNotGPU())
+        return amrex::IntVect{amrex::FabArrayBase::mfiter_tile_size[0],
+                              amrex::FabArrayBase::mfiter_tile_size[1],
+                              1024000};
+    else
+        return amrex::IntVect::TheZeroVector();
+}
+// should we do this based on number of dims?
+/*
+#if AMREX_SPACEDIM == 1
+    return IntVect{FabArrayBase::mfiter_tile_size[0],
+                   FabArrayBase::mfiter_tile_size[1],
+                   1024000};
+#elif AMREX_SPACEDIM == 2
+    return IntVect{FabArrayBase::mfiter_tile_size[0],
+                   1024000};
+#else
+return IntVect{FabArrayBase::mfiter_tile_size[0],
+                   FabArrayBase::mfiter_tile_size[1],
+                   1024000};
+#endif
+*/
+
 #endif
 
