### 2023-03-29 1c9c85a
Rollback an automatic, erroneous clang-tidy change.
