### 2023-03-29 9d1f93d
Remove conflicting 'static' declaration from several ERF functions.
