### 2023-03-29 7f16cb3
Remove 'static' from define_grids_to_evolve (AMReX compatibility).
