### 2023-03-29 0f40e24
Remove static from AverageDownTo (AMReX compatibility).
