### 2023-03-29 3634925
Add missing 'static' declaration.
