### 2023-03-29 4ee4d61
MakeDiagnosticAverage no longer static (AMReX compatibility).
