+++ Source/ERF_Constants.H
@@ -1,5 +1,5 @@
-#ifndef CONSTANTS_H_
-#define CONSTANTS_H_
+#ifndef ERF_CONSTANTS_H_
+#define ERF_CONSTANTS_H_
 
 #include <AMReX_REAL.H>
 
