### 2023-03-29 7674bff
Clang tidy fixes (#1025)
* Some automatic clang-tidy fixes.

* Fix clang-tidy for MOSTAverage class.

* Rollback an automatic, erroneous clang-tidy change.

* Remove conflicting 'static' declaration from several ERF functions.

* Remove 'static' from define_grids_to_evolve (AMReX compatibility).

* Remove static from AverageDownTo (AMReX compatibility).

* Add missing 'static' declaration.

* MakeDiagnosticAverage no longer static (AMReX compatibility).

* Fix header ifdef statement.
