### 2023-03-30 cc9ae3d
these routines will be overhauled but if these fixes break the benchmark I'd like it to be done now.
