### 2023-03-30 2565d24
Use NOLINT to avoid erroneous clang-tidy autofixes.
