### 2023-03-30 962b74f
Fix Clang tidy autofixes (#1027)
* Use NOLINT to avoid erroneous clang-tidy autofixes.

* More NOLINT...
