+++ Source/TimeIntegration/ERF_make_buoyancy.cpp
@@ -132,7 +132,7 @@ void make_buoyancy (BoxArray& grids_to_evolve,
 #if defined(ERF_USE_MOISTURE)
     if (solverChoice.buoyancy_type == 1) {
 
-        for ( MFIter mfi(buoyancy,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        for ( MFIter mfi(buoyancy,TileNoZ()); mfi.isValid(); ++mfi)
         {
             const Box& valid_bx = surroundingNodes(grids_to_evolve[mfi.index()],2);
 
@@ -204,7 +204,7 @@ void make_buoyancy (BoxArray& grids_to_evolve,
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
 #endif
-        for ( MFIter mfi(buoyancy,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        for ( MFIter mfi(buoyancy,TileNoZ()); mfi.isValid(); ++mfi)
         {
             const Box& valid_bx = surroundingNodes(grids_to_evolve[mfi.index()],2);
 
@@ -255,7 +255,7 @@ void make_buoyancy (BoxArray& grids_to_evolve,
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
 #endif
-        for ( MFIter mfi(buoyancy,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        for ( MFIter mfi(buoyancy,TileNoZ()); mfi.isValid(); ++mfi)
         {
             const Box& valid_bx = surroundingNodes(grids_to_evolve[mfi.index()],2);
 
@@ -302,7 +302,7 @@ void make_buoyancy (BoxArray& grids_to_evolve,
 
     AMREX_ALWAYS_ASSERT(solverChoice.buoyancy_type == 1);
 
-        for ( MFIter mfi(buoyancy,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        for ( MFIter mfi(buoyancy,TileNoZ()); mfi.isValid(); ++mfi)
         {
             const Box& valid_bx = surroundingNodes(grids_to_evolve[mfi.index()],2);
 
