### 2023-03-31 c675d4e
add link checking to docs CI
