+++ .github/workflows/docs.yml
@@ -26,7 +26,7 @@ jobs:
 
       - name: Check links
         run: |
-          cd Docs
+          cd Docs/sphinx_doc
           make SPHINXOPTS=-v NO_DOXYGEN=TRUE linkcheck
 
       - name: Deploy to Webpage
