+++ Source/BoundaryConditions/MOSTAverage.cpp
@@ -1,5 +1,6 @@
 #include <MOSTAverage.H>
 #include <utility>
+#include <EOS.H>
 
 // Constructor
 MOSTAverage::MOSTAverage (amrex::Vector<amrex::Geometry>  geom,
@@ -259,7 +260,7 @@ MOSTAverage::set_k_indices_T()
     if (read_z) {
         for (int lev(0); lev < m_maxlev; lev++) {
             int kmax = m_geom[lev].Domain().bigEnd(2);
-            for (amrex::MFIter mfi(*m_k_indx[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+            for (amrex::MFIter mfi(*m_k_indx[lev], TileNoZ()); mfi.isValid(); ++mfi) {
                 amrex::Box npbx  = mfi.tilebox(); npbx.convert({1,1,0});
                 const auto z_phys_arr = m_z_phys_nd[lev]->const_array(mfi);
                 auto k_arr = m_k_indx[lev]->array(mfi);
@@ -303,7 +304,7 @@ MOSTAverage::set_norm_indices_T()
         int kmax = m_geom[lev].Domain().bigEnd(2);
         const auto dxInv  = m_geom[lev].InvCellSizeArray();
         amrex::IntVect ng = m_k_indx[lev]->nGrowVect(); ng[2]=0;
-        for (amrex::MFIter mfi(*m_k_indx[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+        for (amrex::MFIter mfi(*m_k_indx[lev], TileNoZ()); mfi.isValid(); ++mfi) {
             amrex::Box npbx  = mfi.tilebox(); npbx.convert({1,1,0});
             amrex::Box gpbx  = mfi.growntilebox(ng);
             const auto z_phys_arr = m_z_phys_nd[lev]->const_array(mfi);
@@ -368,7 +369,7 @@ MOSTAverage::set_z_positions_T()
         amrex::RealVect base;
         const auto dx = m_geom[lev].CellSizeArray();
         amrex::IntVect ng = m_x_pos[lev]->nGrowVect(); ng[2]=0;
-        for (amrex::MFIter mfi(*m_x_pos[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+        for (amrex::MFIter mfi(*m_x_pos[lev], TileNoZ()); mfi.isValid(); ++mfi) {
             amrex::Box npbx  = mfi.tilebox(); npbx.convert({1,1,0});
             amrex::Box gpbx  = mfi.growntilebox(ng);
             amrex::RealBox grb{gpbx,dx.data(),base.dataPtr()};
@@ -409,7 +410,7 @@ MOSTAverage::set_norm_positions_T()
         const auto dx = m_geom[lev].CellSizeArray();
         const auto dxInv  = m_geom[lev].InvCellSizeArray();
         amrex::IntVect ng = m_x_pos[lev]->nGrowVect(); ng[2]=0;
-        for (amrex::MFIter mfi(*m_x_pos[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+        for (amrex::MFIter mfi(*m_x_pos[lev], TileNoZ()); mfi.isValid(); ++mfi) {
             amrex::Box npbx  = mfi.tilebox(); npbx.convert({1,1,0});
             amrex::Box gpbx  = mfi.growntilebox(ng);
             amrex::RealBox grb{gpbx,dx.data(),base.dataPtr()};
@@ -519,7 +520,7 @@ MOSTAverage::compute_plane_averages(int lev)
 #ifdef _OPENMP
[…]
