### 2023-03-31 9573c36
add link checking to docs CI (#1029)
* add link checking to docs CI

* fix directory

* fixes for broken links

* more link fixes

* white spaces

* more link fixes
