+++ Exec/SuperCell/prob.cpp
@@ -200,7 +200,7 @@ erf_init_dens_hse(MultiFab& rho_hse,
 
         if (khi > 255) amrex::Abort("1D Arrays are hard-wired to only 256 high");
 
-        for ( MFIter mfi(rho_hse, TilingIfNotGPU()); mfi.isValid(); ++mfi )
+        for ( MFIter mfi(rho_hse, TileNoZ()); mfi.isValid(); ++mfi )
         {
             Array4<Real      > rho_arr  = rho_hse.array(mfi);
             Array4<Real const> z_cc_arr = z_phys_cc->const_array(mfi);
