+++ Source/Microphysics/IceFall.cpp
@@ -1,6 +1,7 @@
 
 #include <AMReX_ParReduce.H>
 #include "Microphysics.H"
+#include "EOS.H"
 
 using namespace amrex;
 
@@ -64,7 +65,7 @@ void Microphysics::IceFall() {
   //   for (int j=0; j<ny; j++) {
   //     for (int i=0; i<nx; i++) {
   //       for (int icrm=0; icrm<ncrms; icrm++) {
-  for ( amrex::MFIter mfi(*tabs, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+  for ( amrex::MFIter mfi(*tabs, TileNoZ()); mfi.isValid(); ++mfi) {
      //auto qcl_array   = qcl->array(mfi);
      auto qci_array   = qci->array(mfi);
      auto qt_array    = qt->array(mfi);
