+++ Source/TimeIntegration/TimeIntegration.cpp
@@ -84,7 +84,7 @@ void ERF::erf_advance(int level,
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
 #endif
-        for ( MFIter mfi(cons_new,TilingIfNotGPU()); mfi.isValid(); ++mfi)
+        for ( MFIter mfi(cons_new,TileNoZ()); mfi.isValid(); ++mfi)
         {
             Box bxcc  = mfi.growntilebox(IntVect(1,1,0));
             Box tbxxy = mfi.tilebox(IntVect(1,1,0),IntVect(1,1,0));
