+++ Source/Microphysics/IceFall.cpp
@@ -1,7 +1,7 @@
 
 #include <AMReX_ParReduce.H>
 #include "Microphysics.H"
-#include "EOS.H"
+#include "TileNoZ.H"
 
 using namespace amrex;
 
