+++ Source/TimeIntegration/ERF_fast_rhs_N.cpp
@@ -3,7 +3,7 @@
 #include <AMReX_MultiFabUtil.H>
 #include <AMReX_ArrayLim.H>
 #include <AMReX_BC_TYPES.H>
-#include <EOS.H>
+#include <TileNoZ.H>
 #include <ERF_Constants.H>
 #include <IndexDefines.H>
 #include <TimeIntegration.H>
