+++ Source/TimeIntegration/TimeIntegration.cpp
@@ -2,12 +2,14 @@
 #include <AMReX_TimeIntegrator.H>
 #include <ERF_MRI.H>
 #include <EddyViscosity.H>
+#include <EOS.H>
+#include <ERF.H>
 #include <TerrainMetrics.H>
 #include <TimeIntegration.H>
 #include <PlaneAverage.H>
 #include <Diffusion.H>
-#include <ERF.H>
-#include <EOS.H>
+#include <TileNoZ.H>
+#include <Utils.H>
 
 using namespace amrex;
 
