+++ CMake/BuildERFExe.cmake
@@ -141,6 +141,7 @@ function(build_erf_lib erf_lib_name)
        ${SRC_DIR}/Diffusion/NumericalDiffusion.cpp
        ${SRC_DIR}/Utils/MomentumToVelocity.cpp
        ${SRC_DIR}/Diffusion/PBLModels.cpp
+       ${SRC_DIR}/Utils/TileNoZ.H
        ${SRC_DIR}/Utils/Utils.H
        ${SRC_DIR}/Utils/TerrainMetrics.H
        ${SRC_DIR}/Utils/TerrainMetrics.cpp
