+++ Source/BoundaryConditions/MOSTAverage.cpp
@@ -1,17 +1,19 @@
 #include <MOSTAverage.H>
 #include <utility>
-#include <EOS.H>
+#include <TileNoZ.H>
+
+using namespace amrex;
 
 // Constructor
-MOSTAverage::MOSTAverage (amrex::Vector<amrex::Geometry>  geom,
-                          amrex::Vector<amrex::Vector<amrex::MultiFab>>& vars_old,
-                          amrex::Vector<std::unique_ptr<amrex::MultiFab>>& Theta_prim,
-                          amrex::Vector<std::unique_ptr<amrex::MultiFab>>& z_phys_nd)
+MOSTAverage::MOSTAverage (Vector<Geometry>  geom,
+                          Vector<Vector<MultiFab>>& vars_old,
+                          Vector<std::unique_ptr<MultiFab>>& Theta_prim,
+                          Vector<std::unique_ptr<MultiFab>>& z_phys_nd)
   : m_geom(std::move(geom))
 {
     // Get basic info
     //--------------------------------------------------------
-    amrex::ParmParse pp(m_pp_prefix);
+    ParmParse pp(m_pp_prefix);
     pp.query("most.radius",m_radius);
     pp.query("most.time_average",m_t_avg);
     pp.query("most.average_policy",m_policy);
@@ -46,70 +48,70 @@ MOSTAverage::MOSTAverage (amrex::Vector<amrex::Geometry>  geom,
       m_z_phys_nd[lev] = z_phys_nd[lev].get();
       { // Nodal in x
         auto& mf  = vars_old[lev][Vars::xvel];
-        amrex::MultiFab* mfp = &vars_old[lev][Vars::xvel];
+        MultiFab* mfp = &vars_old[lev][Vars::xvel];
         // Create a 2D ba, dm, & ghost cells
-        const amrex::BoxArray& ba  = mf.boxArray();
-        amrex::BoxList bl2d = ba.boxList();
+        const BoxArray& ba  = mf.boxArray();
+        BoxList bl2d = ba.boxList();
         for (auto& b : bl2d) b.setRange(2,0);
-        amrex::BoxArray ba2d(std::move(bl2d));
-        const amrex::DistributionMapping& dm = mf.DistributionMap();
+        BoxArray ba2d(std::move(bl2d));
+        const DistributionMapping& dm = mf.DistributionMap();
         const int ncomp   = 1;
-        amrex::IntVect ng = mf.nGrowVect(); ng[2]=0;
+        IntVect ng = mf.nGrowVect(); ng[2]=0;
 
           m_fields[lev][0] = mfp;
-        m_averages[lev][0] = new amrex::MultiFab(ba2d,dm,ncomp,ng);
+        m_averages[lev][0] = new MultiFab(ba2d,dm,ncomp,ng);
         m_averages[lev][0]->setVal(1.E34);
       }
       { // Nodal in y
         auto& mf  = vars_old[lev][Vars::yvel];
-        amrex::MultiFab* mfp = &vars_old[lev][Vars::yvel];
+        MultiFab* mfp = &vars_old[lev][Vars::yvel];
         // Create a 2D ba, dm, & ghost cells
-        const amrex::BoxArray& ba  = mf.boxArray();
-        amrex::BoxList bl2d = ba.boxList();
+        const BoxArray& ba  = mf.boxArray();
+        BoxList bl2d = ba.boxList();
         for (auto& b : bl2d) b.setRange(2,0);
-        amrex::BoxArray ba2d(std::move(bl2d));
[…]
