### 2023-04-02 #5202 @pratikvn
An update on the review from my side.

1. I think the paper looks good to me.
2. The software is lacking an API documentation. Atleast I cannot seem to find it on the webpage yet. There is feature and functionality documentation, but the API documentation is missing (I believe there is an open PR for this though: https://github.com/erf-model/ERF/issues/165).
3. I had a few installation issues and I have reported them in the repository. I believe some of them have being resolved, but others need attention.
4. I am not sure there are any particular performance claims made. They do mention performance portability and ability to run with the MPI + X model, but there have been no specific comparisons with respect to other libraries or performance in general.
5. There are a few regression and functionality tests which are automated, but I cannot find any real-world examples.

So to summarize, three things are blocking IMO:
* Installation issues (with dependencies documentation)
* API documentation.
* Real world examples with possible some nice output that can be visualized.
