### 2023-04-02 6b9f49c
clean up cmake files
