+++ Source/TimeIntegration/ERF_make_condensation_source.cpp
@@ -1,4 +1,5 @@
 #include <ERF.H>
+#include <EOS.H>
 #include <TileNoZ.H>
 
 using namespace amrex;
