+++ Exec/EkmanSpiral_input_sounding/CMakeLists.txt
@@ -3,7 +3,6 @@ set(erf_exe_name ekman_spiral_input_sounding)
 add_executable(${erf_exe_name} "")
 target_sources(${erf_exe_name}
    PRIVATE
-     prob.H
      prob.cpp
 )
 
