+++ CMake/BuildERFExe.cmake
@@ -15,13 +15,15 @@ function(build_erf_lib erf_lib_name)
   include(${CMAKE_SOURCE_DIR}/CMake/SetERFCompileFlags.cmake)
   set_erf_compile_flags(${erf_lib_name})
 
-  set(ERF_EOS_DIR "${CMAKE_SOURCE_DIR}/Source")
-  target_sources(${erf_lib_name} PRIVATE
-                 ${ERF_EOS_DIR}/EOS.H)
-  target_include_directories(${erf_lib_name} SYSTEM PUBLIC ${ERF_EOS_DIR})
-
   if(ERF_ENABLE_MOISTURE)
     target_compile_definitions(${erf_lib_name} PUBLIC ERF_USE_MOISTURE)
+  endif()
+
+  if(ERF_ENABLE_MULTIBLOCK)
+    target_sources(${erf_lib_name} PRIVATE
+                   ${SRC_DIR}/MultiBlock/MultiBlockContainer.cpp)
+    target_compile_definitions(${erf_lib_name} PUBLIC ERF_USE_MULTIBLOCK)
+    target_include_directories(${erf_lib_name} PUBLIC ${SRC_DIR}/MultiBlock)
   endif()
 
   if(ERF_ENABLE_WARM_NO_PRECIP)
@@ -34,18 +36,8 @@ function(build_erf_lib erf_lib_name)
     target_compile_definitions(${erf_lib_name} PUBLIC ERF_USE_POISSON_SOLVE)
   endif()
 
-  if(ERF_ENABLE_MULTIBLOCK)
-    target_sources(${erf_lib_name} PRIVATE
-                   ${SRC_DIR}/MultiBlock/MultiBlockContainer.H
-                   ${SRC_DIR}/MultiBlock/MultiBlockContainer.cpp)
-    target_compile_definitions(${erf_lib_name} PUBLIC ERF_USE_MULTIBLOCK)
-  endif()
-
   if(ERF_ENABLE_NETCDF)
     target_sources(${erf_lib_name} PRIVATE
-                   ${SRC_DIR}/IO/NCInterface.H
-                   ${SRC_DIR}/IO/NCWpsFile.H
-                   ${SRC_DIR}/IO/NCPlotFile.H
                    ${SRC_DIR}/IO/NCBuildFABs.cpp
                    ${SRC_DIR}/IO/NCInterface.cpp
                    ${SRC_DIR}/IO/NCPlotFile.cpp
@@ -60,7 +52,6 @@ function(build_erf_lib erf_lib_name)
 
   if(ERF_ENABLE_MOISTURE)
     target_sources(${erf_lib_name} PRIVATE
-       ${SRC_DIR}/Microphysics/Microphysics.H
        ${SRC_DIR}/Microphysics/Init.cpp
        ${SRC_DIR}/Microphysics/Cloud.cpp
        ${SRC_DIR}/Microphysics/IceFall.cpp
@@ -77,18 +68,12 @@ function(build_erf_lib erf_lib_name)
 
   target_sources(${erf_lib_name}
      PRIVATE
-       ${SRC_DIR}/DataStruct.H
-       ${SRC_DIR}/ERF_Constants.H
-       ${SRC_DIR}/Derive.H
        ${SRC_DIR}/Derive.cpp
-       ${SRC_DIR}/IndexDefines.H
-       ${SRC_DIR}/ERF.H
        ${SRC_DIR}/ERF.cpp
        ${SRC_DIR}/ERF_Tagging.cpp
-       ${SRC_DIR}/prob_common.H
-       ${SRC_DIR}/BoundaryConditions/ABLMost.H
[…]
