### 2023-04-03 b95e478
clean up cmake files (#1033)
