+++ Docs/sphinx_doc/index.rst
@@ -6,11 +6,14 @@ ERF
 ERF solves the compressible Navier-Stokes on a Arakawa C-grid for large-scale weather modeling.
 
 ERF is built on `AMReX <https://github.com/AMReX-Codes/amrex>`_,
-an adaptive mesh refinement software framework, which provides the underlying software infrastructure for
-block structured AMR operations.
-The full AMReX documentation can be found `here <https://amrex-codes.github.io/amrex/docs_html/>`_ and the tutorials can be found `here <https://amrex-codes.github.io/amrex/tutorials_html/>`_.
+an adaptive mesh refinement software framework, which provides the underlying software infrastructure 
+and performance portability.
+The full AMReX documentation can be found `here <https://amrex-codes.github.io/amrex/docs_html/>`_
+and tutorials can be found `here <https://amrex-codes.github.io/amrex/tutorials_html/>`_.
 
 ERF is designed to run on machines from laptops to multicore CPU and hybrid CPU/GPU systems.
+
+In addition to this documentation, there is API documentation for ERF generated by Doxygen<https://erf-model.github.io/docs/index.html>`_.
 
 .. raw:: html
 
