### 2023-04-03 db23e9b
fixes for tiling in fast_rhs_T
