+++ Source/TimeIntegration/ERF_fast_rhs_MT.cpp
@@ -85,7 +85,8 @@ void erf_fast_rhs_MT (int step, int /*level*/,
     FArrayBox RHS_fab;
     FArrayBox soln_fab;
 
-    for ( MFIter mfi(S_stg_data[IntVar::cons],TileNoZ()); mfi.isValid(); ++mfi)
+//    for ( MFIter mfi(S_stg_data[IntVar::cons],TileNoZ()); mfi.isValid(); ++mfi)
+    for ( MFIter mfi(S_stg_data[IntVar::cons],false); mfi.isValid(); ++mfi)
     {
         // Construct intersection of current tilebox and valid region for updating
         Box valid_bx = grids_to_evolve[mfi.index()];
