+++ Docs/sphinx_doc/index.rst
@@ -13,7 +13,7 @@ and tutorials can be found `here <https://amrex-codes.github.io/amrex/tutorials_
 
 ERF is designed to run on machines from laptops to multicore CPU and hybrid CPU/GPU systems.
 
-In addition to this documentation, there is API documentation for ERF generated by `Doxygen<https://erf-model.github.io/docs/index.html>`_.
+In addition to this documentation, there is API documentation for ERF generated by `Doxygen <https://erf-model.github.io/docs/index.html>`_.
 
 .. raw:: html
 
