### 2023-04-04 8d85d12
docstrings for the rest of Initialization.
