+++ Source/BoundaryConditions/BoundaryConditions_wrfbdy.cpp
@@ -1,4 +1,5 @@
 #include "ERF.H"
+#include "Utils.H"
 
 using namespace amrex;
 
@@ -34,141 +35,50 @@ ERF::fill_from_wrfbdy (const Vector<MultiFab*>& mfs, const Real time)
 
         if (var_idx == Vars::cons || var_idx == Vars::zvel)
         {
-
             const auto& dom_lo = amrex::lbound(domain);
             const auto& dom_hi = amrex::ubound(domain);
 
             MultiFab& mf = *mfs[var_idx];
 
-            IntVect ng_vect = mf.nGrowVect();
+            IntVect ng_vect = mf.nGrowVect(); ng_vect[2] = 0;
 
 #ifdef AMREX_USE_OMP
 #pragma omp parallel if (Gpu::notInLaunchRegion())
 #endif
-            for (MFIter mfi(mf); mfi.isValid(); ++mfi)
+            for (MFIter mfi(mf,TilingIfNotGPU()); mfi.isValid(); ++mfi)
             {
-                const Array4<Real>& dest_arr = mf.array(mfi);
-                Box bx  = mfi.tilebox();
+                // Grown tilebox so we fill exterior ghost cells as well
                 Box gbx = mfi.growntilebox(ng_vect);
+                const Array4<Real>& dest_arr = mf.array(mfi);
 
-                const auto& bx_lo = amrex::lbound(bx);
-                const auto& bx_hi = amrex::ubound(bx);
+                Box bx_xlo, bx_xhi, bx_ylo, bx_yhi;
+                compute_interior_ghost_bxs_xy(gbx, domain, width,
+                                              bx_xlo, bx_xhi,
+                                              bx_ylo, bx_yhi, ng_vect);
 
-                // Note: "domain" here is the domain for the centering of this variable, i.e.
-                //       it is z-face centered for w and cell-centered for the state
-
-                // x-faces
+                // x-faces (includes y ghost cells)
+                ParallelFor(bx_xlo, ncomp, [=] AMREX_GPU_DEVICE (int i, int j, int k, int n)
                 {
-                    if (bx_lo.x == dom_lo.x)
-                    {
-                        Box bx_xlo(gbx & domain);
-                        bx_xlo.setSmall(0,dom_lo.x-ng_vect[0]);
-                        bx_xlo.setBig(0,dom_lo.x+width-1);
-
-                        ParallelFor(bx_xlo, ncomp, [=] AMREX_GPU_DEVICE (int i, int j, int k, int n)
-                        {
-                            dest_arr(i,j,k,icomp+n) = dest_arr(dom_lo.x+width,j,k,icomp+n);
-                        });
-                    } // bx
-
-                    if (bx_hi.x == dom_hi.x)
-                    {
-                        Box bx_xhi(gbx & domain);
-                        bx_xhi.setSmall(0,dom_hi.x-width+1);
-                        bx_xhi.setBig(0,dom_hi.x+ng_vect[0]);
-
-                        ParallelFor(bx_xhi, ncomp, [=] AMREX_GPU_DEVICE (int i, int j, int k, int n)
-                        {
-                            dest_arr(i,j,k,icomp+n) = dest_arr(dom_hi.x-width,j,k,icomp+n);
-                        });
-                    } // bx
-                } // x-faces
-
-                // y-faces
+                    int jj = std::max(j , dom_lo.y);
+                        jj = std::min(jj, dom_hi.y);
+                    dest_arr(i,j,k,icomp+n) = dest_arr(dom_lo.x+width,jj,k,icomp+n);
+                });
[…]
