### 2023-04-06 a5a0fb2
Fix norm with periodicity.
