### 2023-04-07 erf-model/ERF#1041 @adam-m-jcbs
[WIP] JOSS Reviewer 2 Omnibus Review Issue
Hello ERF team!  I am serving as a reviewer for this JOSS submission:
https://github.com/openjournals/joss-reviews/issues/5202

I'm posting this issue as a single "omnibus" overview of my review.  Please do let me know if you would prefer any topics be broken out into dedicated issues.

Resolution of this issue will indicate that this submission satisfied all JOSS criteria for publication and that I recommend it for publication.

Please see my review checklist here for all details and items I've verified as resolved:
https://github.com/openjournals/joss-reviews/issues/5202#issuecomment-1477008548

This issue is currently a WIP!  Please bear with me as I work to complete the review, but I wanted to go ahead and get started.  I will remove the [WIP] tag and update here when this issue's ready for review.

### Minor issue: copyright
The copyright date in the license is 2022.  Is this intended?  There may be no need or desire to update the copyright date, but I wanted to note it and affirm that 2022 is correct (or that it will be updated).

- [ ] Verify copyright date
