+++ Source/Utils/MomentumToVelocity.cpp
@@ -25,10 +25,7 @@ MomentumToVelocity( BoxArray& grids_to_evolve,
     for ( MFIter mfi(cons_in,TilingIfNotGPU()); mfi.isValid(); ++mfi)
     {
         // We need velocity in the interior ghost cells (init == real)
-        const Box& valid_bx = mfi.validbox();
-
-        // Construct intersection of current tilebox and valid region for updating
-        Box bx = mfi.tilebox() & valid_bx;
+        Box bx = mfi.tilebox();
 
         const Box& tbx = surroundingNodes(bx,0);
         const Box& tby = surroundingNodes(bx,1);
