+++ Source/Utils/InteriorGhostCells.cpp
@@ -85,9 +85,6 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
     Real F1 = 1./(10.*delta_t);
     Real F2 = 1./(50.*delta_t);
 
-    // Copy of the Domain
-    Box domain = geom.Domain();
-
     // Compute interpolation factors
     Real dT = bdy_time_interval;
     int n_time = static_cast<int>(time / dT);
@@ -99,9 +96,6 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
     FArrayBox V_xlo, V_xhi, V_ylo, V_yhi;
     FArrayBox R_xlo, R_xhi, R_ylo, R_yhi;
     FArrayBox T_xlo, T_xhi, T_ylo, T_yhi;
-
-    // 4 halo boxes
-    Box bx_xlo, bx_xhi, bx_ylo, bx_yhi;
 
     // Variable index map
     Vector<int> var_map = {Vars::xvel, Vars::yvel, Vars::cons, Vars::cons};
@@ -115,6 +109,7 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
     {
         // Convert the domain to the ixtype of the variable
         int var_idx = var_map[ivar];
+        Box domain  = geom.Domain();
         domain.convert(S_data[var_idx].boxArray().ixType());
 
         // Grown domain to get the 4 halo boxes w/ ghost cells
@@ -122,6 +117,7 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
         Box gdom(domain); gdom.grow(ng_vect);
 
         // 4 halo boxes w/ interior ghost cells
+        Box bx_xlo, bx_xhi, bx_ylo, bx_yhi;
         compute_interior_ghost_bxs_xy(gdom, domain, width,
                                       bx_xlo, bx_xhi,
                                       bx_ylo, bx_yhi,
@@ -129,17 +125,17 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
 
         // Size the FABs
         if (ivar  == WRFBdyVars::U) {
-            U_xlo.resize(bx_xlo,2); U_xhi.resize(bx_xhi,2);
-            U_ylo.resize(bx_ylo,2); U_yhi.resize(bx_yhi,2);
+            U_xlo.resize(bx_xlo,1); U_xhi.resize(bx_xhi,1);
+            U_ylo.resize(bx_ylo,1); U_yhi.resize(bx_yhi,1);
         } else if (ivar  == WRFBdyVars::V) {
-            V_xlo.resize(bx_xlo,2); V_xhi.resize(bx_xhi,2);
-            V_ylo.resize(bx_ylo,2); V_yhi.resize(bx_yhi,2);
+            V_xlo.resize(bx_xlo,1); V_xhi.resize(bx_xhi,1);
+            V_ylo.resize(bx_ylo,1); V_yhi.resize(bx_yhi,1);
         } else if (ivar  == WRFBdyVars::R) {
-            R_xlo.resize(bx_xlo,2); R_xhi.resize(bx_xhi,2);
-            R_ylo.resize(bx_ylo,2); R_yhi.resize(bx_yhi,2);
+            R_xlo.resize(bx_xlo,1); R_xhi.resize(bx_xhi,1);
+            R_ylo.resize(bx_ylo,1); R_yhi.resize(bx_yhi,1);
         } else {
-            T_xlo.resize(bx_xlo,2); T_xhi.resize(bx_xhi,2);
-            T_ylo.resize(bx_ylo,2); T_yhi.resize(bx_yhi,2);
[…]
