### 2023-04-08 73fef82
shorten dt for MOST reg test and clean up unused variables
