+++ Source/Diffusion/DiffusionSrcForState_N.cpp
@@ -9,7 +9,6 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain,
                         int start_comp, int num_comp,
                         const Array4<const Real>& u,
                         const Array4<const Real>& v,
-                        const Array4<const Real>& w,
                         const Array4<const Real>& cell_data,
                         const Array4<const Real>& cell_prim,
                         const Array4<Real>& cell_rhs,
@@ -21,8 +20,6 @@ DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain,
                         const Array4<const Real>& mf_m,
                         const Array4<const Real>& mf_u,
                         const Array4<const Real>& mf_v,
-                              Array4<      Real>& hfx_x,
-                              Array4<      Real>& hfx_y,
                               Array4<      Real>& hfx_z,
                               Array4<      Real>& diss,
                         const Array4<const Real>& mu_turb,
