### 2023-04-08 4cb6afe
shorten dt for MOST reg test and clean up unused variables (#1043)
