### 2023-04-08 5486b09
Allow option to not include lagged rt when doing moving terrain
