### 2023-04-10 b089c4f
fix OMP directive placement
