### 2023-04-13 4c119df
Use MPI by default with the CUDA CI build.
