+++ .github/workflows/cuda-ci.yml
@@ -51,7 +51,7 @@ jobs:
           cmake -Bbuild-${{matrix.cuda_pkg}} \
             -DCMAKE_BUILD_TYPE:STRING=RelWithDebInfo \
             -DERF_DIM:STRING=3 \
-            -DERF_ENABLE_MPI:BOOL=OFF \
+            -DERF_ENABLE_MPI:BOOL=ON \
             -DERF_ENABLE_CUDA:BOOL=ON .
           cmake --build build-${{matrix.cuda_pkg}} -- -j $(nproc)
 
