### 2023-04-13 0ad91c5
Update docs for minimum dependencies.
