### 2023-04-13 36bb0a9
add cmake scripts with hip and sycl.
