### 2023-04-13 7be2a57
Adds lots of Doxygen annotations (#1048)
* add docstrings

* add docstrings

* docstrings for the rest of Initialization.

* Saving more progress on docstrings ...

* more docs ...

* Fix style

---------

Co-authored-by: asalmgren <asalmgren@lbl.gov>
