+++ Source/ERF.cpp
@@ -1595,6 +1595,8 @@ ERF::ERF (const amrex::RealBox& rb, int max_level_in,
         nsubsteps[lev] = MaxRefRatio(lev-1);
     }
 
+    grids_to_evolve.resize(nlevs_max);
+
     t_new.resize(nlevs_max, 0.0);
     t_old.resize(nlevs_max, -1.e100);
     dt.resize(nlevs_max, 1.e100);
