### 2023-04-14 e26fc8e
Cut out last cell and account for case with only set zones.
