+++ Source/TimeIntegration/TI_slow_rhs_fun.H
@@ -195,7 +195,7 @@
         // Populate RHS for relaxation zones
         if (init_type == "real" && level == 0)
             compute_interior_ghost_RHS(bdy_time_interval, old_stage_time, slow_dt,
-                                       wrfbdy_width, wrfbdy_set_width, fine_geom,
+                                       wrfbdy_width-1, wrfbdy_set_width, fine_geom,
                                        S_rhs, S_data, bdy_data_xlo, bdy_data_xhi,
                                        bdy_data_ylo, bdy_data_yhi);
 #endif
