+++ Source/Utils/Utils.H
@@ -42,6 +42,7 @@ void VelocityToMomentum (amrex::BoxArray& grids_to_evolve,
 void compute_interior_ghost_bxs_xy(const amrex::Box& bx,
                                    const amrex::Box& domain,
                                    const int& width,
+                                   const int& set_width,
                                    amrex::Box& bx_xlo,
                                    amrex::Box& bx_xhi,
                                    amrex::Box& bx_ylo,
