+++ Source/IO/NCCheckpoint.cpp
@@ -87,7 +87,7 @@ ERF::WriteNCCheckpointFile () const
        for (auto lev{0}; lev <= finest_level; ++lev) {
            auto box_array = boxArray(lev);
            for (int nb(0); nb < box_array.size(); ++nb) {
-              long unsigned int nbb = static_cast<long unsigned int>(nb);
+              auto nbb = static_cast<long unsigned int>(nb);
               auto box = box_array[nb];
               ncf.var(lo_names[lev][nb] ).put(box.smallEnd().begin(), {nbb, 0}, {1, AMREX_SPACEDIM});
               ncf.var(hi_names[lev][nb] ).put(box.bigEnd().begin()  , {nbb, 0}, {1, AMREX_SPACEDIM});
