### 2023-04-14 8146dfd
Run regtests for HIP CI test.
