### 2023-04-14 cbafb67
small clean-up and update to docs re SYCL
