### 2023-04-14 05cbae9
small clean-up and update to docs re SYCL (#1057)
