+++ Source/Utils/InteriorGhostCells.cpp
@@ -210,7 +210,8 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
         const auto& bdatyhi_np1 = bdy_data_yhi[n_time+1][ivar].const_array();
 
         // Populate with interpolation (protect from ghost cells)
-        amrex::ParallelFor(bx_xlo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+        amrex::ParallelFor(bx_xlo, bx_xhi,
+        [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
         {
             int ii = std::max(i , dom_lo.x);
                 ii = std::min(ii, dom_lo.x+width-1);
@@ -218,8 +219,8 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
                 jj = std::min(jj, dom_hi.y);
             arr_xlo(i,j,k) = oma   * bdatxlo_n  (ii,jj,k,0)
                            + alpha * bdatxlo_np1(ii,jj,k,0);
-        });
-        amrex::ParallelFor(bx_xhi, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+        },
+        [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
         {
             int ii = std::max(i , dom_hi.x-width+1);
                 ii = std::min(ii, dom_hi.x);
@@ -228,7 +229,9 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
             arr_xhi(i,j,k) = oma   * bdatxhi_n  (ii,jj,k,0)
                            + alpha * bdatxhi_np1(ii,jj,k,0);
         });
-        amrex::ParallelFor(bx_ylo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+
+        amrex::ParallelFor(bx_ylo, bx_yhi,
+        [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
         {
             int ii = std::max(i , dom_lo.x);
                 ii = std::min(ii, dom_hi.x);
@@ -236,8 +239,8 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
                 jj = std::min(jj, dom_lo.y+width-1);
             arr_ylo(i,j,k) = oma   * bdatylo_n  (ii,jj,k,0)
                            + alpha * bdatylo_np1(ii,jj,k,0);
-        });
-        amrex::ParallelFor(bx_yhi, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+        },
+        [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
         {
             int ii = std::max(i , dom_lo.x);
                 ii = std::min(ii, dom_hi.x);
@@ -280,22 +283,25 @@ compute_interior_ghost_RHS(const Real& bdy_time_interval,
             arr_xlo = U_xlo.array(); arr_xhi = U_xhi.array();
             arr_ylo = U_ylo.array(); arr_yhi = U_yhi.array();
 
-            amrex::ParallelFor(bx_xlo, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+            amrex::ParallelFor(bx_xlo, bx_xhi,
+            [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
             {
                 Real rho_interp = 0.5 * ( rarr_xlo(i-1,j,k) + rarr_xlo(i,j,k) );
                 arr_xlo(i,j,k) *= rho_interp;
-            });
-            amrex::ParallelFor(bx_xhi, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+            },
+            [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
             {
[…]
