### 2023-04-14 f6e71c8
Fix an issue with WPS compilation caused by clang-tidy autofixes.
