### 2023-04-14 c23b711
Fix wps test compile (from clang-tidy autofixes) (#1058)
* Fix an issue with WPS compilation caused by clang-tidy autofixes.

* Add missing reference to previous fix.
