### 2023-04-14 7965743
Try running HIP regtests without MPI.
