### 2023-04-15 addec82
Fix missed copy for terrain stress.
