+++ Source/TimeIntegration/TI_no_substep_fun.H
@@ -18,11 +18,10 @@
         {
             for ( MFIter mfi(S_sum[IntVar::cons],TilingIfNotGPU()); mfi.isValid(); ++mfi)
             {
-                const Box bx = mfi.tilebox() & grids_to_evolve[level][mfi.index()];
-
-                Box tbx = surroundingNodes((mfi.tilebox() & grids_to_evolve[level][mfi.index()]),0);
-                Box tby = surroundingNodes((mfi.tilebox() & grids_to_evolve[level][mfi.index()]),1);
-                Box tbz = surroundingNodes((mfi.tilebox() & grids_to_evolve[level][mfi.index()]),2);
+                const Box bx = mfi.tilebox();
+                Box tbx = surroundingNodes(mfi.tilebox(),0);
+                Box tby = surroundingNodes(mfi.tilebox(),1);
+                Box tbz = surroundingNodes(mfi.tilebox(),2);
 
                 Vector<Array4<Real> >  ssum_h(n_data);
                 Vector<Array4<Real> >  sold_h(n_data);
@@ -61,7 +60,7 @@
                     [=] AMREX_GPU_DEVICE (int i, int j, int k, int nn) {
                         const int n = scomp_fast[IntVar::cons] + nn;
                         ssum[IntVar::cons](i,j,k,n) = dJ_old(i,j,k) *  sold[IntVar::cons](i,j,k,n)
-                                          + slow_dt                 * fslow[IntVar::cons](i,j,k,n);
+                                                          + slow_dt * fslow[IntVar::cons](i,j,k,n);
                         ssum[IntVar::cons](i,j,k,n) /= dJ_new(i,j,k);
                     });
 
@@ -71,14 +70,14 @@
                         Real dJ_old_xface = 0.5 * (dJ_old(i,j,k) + dJ_old(i-1,j,k));
                         Real dJ_new_xface = 0.5 * (dJ_new(i,j,k) + dJ_new(i-1,j,k));
                         ssum[IntVar::xmom](i,j,k) = dJ_old_xface *  sold[IntVar::xmom](i,j,k)
-                                        + slow_dt                * fslow[IntVar::xmom](i,j,k);
+                                                       + slow_dt * fslow[IntVar::xmom](i,j,k);
                         ssum[IntVar::xmom](i,j,k) /= dJ_new_xface;
                     },
                     [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept {
                         Real dJ_old_yface = 0.5 * (dJ_old(i,j,k) + dJ_old(i,j-1,k));
                         Real dJ_new_yface = 0.5 * (dJ_new(i,j,k) + dJ_new(i,j-1,k));
                         ssum[IntVar::ymom](i,j,k) = dJ_old_yface *  sold[IntVar::ymom](i,j,k)
-                                        + slow_dt                * fslow[IntVar::ymom](i,j,k);
+                                                       + slow_dt * fslow[IntVar::ymom](i,j,k);
                         ssum[IntVar::ymom](i,j,k) /= dJ_new_yface;
                     });
                     ParallelFor(tbz,
@@ -90,13 +89,12 @@
                             Real dJ_old_zface = 0.5 * (dJ_old(i,j,k) + dJ_old(i,j,k-1));
                             Real dJ_new_zface = 0.5 * (dJ_new(i,j,k) + dJ_new(i,j,k-1));
                             ssum[IntVar::zmom](i,j,k) = dJ_old_zface *  sold[IntVar::zmom](i,j,k)
-                                            + slow_dt                * fslow[IntVar::zmom](i,j,k);
+                                                           + slow_dt * fslow[IntVar::zmom](i,j,k);
                             ssum[IntVar::zmom](i,j,k) /= dJ_new_zface;
                         }
[…]
