+++ Source/TimeIntegration/ERF_slow_rhs_post.cpp
@@ -112,8 +112,10 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
 
         const Box& valid_bx = grids_to_evolve[mfi.index()];
 
+        const Box& tbx = mfi.tilebox();
+
         // Construct intersection of current tilebox and valid region for updating
-        Box bx = mfi.tilebox() & valid_bx;
+        Box bx = tbx & valid_bx;
 
         const Array4<      Real> & old_cons   = S_old[IntVar::cons].array(mfi);
         const Array4<      Real> & cell_rhs   = S_rhs[IntVar::cons].array(mfi);
@@ -224,7 +226,7 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
                                            mu_turb, solverChoice, tm_arr, grav_gpu, bc_ptr);
                 }
                 if (l_use_ndiff) {
-                    NumericalDiffusion(valid_bx, start_comp, num_comp, dt, solverChoice,
+                    NumericalDiffusion(bx, start_comp, num_comp, dt, solverChoice,
                                        new_cons, cell_rhs, mf_u, mf_v, false, false);
                 }
             }
@@ -247,7 +249,7 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
                                            mu_turb, solverChoice, tm_arr, grav_gpu, bc_ptr);
                 }
                 if (l_use_ndiff) {
-                    NumericalDiffusion(valid_bx, start_comp, num_comp, dt, solverChoice,
+                    NumericalDiffusion(bx, start_comp, num_comp, dt, solverChoice,
                                        new_cons, cell_rhs, mf_u, mf_v, false, false);
                 }
             }
@@ -269,11 +271,14 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
                                        mu_turb, solverChoice, tm_arr, grav_gpu, bc_ptr);
             }
             if (l_use_ndiff) {
-                NumericalDiffusion(valid_bx, start_comp, num_comp, dt, solverChoice,
+                NumericalDiffusion(bx, start_comp, num_comp, dt, solverChoice,
                                    new_cons, cell_rhs, mf_u, mf_v, false, false);
             }
         }
 
+        // NOTE: Computing the RHS is done over bx (union w/ grids to evolve).
+        //       However, the update is over tbx (no union). The interior ghost
+        //       cells have their RHS populated already.
 
         // This updates just the "slow" conserved variables
         {
@@ -283,7 +288,7 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
         {
             auto const& src_arr = source.const_array(mfi);
             num_comp = S_data[IntVar::cons].nComp() - start_comp;
-            ParallelFor(bx, num_comp,
+            ParallelFor(tbx, num_comp,
             [=] AMREX_GPU_DEVICE (int i, int j, int k, int nn) noexcept {
                 const int n = start_comp + nn;
                 // NOTE: we don't include additional source terms when terrain is moving
@@ -294,7 +299,7 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
             if (l_use_deardorff) {
               start_comp = RhoKE_comp;
               num_comp   = 1;
-              ParallelFor(bx, num_comp,
+              ParallelFor(tbx, num_comp,
               [=] AMREX_GPU_DEVICE (int i, int j, int k, int nn) noexcept {
                 const int n = start_comp + nn;
[…]
