### 2023-04-17 06dbd69
small fixes to quiet compiler warnings, plus adding cmake_hip_crusher.sh
