### 2023-04-18 360d5a4
small fixes to quiet compiler warnings, plus adding cmake_hip_crusher.sh (#1062)
