### 2023-04-18 erf-model/ERF#1010 @pratikvn
I see that some basic doxygen generated API documentation has been added. I think it would also be helpful for users to have some class and function documentation in the source files which is then also in the API documentation, but I understand that requires more substantial work, so I think it is okay to not have it as a requirement for the paper.

I think it would be useful to have atleast a couple of examples. You could even just copy and repurpose your most representative regression tests, and combine that with the theory you have on your webpage, showing how to run it and what is the expected output.
