### 2023-04-20 #5202 @adam-m-jcbs
@kthyng yes, I was recently out sick last week and that has led to many things coming up.

I apologize for the slip. I can continue to work to complete this review and would like to, but have been disrupted and do not want to slow publication.

Being extremely conservative, I estimate I can complete this review at the latest by Friday May 19. I'll plan on sooner, but want to set realistic expectations.

I understand if that's too great a delay and an alternative reviewer is needed. I have prepared the environment for and started some of the functional review.

Just let me know how JOSS would like to move forward. I apologize to the editors and authors for this delay and disruption. For now, I'll plan to continue working on the review as able, the as able just became less than it was when I originally agreed to do the review.
