### 2023-04-21 #5202 @kthyng
@yhtang or @navidcy — would either of you be interested in reviewing this submission to JOSS? The review would take place in this github issue and issues created in the software repo itself. This review is underway but I'm recruiting a replacement reviewer, if possible.
