+++ Source/Diffusion/DiffusionSrcForState_N.cpp
@@ -4,6 +4,34 @@
 
 using namespace amrex;
 
+/**
+ * Function for computing the scalar RHS for diffusion operator without terrain.
+ *
+ * @param bx cell center box to loop over
+ * @param domain box of the whole domain
+ * @param start_comp starting component index
+ * @param num_comp number of components
+ * @param u velocity in x-dir
+ * @param v velocity in y-dir
+ * @param cell_data conserved cell center vars
+ * @param cell_prim primitive cell center vars
+ * @param cell_rhs RHS for cell center vars
+ * @param xflux flux in x-dir
+ * @param yflux flux in y-dir
+ * @param zflux flux in z-dir
+ * @param cellSizeInv inverse cell size array
+ * @param SmnSmn_a strain rate magnitude
+ * @param mf_m map factor at cell center
+ * @param mf_u map factor at x-face
+ * @param mf_v map factor at y-face
+ * @param hfx_z heat flux in z-dir
+ * @param diss dissipation of TKE
+ * @param mu_turb turbulent viscosity
+ * @param solverChoice container of solver params
+ * @param tm_arr theta mean array
+ * @param grav_gpu gravity vector
+ * @param bc_ptr container with boundary conditions
+ */
 void
 DiffusionSrcForState_N (const amrex::Box& bx, const amrex::Box& domain,
                         int start_comp, int num_comp,
