+++ Source/Diffusion/PBLModels.cpp
@@ -2,6 +2,17 @@
 #include "DirectionSelector.H"
 #include "Diffusion.H"
 
+/**
+ * Function to compute turbulent viscosity with PBL.
+ *
+ * @param xvel velocity in x-dir
+ * @param yvel velocity in y-dir
+ * @param cons_in cell center conserved quantities
+ * @param eddyViscosity holds turbulent viscosity
+ * @param geom problem geometry
+ * @param solverChoic container with solver parameters
+ * @param most pointer to Monin-Obukhov class if instantiated
+ */
 void
 ComputeTurbulentViscosityPBL (const amrex::MultiFab& xvel,
                               const amrex::MultiFab& yvel,
