+++ Source/main.cpp
@@ -14,10 +14,12 @@ std::string inputs_name;
 
 using namespace amrex;
 
-// Set the refine_grid_layout flags to (1,1,0) by default
-// since the ERF default is different from the amrex default (1,1,1)
-// Also set max_grid_size to very large since the only reason for
-// chopping grids is if Nprocs > Ngrids
+/**
+ * Function to set the refine_grid_layout flags to (1,1,0) by default
+ * since the ERF default is different from the amrex default (1,1,1)
+ * Also set max_grid_size to very large since the only reason for
+ * chopping grids is if Nprocs > Ngrids
+*/
 void add_par () {
    ParmParse pp("amr");
 
@@ -39,6 +41,10 @@ void add_par () {
    pp.add("n_error_buf",0);
 }
 
+/**
+ * Main driver -- creates the ERF object, calls ERF.InitData() and ERF.Evolve()
+ * Also includes the multiblock interface in the case where there is more than one ERF object
+*/
 int main(int argc, char* argv[])
 {
     // Check to see if the command line contains --describe
