+++ Source/TimeIntegration/ERF_slow_rhs_post.cpp
@@ -12,38 +12,6 @@
 
 using namespace amrex;
 
-/**
- * Function for computing the slow RHS for the evolution equations for the scalars other than density or potential temperature
- *
- * @param[in]  level level of resolution
- * @param[in]  dt    slow time step
- * @param[in]  grids_to_evolve the region in the domain excluding the relaxation and specified zones
- * @param[out]  S_rhs RHS computed here
- * @param[in]  S_old solution at start of time step
- * @param[in]  S_new solution at end of current RK stage
- * @param[in]  S_data current solution
- * @param[in]  S_prim primitive variables (i.e. conserved variables divided by density)
- * @param[in]  S_scratch scratch space
- * @param[in]  xvel x-component of velocity
- * @param[in]  yvel y-component of velocity
- * @param[in]  zvel z-component of velocity
- * @param[in] source source terms for conserved variables
- * @param[in] SmnSmn SmnSmn
- * @param[in] eddyDiffs diffusion coefficients for LES turbulence models
- * @param[in] Hfx3
- * @param[in] Diss
- * @param[in]  geom   Container for geometric informaiton
- * @param[in]  solverChoice  Container for solver parameters
- * @param[in]  most  Pointer to MOST class for Monin-Obukhov Similarity Theory boundary condition
- * @param[in]  domain_bcs_type_d device vector for domain boundary conditions
- * @param[in] z_phys_nd height coordinate at nodes
- * @param[in] detJ     Jacobian of the metric transformation at start of time step (= 1 if use_terrain is false)
- * @param[in] detJ_new Jacobian of the metric transformation at new RK stage time (= 1 if use_terrain is false)
- * @param[in] mapfac_m map factor at cell centers
- * @param[in] mapfac_u map factor at x-faces
- * @param[in] mapfac_v map factor at y-faces
- */
-
 void erf_slow_rhs_post (int /*level*/, Real dt,
                         BoxArray& grids_to_evolve,
                         Vector<MultiFab>& S_rhs,
@@ -64,8 +32,8 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
                         std::unique_ptr<ABLMost>& most,
                         const Gpu::DeviceVector<amrex::BCRec>& domain_bcs_type_d,
                         std::unique_ptr<MultiFab>& z_phys_nd,
-                        std::unique_ptr<MultiFab>& detJ,
-                        std::unique_ptr<MultiFab>& detJ_new,
+                        std::unique_ptr<MultiFab>& dJ,
+                        std::unique_ptr<MultiFab>& dJ_new,
                         std::unique_ptr<MultiFab>& mapfac_m,
                         std::unique_ptr<MultiFab>& mapfac_u,
                         std::unique_ptr<MultiFab>& mapfac_v)
@@ -173,9 +141,9 @@ void erf_slow_rhs_post (int /*level*/, Real dt,
         const Array4<Real const>& mu_turb = l_use_turb ? eddyDiffs->const_array(mfi) : Array4<const Real>{};
 
         // Metric terms
-        const Array4<const Real>& z_nd         = l_use_terrain    ? z_phys_nd->const_array(mfi) : Array4<const Real>{};
-        const Array4<const Real>& detJ_arr     = l_use_terrain    ? detJ->const_array(mfi)        : Array4<const Real>{};
-        const Array4<const Real>& detJ_new_arr = l_moving_terrain ? detJ_new->const_array(mfi)    : Array4<const Real>{};
+        const Array4<const Real>& z_nd     = l_use_terrain    ? z_phys_nd->const_array(mfi) : Array4<const Real>{};
[…]
