+++ Source/TimeIntegration/TI_fast_rhs_fun.H
@@ -1,6 +1,3 @@
-/**
- *  Wrapper for calling the routine that creates the fast RHS
- */
 auto fast_rhs_fun = [&](int fast_step, int n_sub,
                             Vector<MultiFab>& S_slow_rhs,
                             const Vector<MultiFab>& S_old,
