+++ Source/TimeIntegration/ERF_ComputeTimestep.cpp
@@ -3,10 +3,7 @@
 
 using namespace amrex;
 
-/**
- * Function that calls estTimeStep for each level
- *
- */
+// a wrapper for estTimeStep
 void
 ERF::ComputeDt ()
 {
@@ -39,12 +36,6 @@ ERF::ComputeDt ()
     }
 }
 
-/**
- * Function that calls estTimeStep for each level
- *
- * @param[in] level level of refinement (coarsest level i 0)
- * @param[out] dt_fast_ratio ratio of slow to fast time step
- */
 Real
 ERF::estTimeStep(int level, long& dt_fast_ratio) const
 {
