+++ Source/TimeIntegration/ERF_TimeStepping.cpp
@@ -4,8 +4,15 @@
 
 using namespace amrex;
 
-// Advance a level by dt
-// includes a recursive call for finer levels
+/**
+ * Function that coordinates the evolution across levels -- this calls Advance to do the
+ * actual advance at this level,  then recursively calls itself at finer levels
+ *
+ * @param[in] lev level of refinement (coarsest level is 0)
+ * @param[in] time start time for time advance
+ * @param[in] iteration time step counter
+ */
+
 void
 ERF::timeStep (int lev, Real time, int iteration)
 {
@@ -73,7 +80,15 @@ ERF::timeStep (int lev, Real time, int iteration)
     }
 }
 
-// advance a single level for a single time step
+/**
+ * Function that advances the solution at one level for a single time step --
+ * this does some preliminaries then calls erf_advance
+ *
+ * @param[in] lev level of refinement (coarsest level is 0)
+ * @param[in] time start time for time advance
+ * @param[in] dt_lev time step for this time advance
+ */
+
 void
 ERF::Advance (int lev, Real time, Real dt_lev, int /*iteration*/, int /*ncycle*/)
 {
