+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -17,6 +17,57 @@
 
 using namespace amrex;
 
+/**
+ * Function for computing the slow RHS for the evolution equations for the density, potential temperature and momentum.
+ *
+ * @param[in]  level level of resolution
+ * @param[in]  nrk   which RK stage
+ * @param[in]  dt    slow time step
+ * @param[in]  grids_to_evolve the region in the domain excluding the relaxation and specified zones
+ * @param[out]  S_rhs RHS computed here
+ * @param[in]  S_data current solution
+ * @param[in]  S_prim primitive variables (i.e. conserved variables divided by density)
+ * @param[in]  S_scratch scratch space
+ * @param[in]  xvel x-component of velocity
+ * @param[in]  yvel y-component of velocity
+ * @param[in]  zvel z-component of velocity
+ * @param[in]  qv   water vapor
+ * @param[in]  z_t_ mf rate of change of grid height -- only relevant for moving terrain
+ * @param[in] Omega component of the momentum normal to the z-coordinate surface
+ * @param[in] source source terms for conserved variables
+ * @param[in] buoyancy buoyancy source term
+ * @param[in] Tau11 tau_11 component of stress tensor
+ * @param[in] Tau22 tau_22 component of stress tensor
+ * @param[in] Tau33 tau_33 component of stress tensor
+ * @param[in] Tau12 tau_12 component of stress tensor
+ * @param[in] Tau12 tau_13 component of stress tensor
+ * @param[in] Tau21 tau_21 component of stress tensor
+ * @param[in] Tau23 tau_23 component of stress tensor
+ * @param[in] Tau31 tau_31 component of stress tensor
+ * @param[in] Tau32 tau_32 component of stress tensor
+ * @param[in] SmnSmn SmnSmn
+ * @param[in] eddyDiffs diffusion coefficients for LES turbulence models
+ * @param[in] Hfx3
+ * @param[in] Diss
+ * @param[in]  geom   Container for geometric informaiton
+ * @param[in]  solverChoice  Container for solver parameters
+ * @param[in]  most  Pointer to MOST class for Monin-Obukhov Similarity Theory boundary condition
+ * @param[in]  domain_bcs_type_d device vector for domain boundary conditions
+ * @param[in]  domain_bcs_type     host vector for domain boundary conditions
+ * @param[in]  domain_bcs_type     host vector for domain boundary conditions
+ * @param[in] z_phys_nd height coordinate at nodes
+ * @param[in] detJ Jacobian of the metric transformation (= 1 if use_terrain is false)
+ * @param[in]  p0     Reference (hydrostatically stratified) pressure
+ * @param[in] mapfac_m map factor at cell centers
+ * @param[in] mapfac_u map factor at x-faces
+ * @param[in] mapfac_v map factor at y-faces
+ * @param[in] dptr_rayleigh_tau  strength of Rayleigh damping
+ * @param[in] dptr_rayleigh_ubar reference value for x-velocity used to define Rayleigh damping
+ * @param[in] dptr_rayleigh_vbar reference value for y-velocity used to define Rayleigh damping
+ * @param[in] dptr_rayleigh_wbar reference value for z-velocity used to define Rayleigh damping
+ * @param[in] dptr_rayleigh_thetabar reference value for potential temperature used to define Rayleigh damping
+ */
+
 void erf_slow_rhs_pre (int /*level*/, int nrk,
                        amrex::Real dt,
                        BoxArray& grids_to_evolve,
[…]
