+++ Source/TimeIntegration/NEW/ERF_MRI.H
@@ -1,0 +1,385 @@
+#ifndef ERF_MRI_H
+#define ERF_MRI_H
+#include <AMReX_REAL.H>
+#include <AMReX_Vector.H>
+#include <AMReX_ParmParse.H>
+#include <AMReX_IntegratorBase.H>
+#include <TimeIntegration.H>
+#include <functional>
+
+template<class T>
+class MRISplitIntegrator : public amrex::IntegratorBase<T>
+{
+private:
+   /**
+    * \brief rhs is the right-hand-side function the integrator will use.
+    */
+    std::function<void(T&, const T&,     const amrex::Real, const amrex::Real     )> rhs;
+    std::function<void(T&, T&, T&, const amrex::Real, const amrex::Real, const amrex::Real, const int)> slow_rhs_pre;
+    std::function<void(T&, T&, T&, T&, T&, const amrex::Real, const amrex::Real, const amrex::Real     )> slow_rhs_post;
+    std::function<void(int, int, T&, const T&, T&, T&, T&, const amrex::Real, const amrex::Real,
+                                                           const amrex::Real, const amrex::Real)> fast_rhs;
+
+   /**
+    * \brief Integrator timestep size (Real)
+    */
+    amrex::Real timestep;
+
+   /**
+    * \brief The ratio of slow timestep size / fast timestep size (int)
+    */
+    int slow_fast_timestep_ratio = 0;
+
+   /**
+    * \brief Should we not do acoustic substepping
+    */
+    int no_substepping;
+
+   /**
+    * \brief Do we follow the recommendation to only perform a single substep in the first RK stage
+    */
+    int force_stage1_single_substep;
+
+   /**
+    * \brief The  pre_update function is called by the integrator on stage data before using it to evaluate a right-hand side.
+    * \brief The post_update function is called by the integrator on stage data at the end of the stage
+    */
+    std::function<void (T&, int)> pre_update;
+    std::function<void (T&, amrex::Real, int, int)> post_update;
+    std::function<void (T&, T&, T&, amrex::Real, amrex::Real)>   no_substep;
+
+
+    amrex::Vector<std::unique_ptr<T> > T_store;
+    T* S_sum;
+    T* S_scratch;
+    T* F_slow;
+
+    void initialize_data (const T& S_data)
+    {
+        // TODO: We can optimize memory by making the cell-centered part of S_sum, S_scratch and F_pert
+        //       have only 2 components, not Cons::NumVars components
+        const bool include_ghost = true;
+        amrex::IntegratorOps<T>::CreateLike(T_store, S_data, include_ghost);
+        S_sum = T_store[0].get();
+        amrex::IntegratorOps<T>::CreateLike(T_store, S_data, include_ghost);
+        S_scratch = T_store[1].get();
+        amrex::IntegratorOps<T>::CreateLike(T_store, S_data, include_ghost);
+        F_slow = T_store[2].get();
+    }
+
+public:
+    MRISplitIntegrator () = default;
+
+    MRISplitIntegrator (const T& S_data)
+    {
+        initialize_data(S_data);
+    }
+
+    void initialize (const T& S_data) override
+    {
+        initialize_data(S_data);
+    }
+
+    ~MRISplitIntegrator () override = default;
+
+    // Declare a default move constructor so we ensure the destructor is
+    // not called when we return an object of this class by value
+    MRISplitIntegrator(MRISplitIntegrator&&)  noexcept = default;
+
+    // Declare a default move assignment operator
+    MRISplitIntegrator& operator=(MRISplitIntegrator&& other)  noexcept = default;
+
+    // Delete the copy constructor and copy assignment operators because
+    // the integrator allocates internal memory that is best initialized
+    // from scratch when needed instead of making a copy.
+
+    // Delete the copy constructor
[…]

[…]
