+++ Source/TimeIntegration/NEW/ERF_make_fast_coeffs.cpp
@@ -1,274 +1,0 @@
-#include <AMReX.H>
-#include <AMReX_MultiFab.H>
-#include <IndexDefines.H>
-#include <TimeIntegration.H>
-#include <prob_common.H>
-#include <TileNoZ.H>
-
-using namespace amrex;
-
-void make_fast_coeffs (int /*level*/,
-                       BoxArray& grids_to_evolve,
-                       MultiFab& fast_coeffs,
-                       Vector<MultiFab>& S_stage_data,                 // S_bar = S^n, S^* or S^**
-                       const MultiFab& S_stage_prim,
-                       const MultiFab& pi_stage,                       // Exner function evaluted at least stage
-                       const amrex::Geometry geom,
-                       const SolverChoice& solverChoice,
-                       std::unique_ptr<MultiFab>& detJ_cc,
-                       const MultiFab* r0, const MultiFab* pi0,
-                       amrex::Real dtau)
-{
-    BL_PROFILE_VAR("make_fast_coeffs()",make_fast_coeffs);
-
-    // beta_s = -1.0 : fully explicit
-    // beta_s =  1.0 : fully implicit
-    // Real beta_s = 0.1;
-    // HACK
-    Real beta_s = -1.0;
-    Real beta_2 = 0.5 * (1.0 + beta_s);  // multiplies implicit terms
-
-    Real c_v = solverChoice.c_p - R_d;
-
-    bool l_use_terrain    = solverChoice.use_terrain;
-
-    const Box domain(geom.Domain());
-    const GpuArray<Real, AMREX_SPACEDIM> dxInv = geom.InvCellSizeArray();
-
-    Real dzi = dxInv[2];
-
-    MultiFab coeff_A_mf(fast_coeffs, amrex::make_alias, 0, 1);
-    MultiFab coeff_B_mf(fast_coeffs, amrex::make_alias, 1, 1);
-    MultiFab coeff_C_mf(fast_coeffs, amrex::make_alias, 2, 1);
-    MultiFab coeff_P_mf(fast_coeffs, amrex::make_alias, 3, 1);
-    MultiFab coeff_Q_mf(fast_coeffs, amrex::make_alias, 4, 1);
-
-
-    // *************************************************************************
-    // Set gravity as a vector
-    const    Array<Real,AMREX_SPACEDIM> grav{0.0, 0.0, -solverChoice.gravity};
-    const GpuArray<Real,AMREX_SPACEDIM> grav_gpu{grav[0], grav[1], grav[2]};
-
-    // *************************************************************************
-    // Define updates in the current RK stage
-    // *************************************************************************
-#ifdef _OPENMP
-#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
-#endif
-    {
-
-    for ( MFIter mfi(S_stage_data[IntVar::cons],TileNoZ()); mfi.isValid(); ++mfi)
-    {
-        const Box& valid_bx = grids_to_evolve[mfi.index()];
-
-        // Construct intersection of current tilebox and valid region for updating
-        Box bx = mfi.tilebox() & valid_bx;
-
-        Box tbz = surroundingNodes(bx,2);
-
-        const Array4<const Real> & stage_cons = S_stage_data[IntVar::cons].const_array(mfi);
-        const Array4<const Real> & prim       = S_stage_prim.const_array(mfi);
-
[…]
