### 2023-04-24 0787ef5
More doxygen updates (#1066)
* more doxygenating

* Revert "more doxygenating"

This reverts commit 4bee66421cf99c6ecd3c99eae11bb9bbd1409344.

* Revert "Revert "more doxygenating""

This reverts commit 6fb65167732caefc4f8d82576feb914f63d6fb1a.

* Don't commit temp files
