+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -45,10 +45,10 @@ using namespace amrex;
  * @param[in] Tau23 tau_23 component of stress tensor
  * @param[in] Tau31 tau_31 component of stress tensor
  * @param[in] Tau32 tau_32 component of stress tensor
- * @param[in] SmnSmn SmnSmn
+ * @param[in] SmnSmn magnitude of strain rate
  * @param[in] eddyDiffs diffusion coefficients for LES turbulence models
- * @param[in] Hfx3
- * @param[in] Diss
+ * @param[in] Hfx3 z-dir heat flux
+ * @param[in] Diss dissipation of turbulent kinetic energy
  * @param[in]  geom   Container for geometric informaiton
  * @param[in]  solverChoice  Container for solver parameters
  * @param[in]  most  Pointer to MOST class for Monin-Obukhov Similarity Theory boundary condition
