+++ Source/Diffusion/ComputeStress_T.cpp
@@ -6,23 +6,23 @@ using namespace amrex;
 /**
  * Function for computing the stress with constant viscosity and with terrain.
  *
- * @param bxcc cell center box for tau_ii
- * @param tbxxy nodal xy box for tau_12
- * @param tbxxz nodal xz box for tau_13
- * @param tbxyz nodal yz box for tau_23
- * @param mu_eff constant molecular viscosity
- * @param tau11 hold 11 strain
- * @param tau22 hold 22 strain
- * @param tau33 hold 33 strain
- * @param tau12 hold 12 strain
- * @param tau13 hold 13 strain
- * @param tau21 hold 21 strain
- * @param tau23 hold 23 strain
- * @param tau31 hold 31 strain
- * @param tau32 hold 32 strain
- * @param er_arr expansion rate
- * @param z_nd nodal array of physical z heights
- * @param dxInv inverse cell size array
+ * @param[in]  bxcc cell center box for tau_ii
+ * @param[in]  tbxxy nodal xy box for tau_12
+ * @param[in]  tbxxz nodal xz box for tau_13
+ * @param[in]  tbxyz nodal yz box for tau_23
+ * @param[in]  mu_eff constant molecular viscosity
+ * @param[out] tau11 11 stress
+ * @param[out] tau22 22 stress
+ * @param[out] tau33 33 stress
+ * @param[out] tau12 12 stress
+ * @param[out] tau13 13 stress
+ * @param[out] tau21 21 stress
+ * @param[out] tau23 23 stress
+ * @param[out] tau31 31 stress
+ * @param[out] tau32 32 stress
+ * @param[in]  er_arr expansion rate
+ * @param[in]  z_nd nodal array of physical z heights
+ * @param[in]  dxInv inverse cell size array
  */
 void
 ComputeStressConsVisc_T(Box bxcc, Box tbxxy, Box tbxxz, Box tbxyz, Real mu_eff,
@@ -250,24 +250,24 @@ ComputeStressConsVisc_T(Box bxcc, Box tbxxy, Box tbxxz, Box tbxyz, Real mu_eff,
 /**
  * Function for computing the stress with constant viscosity and with terrain.
  *
- * @param bxcc cell center box for tau_ii
- * @param tbxxy nodal xy box for tau_12
- * @param tbxxz nodal xz box for tau_13
- * @param tbxyz nodal yz box for tau_23
- * @param mu_eff constant molecular viscosity
- * @param mu_turb variable turbulent viscosity
- * @param tau11 hold 11 strain
- * @param tau22 hold 22 strain
- * @param tau33 hold 33 strain
- * @param tau12 hold 12 strain
- * @param tau13 hold 13 strain
- * @param tau21 hold 21 strain
- * @param tau23 hold 23 strain
- * @param tau31 hold 31 strain
- * @param tau32 hold 32 strain
- * @param er_arr expansion rate
- * @param z_nd nodal array of physical z heights
- * @param dxInv inverse cell size array
+ * @param[in]  bxcc cell center box for tau_ii
+ * @param[in]  tbxxy nodal xy box for tau_12
+ * @param[in]  tbxxz nodal xz box for tau_13
+ * @param[in]  tbxyz nodal yz box for tau_23
+ * @param[in]  mu_eff constant molecular viscosity
+ * @param[in] mu_turb variable turbulent viscosity
+ * @param[out] tau11 11 stress
+ * @param[out] tau22 22 stress
+ * @param[out] tau33 33 stress
+ * @param[out] tau12 12 stress
+ * @param[out] tau13 13 stress
+ * @param[out] tau21 21 stress
+ * @param[out] tau23 23 stress
[…]
