+++ Source/Microphysics/IceFall.cpp
@@ -5,6 +5,9 @@
 
 using namespace amrex;
 
+/**
+ * Computes contributions to Microphysics and thermodynamic variables from falling cloud ice in each column.
+ */
 void Microphysics::IceFall() {
 
   Real dz  = m_geom.CellSize(2);
