+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -47,7 +47,7 @@ using namespace amrex;
  * @param[in] Tau32 tau_32 component of stress tensor
  * @param[in] SmnSmn strain rate magnitude
  * @param[in] eddyDiffs diffusion coefficients for LES turbulence models
- * @param[in] Hfx3 z-dir heat flux
+ * @param[in] Hfx3 heat flux in z-dir
  * @param[in] Diss dissipation of turbulent kinetic energy
  * @param[in]  geom   Container for geometric informaiton
  * @param[in]  solverChoice  Container for solver parameters
