+++ Source/BoundaryConditions/BoundaryConditions_zvel.cpp
@@ -4,16 +4,20 @@
 
 using namespace amrex;
 
-//
-// dest_arr is the Array4 to be filled
-// time is the time at which the data should be filled
-// bccomp is the index into both domain_bcs_type_bcr and bc_extdir_vals
-//     so this follows the BCVars enum
-//
+/*
+ * Impose lateral boundary conditions on z-component of velocity
+ *
+ * @param[in] dest_arr  Array4 of the quantity to be filled
+ * @param[in] bx        box associated with this data
+ * @param[in] domain    compuational domain
+ * @param[in] z_phys_nd height coordinate at nodes
+ * @param[in] time      time at which the data should be filled
+ * @param[in] bccomp    index into m_domain_bcs_type
+ */
 void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr, const Box& bx, const Box& domain,
-                                              const Array4<Real const>& z_nd_arr,
+                                              const Array4<Real const>& z_phys_nd,
                                               const GpuArray<Real,AMREX_SPACEDIM> dxInv,
-                                              Real /*time*/, int bccomp_w)
+                                              Real /*time*/, int bccomp)
 {
     BL_PROFILE_VAR("impose_lateral_zvel_bcs()",impose_lateral_zvel_bcs);
     const auto& dom_lo = amrex::lbound(domain);
@@ -24,7 +28,7 @@ void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr, cons
     //      0 is used as starting index for bcrs
     int ncomp = 1;
     Vector<BCRec> bcrs_w(1);
-    amrex::setBC(bx, domain, bccomp_w, 0, 1, m_domain_bcs_type, bcrs_w);
+    amrex::setBC(bx, domain, bccomp, 0, 1, m_domain_bcs_type, bcrs_w);
 
     // xlo: ori = 0
     // ylo: ori = 1
@@ -47,7 +51,7 @@ void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr, cons
 
     for (int i = 0; i < ncomp; i++)
         for (int ori = 0; ori < 2*AMREX_SPACEDIM; ori++)
-            l_bc_extdir_vals_d[i][ori] = m_bc_extdir_vals[bccomp_w+i][ori];
+            l_bc_extdir_vals_d[i][ori] = m_bc_extdir_vals[bccomp+i][ori];
 
     GeometryData const& geomdata = m_geom.data();
     bool is_periodic_in_x = geomdata.isPeriodic(0);
@@ -66,7 +70,7 @@ void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr, cons
                 if (bc_ptr_w[n].lo(0) == ERFBCType::ext_dir) {
                     dest_arr(i,j,k) = l_bc_extdir_vals_d[n][0];
                     if (l_use_terrain) {
-                        dest_arr(i,j,k) = WFromOmega(i,j,k,dest_arr(i,j,k),z_nd_arr,dxInv);
+                        dest_arr(i,j,k) = WFromOmega(i,j,k,dest_arr(i,j,k),z_phys_nd,dxInv);
                     }
                 } else if (bc_ptr_w[n].lo(0) == ERFBCType::foextrap) {
                     dest_arr(i,j,k) =  dest_arr(dom_lo.x,j,k);
@@ -81,7 +85,7 @@ void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr, cons
[…]
