### 2023-04-26 erf-model/ERF#1010 @asalmgren
@pratikvn -- I believe we have completed the doxygen documentation (it's not perfect but relatively comprehensive) -- I will consider that part of the issue closed unless you tell us otherwise.   We are still working on the examples.
