### 2023-04-26 9b3e765
fix some typos in docs
