+++ Source/BoundaryConditions/BoundaryConditions_yvel.cpp
@@ -8,7 +8,7 @@ using namespace amrex;
  *
  * @param[in] dest_arr Array4 of the quantity to be filled
  * @param[in] bx       box associated with this data
- * @param[in] domain   compuational domain
+ * @param[in] domain   computational domain
  * @param[in] time     time at which the data should be filled
  * @param[in] bccomp   index into m_domain_bcs_type
  */
@@ -139,10 +139,9 @@ void ERFPhysBCFunct::impose_lateral_yvel_bcs (const Array4<Real>& dest_arr,
 /*
  * Impose vertical boundary conditions on y-component of velocity
  *
- * param
  * @param[in] dest_arr  the Array4 of the quantity to be filled
  * @param[in] bx        the box associated with this data
- * @param[in] domain    the compuational domain
+ * @param[in] domain    the computational domain
  * @param[in] z_phys_nd height coordinate at nodes
  * @param[in] dxInv     inverse cell size array
  * @param[in] time      the time at which the data should be filled
