### 2023-04-27 erf-model/ERF#1010 @pratikvn
@asalmgren , Thank you. I see that it is mentioned on your main webpage (https://erf.readthedocs.io/en/latest/index.html) as a line at the end, but that is easy to miss. Maybe you can also add it to your README on github ? But otherwise, it looks good to me.
