### 2023-04-27 08e7fed
fix some typos in docs (#1074)
