+++ Source/BoundaryConditions/BoundaryConditions_cons.cpp
@@ -6,12 +6,14 @@ using namespace amrex;
 /*
  * Impose lateral boundary conditions on conserved scalars (at cell centers)
  *
- * @param[in] mf     the MultiFab to be filled
- * @param[in] icomp  the index into the MultiFab -- this can be any value from 0 to NVAR-1
- * @param[in] ncomp  the number of components -- this can be any value from 1 to NVAR
- *                   as long as icomp+ncomp <= NVAR-1.
- * @param[in] time   the time at which the data should be filled
- * @param[in] bccomp index into m_domain_bcs_type
+ * @param[in,out] dest_arr cell-centered data to be filled
+ * @param[in]     bx       box holding data to be filled
+ * @param[in]     domain   simulation domain
+ * @param[in]     icomp    index into the MultiFab -- this can be any value from 0 to NVAR-1
+ * @param[in]     ncomp    the number of components -- this can be any value from 1 to NVAR
+ *                         as long as icomp+ncomp <= NVAR-1.
+ * @param[in]     time     time at which the data should be filled
+ * @param[in]     bccomp   index into m_domain_bcs_type
  */
 
 void ERFPhysBCFunct::impose_lateral_cons_bcs (const Array4<Real>& dest_arr, const Box& bx, const Box& domain,
@@ -158,10 +160,9 @@ void ERFPhysBCFunct::impose_lateral_cons_bcs (const Array4<Real>& dest_arr, cons
 /*
  * Impose vertical boundary conditions on conserved scalars (at cell centers)
  *
- * param
  * @param[in] dest_arr  the Array4 of the quantity to be filled
  * @param[in] bx        the box associated with this data
- * @param[in] domain    the compuational domain
+ * @param[in] domain    the computational domain
  * @param[in] z_phys_nd height coordinate at nodes
  * @param[in] dxInv     inverse cell size array
  * @param[in] icomp     the index of the first component to be filled
