+++ Source/TimeIntegration/ERF_slow_rhs_pre.cpp
@@ -320,7 +320,7 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
                 tbxyz.grow(IntVect(-1,-1,0));
 
                 Real mu_eff = solverChoice.dynamicViscosity; // Initialized to 0
-                if (!l_use_turb) {
+                if (cons_visc && !l_use_turb) {
                     ComputeStressConsVisc_T(bxcc, tbxxy, tbxxz, tbxyz, mu_eff,
                                             s11, s22, s33,
                                             s12, s13,
@@ -417,7 +417,7 @@ void erf_slow_rhs_pre (int /*level*/, int nrk,
                 tbxyz.grow(IntVect(-1,-1,0));
 
                 Real mu_eff = solverChoice.dynamicViscosity; // Initialized to 0
-                if (!l_use_turb) {
+                if (cons_visc && !l_use_turb) {
                     ComputeStressConsVisc_N(bxcc, tbxxy, tbxxz, tbxyz, mu_eff,
                                             s11, s22, s33,
                                             s12, s13, s23,
