### 2023-05-01 248f262
Clean up last PR (#1078)
* Streamline.

* Correction to logic.

* This should work.
