### 2023-05-02 6d24f22
move beta_s to TI_fast_rhs_fun.H so we don't define it in two separate places (#1079)
