### 2023-05-03 #5202 @rafmudaf
I work within similar domains and sphere's of collaborators as the submitter, @asalmgren, but we have not directly worked together. I do, however, work with other contributors to ERF both within the same organization (NREL) and, at times, on shared projects. I should also note that I contributed to ERF very early in the project. The [contributors graph](https://github.com/erf-model/ERF/graphs/contributors) shows that I made 33 commits in 2021, but the state of the software is considerably different now than it was then.

@kthyng I feel that I am able to remain impartial given that my work on ERF was a couple of years ago and the software has matured significantly in that time. As for my relationship with contributors, I feel entirely comfortable conducting a candid review and providing direct feedback. However, I also understand that one aspect of conflicts of interest is that their impact is not always evident to the people involved. I'll look out for your guidance on whether this COI can be waived, and please let me know if I can provide more information.
