### 2023-05-07 1a9d4b5
set up TinyProfiler to track Evolve() and InitData()
