### 2023-05-07 bffd16c
Create Performance.rst to add GPU weak scaling
