+++ Docs/sphinx_doc/Performance.rst
@@ -1,0 +1,19 @@
+.. role:: cpp(code)
+    :language: c++
+
+ .. role:: f(code)
+    :language: fortran
+
+.. _sec:Performance:
+
+***********
+Performance
+***********
+.. toctree::
+   :maxdepth: 1
+   
+GPU weak scaling
+================
+
+.. image:: figures/scaling.png
+  :width: 400
