### 2023-05-07 f9142c5
added the weak scaling setup information
