+++ Docs/sphinx_doc/Performance.rst
@@ -4,20 +4,22 @@
  .. role:: f(code)
     :language: fortran
 
-.. _sec:Performance:
+.. _Performance:
 
 ***********
 Performance
 ***********
 .. toctree::
    :maxdepth: 1
-   
+
 GPU weak scaling
 ================
 
-The plot shows weak scaling of the ABL application using A100 GPUs on the Perlmutter system. The domain size is **amr.n_cell = 32 32 128** for a **single GPU** which is progressively scaled up to **512 256 128** for **128 GPUs** while using all 4 GPUs per node to perform 100 time steps.
+The plot shows weak scaling of the ABL application with the Smagorinsky LES model using A100 GPUs on the Perlmutter system at NERSC.
+The domain size is **amr.n_cell = 32 32 128** for a **single GPU**; this is progressively scaled up to **512 256 128** for **128 GPUs**.
+This test uses all 4 GPUs per node and runs 100 time steps.
 
-The test turns off molecular diffusion and activates the Smagorinsky LES model. The processing is limited to the initialization and time-stepping functions by turning off diagnostic calculations, such as the mass summations, and file I/O such as writing checkpoints or plotfiles.
+The times shown includes both initialization and time evolution. All I/O and diagnostic calculations are turned off.
 
 .. image:: figures/scaling.png
   :width: 400
