+++ Docs/sphinx_doc/index.rst
@@ -89,8 +89,8 @@ In addition to this documentation, there is API documentation for ERF generated 
    :maxdepth: 1
    :hidden:
 
-   RegressionTests.rst
    Performance.rst
+   RegressionTests.rst
 
 .. toctree::
    :caption: GOALS
