+++ Docs/sphinx_doc/TimeAdvance.rst
@@ -2,11 +2,7 @@
  .. role:: cpp(code)
     :language: c++
 
- .. role:: f(code)
-    :language: fortran
-
-.. _Algorithms:
-
+ .. _TimeAdvance:
 
 Time Advance
 ============
