+++ Source/Utils/Interpolation.H
@@ -2,6 +2,9 @@
 #define INTERPOLATE_H_
 
 #include "DataStruct.H"
+#include "Interpolation_UPW.H"
+#include "Interpolation_WENO.H"
+#include "Interpolation_WENO_Z.H"
 
 /**
  * Interpolation operators used in construction of advective fluxes using non-WENO schemes
