+++ Source/Advection/AdvectionSrcForState.cpp
@@ -63,252 +63,102 @@ AdvectionSrcForRhoAndTheta (const Box& bx, const Box& valid_bx,
     // We note that valid_bx is the actual grid, while bx may be a tile within that grid
     const auto& vbx_hi = amrex::ubound(valid_bx);
 
-    if ( use_terrain && (std::max(horiz_spatial_order,vert_spatial_order) == 2) && !all_use_WENO) {
-        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-        {
-            Real invdetJ = 1./ detJ(i,j,k);
+    // Not using terrain height coords
+    if (!use_terrain) {
+        // Directly compute RHS w/ second order for efficiency
+        if (std::max(horiz_spatial_order,vert_spatial_order) == 2) {
+            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+            {
+                Real xflux_lo = rho_u(i  ,j,k) / mf_u(i  ,j  ,0);
+                Real xflux_hi = rho_u(i+1,j,k) / mf_u(i+1,j  ,0);
+                Real yflux_lo = rho_v(i,j  ,k) / mf_v(i  ,j  ,0);
+                Real yflux_hi = rho_v(i,j+1,k) / mf_v(i  ,j+1,0);
+                Real zflux_lo = Omega(i,j,k  );
+                Real zflux_hi = Omega(i,j,k+1);
 
-            Real xflux_lo = rho_u(i  ,j,k) / mf_u(i  ,j  ,0);
-            Real xflux_hi = rho_u(i+1,j,k) / mf_u(i+1,j  ,0);
-            Real yflux_lo = rho_v(i,j  ,k) / mf_v(i  ,j  ,0);
-            Real yflux_hi = rho_v(i,j+1,k) / mf_v(i  ,j+1,0);
-            Real zflux_lo = Omega(i,j,k  );
-            Real zflux_hi = Omega(i,j,k+1);
+                avg_xmom(i  ,j,k) += fac*xflux_lo;
+                if (i == vbx_hi.x)
+                    avg_xmom(i+1,j,k) += fac*xflux_hi;
+                avg_ymom(i,j  ,k) += fac*yflux_lo;
+                if (j == vbx_hi.y)
+                    avg_ymom(i,j+1,k) += fac*yflux_hi;
+                avg_zmom(i,j,k  ) += fac*zflux_lo;
+                if (k == vbx_hi.z)
+                    avg_zmom(i,j,k+1) += fac*zflux_hi;
 
-            Real met_h_zeta_xlo = Compute_h_zeta_AtIface(i  ,j  ,k,cellSizeInv,z_nd);
-            xflux_lo *= met_h_zeta_xlo;
-            Real met_h_zeta_xhi = Compute_h_zeta_AtIface(i+1,j  ,k,cellSizeInv,z_nd);
-            xflux_hi *= met_h_zeta_xhi;
+                Real mfsq = mf_m(i,j,0) * mf_m(i,j,0);
 
-            Real met_h_zeta_ylo = Compute_h_zeta_AtJface(i  ,j  ,k,cellSizeInv,z_nd);
-            yflux_lo *= met_h_zeta_ylo;
-            Real met_h_zeta_yhi = Compute_h_zeta_AtJface(i  ,j+1,k,cellSizeInv,z_nd);
-            yflux_hi *= met_h_zeta_yhi;
+                advectionSrc(i,j,k,0) = -(
+                                          ( xflux_hi - xflux_lo ) * dxInv * mfsq +
[…]
