+++ .readthedocs.yml
@@ -1,0 +1,18 @@
+version: 2
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
+
+sphinx:
+   configuration: Docs/source/conf.py
+
+python:
+   install:
+   - requirements: Docs/requirements.txt
+
+formats:
+  - htmlzip
+#  - pdf
+#  - epub
