### 2023-05-09 d225a30
Mom N & T implemented; state still left to tod0.
