+++ Source/Utils/Interpolation_WENO_Z.H
@@ -8,7 +8,8 @@
  */
 struct WENO_Z3
 {
-    WENO_Z3(const amrex::Array4<const amrex::Real>& phi): m_phi(phi) {}
+    WENO_Z3(const amrex::Array4<const amrex::Real>& phi)
+        : m_phi(phi) {}
 
     AMREX_GPU_DEVICE
     AMREX_FORCE_INLINE
@@ -17,8 +18,10 @@ struct WENO_Z3
                    const int& j,
                    const int& k,
                    const int& qty_index,
-                   amrex::Real val_hi,
-                   amrex::Real val_lo)
+                   amrex::Real& val_hi,
+                   amrex::Real& val_lo,
+                   amrex::Real /*upw_hi*/,
+                   amrex::Real /*upw_lo*/)
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i+1, j  , k  , qty_index);
@@ -36,8 +39,10 @@ struct WENO_Z3
                    const int& j,
                    const int& k,
                    const int& qty_index,
-                   amrex::Real val_hi,
-                   amrex::Real val_lo)
+                   amrex::Real& val_hi,
+                   amrex::Real& val_lo,
+                   amrex::Real /*upw_hi*/,
+                   amrex::Real /*upw_lo*/)
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i  , j+1, k  , qty_index);
@@ -51,12 +56,30 @@ struct WENO_Z3
     AMREX_GPU_DEVICE
     AMREX_FORCE_INLINE
     void
+    InterpolateInZ_lo(const int& i,
+                      const int& j,
+                      const int& k,
+                      const int& qty_index,
+                      amrex::Real& val_lo,
+                      amrex::Real /*upw_lo*/)
+    {
+        // Data to interpolate on
+        amrex::Real sp1 = m_phi(i  , j  , k+1, qty_index);
+        amrex::Real s   = m_phi(i  , j  , k  , qty_index);
+        amrex::Real sm1 = m_phi(i  , j  , k-1, qty_index);
+
+        val_lo = Evaluate(sp1,s,sm1);
+    }
+
+    AMREX_GPU_DEVICE
+    AMREX_FORCE_INLINE
+    void
     InterpolateInZ(const int& i,
                    const int& j,
                    const int& k,
                    const int& qty_index,
-                   amrex::Real val_hi,
-                   amrex::Real val_lo)
+                   amrex::Real& val_hi,
+                   amrex::Real /*upw_hi*/)
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i  , j  , k+1, qty_index);
@@ -64,7 +87,6 @@ struct WENO_Z3
         amrex::Real sm1 = m_phi(i  , j  , k-1, qty_index);
 
         val_hi = Evaluate(sm1,s,sp1);
-        val_lo = Evaluate(sp1,s,sm1);
     }
 
     AMREX_GPU_DEVICE
@@ -109,7 +131,8 @@ private:
  */
 struct WENO_Z5
 {
-    WENO_Z5(const amrex::Array4<const amrex::Real>& phi): m_phi(phi) {}
+    WENO_Z5(const amrex::Array4<const amrex::Real>& phi)
+        : m_phi(phi) {}
 
     AMREX_GPU_DEVICE
     AMREX_FORCE_INLINE
@@ -118,8 +141,10 @@ struct WENO_Z5
                    const int& j,
                    const int& k,
[…]
