+++ Source/Advection/AdvectionSrcForMom_T.H
@@ -28,13 +28,12 @@ AdvectionSrcForXMom_T (int i, int j, int k,
                        const amrex::Array4<const amrex::Real>& rho_u,
                        const amrex::Array4<const amrex::Real>& rho_v,
                        const amrex::Array4<const amrex::Real>& Omega,
-                       InterpType interp_u,
                        const amrex::Array4<const amrex::Real>& z_nd,
                        const amrex::Array4<const amrex::Real>& detJ,
+                       InterpType interp_u,
                        const amrex::GpuArray<amrex::Real, AMREX_SPACEDIM>& cellSizeInv,
                        const amrex::Array4<const amrex::Real>& mf_u, const
-                       amrex::Array4<const amrex::Real>& mf_v,
-                       int horiz_spatial_order, int vert_spatial_order)
+                       amrex::Array4<const amrex::Real>& mf_v)
 {
     amrex::Real advectionSrc;
     auto dxInv = cellSizeInv[0], dyInv = cellSizeInv[1], dzInv = cellSizeInv[2];
@@ -79,7 +78,8 @@ AdvectionSrcForXMom_T (int i, int j, int k,
     Omega_avg_hi = 0.5 * (Omega(i, j, k+1) + Omega(i-1, j, k+1));
     Omega_avg_lo = 0.5 * (Omega(i, j, k  ) + Omega(i-1, j, k  ));
 
-    interp_u.InterpolateInZ(i,j,k,0,interp_hi,interp_lo,Omega_avg_hi,Omega_avg_lo);
+    interp_u.InterpolateInZ_hi(i,j,k,0,interp_hi,Omega_avg_hi);
+    interp_u.InterpolateInZ_lo(i,j,k,0,interp_lo,Omega_avg_lo);
 
     amrex::Real edgeFluxXZNext = Omega_avg_hi * interp_hi;
     amrex::Real edgeFluxXZPrev = Omega_avg_lo * interp_lo;
@@ -121,13 +121,12 @@ AdvectionSrcForYMom_T (int i, int j, int k,
                        const amrex::Array4<const amrex::Real>& rho_u,
                        const amrex::Array4<const amrex::Real>& rho_v,
                        const amrex::Array4<const amrex::Real>& Omega,
-                       InterpType interp_v,
                        const amrex::Array4<const amrex::Real>& z_nd,
                        const amrex::Array4<const amrex::Real>& detJ,
+                       InterpType interp_v,
                        const amrex::GpuArray<amrex::Real, AMREX_SPACEDIM>& cellSizeInv,
                        const amrex::Array4<const amrex::Real>& mf_u,
-                       const amrex::Array4<const amrex::Real>& mf_v,
-                       int horiz_spatial_order, int vert_spatial_order)
+                       const amrex::Array4<const amrex::Real>& mf_v)
 {
     amrex::Real advectionSrc;
     auto dxInv = cellSizeInv[0], dyInv = cellSizeInv[1], dzInv = cellSizeInv[2];
@@ -171,7 +170,8 @@ AdvectionSrcForYMom_T (int i, int j, int k,
     Omega_avg_hi = 0.5 * (Omega(i, j, k+1) + Omega(i, j-1, k+1));
     Omega_avg_lo = 0.5 * (Omega(i, j, k  ) + Omega(i, j-1, k  ));
 
-    interp_v.InterpolateInZ(i,j,k,0,interp_hi,interp_lo,Omega_avg_hi,Omega_avg_lo);
+    interp_v.InterpolateInZ_hi(i,j,k,0,interp_hi,Omega_avg_hi);
[…]

[…]
