+++ Docs/sphinx_doc/index.rst
@@ -14,7 +14,7 @@ and `AMRex tutorials <https://amrex-codes.github.io/amrex/tutorials_html/>`_ for
 
 ERF is designed to run on machines from laptops to multicore CPU and hybrid CPU/GPU systems.
 
-In addition to this documentation, there is API documentation for ERF generated by `Doxygen <https://erf-model.github.io/docs/index.html>`_.
+In addition to this documentation, there is API documentation for ERF generated by `Doxygen <https://seahorce-scidac.github.io/docs/index.html>`_.
 
 .. raw:: html
 
