### 2023-05-10 50aeb20
Mom and state completed. Needs proof-reading, moisture loop at start of ScalarSrc, and Cmake.
