+++ Source/Utils/Interpolation_WENO_Z.H
@@ -21,7 +21,7 @@ struct WENO_Z3
                    amrex::Real& val_hi,
                    amrex::Real& val_lo,
                    amrex::Real /*upw_hi*/,
-                   amrex::Real /*upw_lo*/)
+                   amrex::Real /*upw_lo*/) const
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i+1, j  , k  , qty_index);
@@ -42,7 +42,7 @@ struct WENO_Z3
                    amrex::Real& val_hi,
                    amrex::Real& val_lo,
                    amrex::Real /*upw_hi*/,
-                   amrex::Real /*upw_lo*/)
+                   amrex::Real /*upw_lo*/) const
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i  , j+1, k  , qty_index);
@@ -61,7 +61,7 @@ struct WENO_Z3
                       const int& k,
                       const int& qty_index,
                       amrex::Real& val_lo,
-                      amrex::Real /*upw_lo*/)
+                      amrex::Real /*upw_lo*/) const
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i  , j  , k+1, qty_index);
@@ -79,7 +79,7 @@ struct WENO_Z3
                    const int& k,
                    const int& qty_index,
                    amrex::Real& val_hi,
-                   amrex::Real /*upw_hi*/)
+                   amrex::Real /*upw_hi*/) const
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i  , j  , k+1, qty_index);
@@ -94,23 +94,23 @@ struct WENO_Z3
     amrex::Real
     Evaluate(const amrex::Real& sm1,
              const amrex::Real& s  ,
-             const amrex::Real& sp1)
+             const amrex::Real& sp1) const
     {
         // Smoothing factors
-        b1 = (s - sm1) * (s - sm1);
-        b2 = (sp1 - s) * (sp1 - s);
+        amrex::Real b1 = (s - sm1) * (s - sm1);
+        amrex::Real b2 = (sp1 - s) * (sp1 - s);
 
         // Weight factors
-        t5 = std::abs(b2 - b1);
-        w1 = g1 * ( 1.0 + (t5*t5) / ((eps + b1) * (eps + b1)) );
-        w2 = g2 * ( 1.0 + (t5*t5) / ((eps + b2) * (eps + b2)) );
+        amrex::Real t5 = std::abs(b2 - b1);
+        amrex::Real w1 = g1 * ( 1.0 + (t5*t5) / ((eps + b1) * (eps + b1)) );
+        amrex::Real w2 = g2 * ( 1.0 + (t5*t5) / ((eps + b2) * (eps + b2)) );
 
         // Weight factor norm
-        wsum = w1 + w2;
+        amrex::Real wsum = w1 + w2;
 
         // Taylor expansions
-        v1 = -sm1 + 3.0 * s;
-        v2 =  s   + sp1;
+        amrex::Real v1 = -sm1 + 3.0 * s;
+        amrex::Real v2 =  s   + sp1;
 
         // Interpolated value
         return ( (w1 * v1 + w2 * v2) / (2.0 * wsum) );
@@ -118,10 +118,7 @@ struct WENO_Z3
 
 private:
     const amrex::Array4<const amrex::Real>& m_phi;
-    amrex::Real eps=1.0e-6;
[…]
