+++ Source/Advection/AdvectionSrcForMom_N.H
@@ -261,13 +261,11 @@ AdvectionSrcForMomWrapper_N(const amrex::Box& bxx, const amrex::Box& bxy, const 
                             const amrex::Array4<amrex::Real>& rho_u_rhs,
                             const amrex::Array4<amrex::Real>& rho_v_rhs,
                             const amrex::Array4<amrex::Real>& rho_w_rhs,
-                            InterpType interp_u,
-                            InterpType interp_v,
-                            InterpType interp_w,
-                            WallInterpType interp_w_wall,
                             const amrex::Array4<const amrex::Real>& rho_u,
                             const amrex::Array4<const amrex::Real>& rho_v,
                             const amrex::Array4<const amrex::Real>& rho_w,
+                            const amrex::Array4<const amrex::Real>& u,
+                            const amrex::Array4<const amrex::Real>& v,
                             const amrex::Array4<const amrex::Real>& w,
                             const amrex::GpuArray<amrex::Real, AMREX_SPACEDIM>& cellSizeInv,
                             const amrex::Array4<const amrex::Real>& mf_m,
@@ -276,6 +274,11 @@ AdvectionSrcForMomWrapper_N(const amrex::Box& bxx, const amrex::Box& bxy, const 
                             const int vert_spatial_order,
                             const int domhi_z)
 {
+    // Instantiate the appropriate structs
+    InterpType interp_u(u); // X-MOM
+    InterpType interp_v(v); // Y-MOM
+    InterpType interp_w(w); // Z-MOM
+    WallInterpType interp_w_wall(w); // Z-MOM @ wall
     amrex::ParallelFor(bxx, bxy, bxz,
     [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
     {
