### 2023-05-12 7bb0582
Use WENO for cc scalars other than rho and theta. Also fix WENO-Z.
