+++ Source/Utils/Interpolation_WENO_Z.H
@@ -20,16 +20,31 @@ struct WENO_Z3
                    const int& qty_index,
                    amrex::Real& val_hi,
                    amrex::Real& val_lo,
-                   amrex::Real /*upw_hi*/,
-                   amrex::Real /*upw_lo*/) const
+                   amrex::Real upw_hi,
+                   amrex::Real upw_lo) const
     {
         // Data to interpolate on
+        amrex::Real sp2 = m_phi(i+2, j  , k  , qty_index);
         amrex::Real sp1 = m_phi(i+1, j  , k  , qty_index);
         amrex::Real s   = m_phi(i  , j  , k  , qty_index);
         amrex::Real sm1 = m_phi(i-1, j  , k  , qty_index);
+        amrex::Real sm2 = m_phi(i-2, j  , k  , qty_index);
 
-        val_hi = Evaluate(sm1,s,sp1);
-        val_lo = Evaluate(sp1,s,sm1);
+        if (upw_hi > tol){
+            val_hi = Evaluate(sm1,s  ,sp1);
+        } else if (upw_hi < -tol) {
+            val_hi = Evaluate(sp2,sp1,s  );
+        } else {
+            val_hi = 0.5 * (s + sp1);
+        }
+
+        if (upw_lo > tol) {
+            val_lo = Evaluate(sm2,sm1,s  );
+        } else if (upw_lo < -tol) {
+            val_lo = Evaluate(sp1,s  ,sm1);
+        } else {
+            val_lo = 0.5 * (s + sm1);
+        }
     }
 
     AMREX_GPU_DEVICE
@@ -41,16 +56,31 @@ struct WENO_Z3
                    const int& qty_index,
                    amrex::Real& val_hi,
                    amrex::Real& val_lo,
-                   amrex::Real /*upw_hi*/,
-                   amrex::Real /*upw_lo*/) const
+                   amrex::Real upw_hi,
+                   amrex::Real upw_lo) const
     {
         // Data to interpolate on
+        amrex::Real sp2 = m_phi(i  , j+2, k  , qty_index);
         amrex::Real sp1 = m_phi(i  , j+1, k  , qty_index);
         amrex::Real s   = m_phi(i  , j  , k  , qty_index);
         amrex::Real sm1 = m_phi(i  , j-1, k  , qty_index);
+        amrex::Real sm2 = m_phi(i  , j-2, k  , qty_index);
 
-        val_hi = Evaluate(sm1,s,sp1);
-        val_lo = Evaluate(sp1,s,sm1);
+        if (upw_hi > tol){
+            val_hi = Evaluate(sm1,s  ,sp1);
+        } else if (upw_hi < -tol) {
+            val_hi = Evaluate(sp2,sp1,s  );
+        } else {
+            val_hi = 0.5 * (s + sp1);
+        }
+
+        if (upw_lo > tol) {
+            val_lo = Evaluate(sm2,sm1,s  );
+        } else if (upw_lo < -tol) {
+            val_lo = Evaluate(sp1,s  ,sm1);
+        } else {
+            val_lo = 0.5 * (s + sm1);
+        }
     }
 
     AMREX_GPU_DEVICE
@@ -61,32 +91,46 @@ struct WENO_Z3
                       const int& k,
                       const int& qty_index,
                       amrex::Real& val_lo,
-                      amrex::Real /*upw_lo*/) const
+                      amrex::Real upw_lo) const
     {
         // Data to interpolate on
         amrex::Real sp1 = m_phi(i  , j  , k+1, qty_index);
[…]
