### 2023-05-12 4913efa
Make advection more modular (#1093)
* This compiled in debug but needs A LOT of work.

* Mom N & T implemented; state still left to tod0.

* Broken with state vars.

* Mom and state completed. Needs proof-reading, moisture loop at start of ScalarSrc, and Cmake.

* Add moist loop for state, fix errors in N/T mom files, general clean up.

* Fix logic with WENO.

* Fix WENO upwinding issue.

* Use WENO for cc scalars other than rho and theta. Also fix WENO-Z.

---------

Co-authored-by: Aaron Lattanzi <amlattanzi@ghidorah.dhcp.lbl.gov>
