+++ Source/Utils/Interpolation_UPW.H
@@ -1,0 +1,675 @@
+#ifndef INTERPOLATE_UPW_H_
+#define INTERPOLATE_UPW_H_
+
+#include "DataStruct.H"
+
+/**
+ * Interpolation operators used for 3rd order upwind scheme
+ */
+struct UPWIND3
+{
+    UPWIND3(const amrex::Array4<const amrex::Real>& phi)
+        : m_phi(phi) {}
+
+    AMREX_GPU_DEVICE
+    AMREX_FORCE_INLINE
+    void
+    InterpolateInX(const int& i,
+                   const int& j,
+                   const int& k,
+                   const int& qty_index,
+                   amrex::Real& val_hi,
+                   amrex::Real& val_lo,
+                   amrex::Real upw_hi,
+                   amrex::Real upw_lo) const
+    {
+        // Data to interpolate on
+        amrex::Real sp2 = m_phi(i+2, j  , k  , qty_index);
+        amrex::Real sp1 = m_phi(i+1, j  , k  , qty_index);
+        amrex::Real s   = m_phi(i  , j  , k  , qty_index);
+        amrex::Real sm1 = m_phi(i-1, j  , k  , qty_index);
+        amrex::Real sm2 = m_phi(i-2, j  , k  , qty_index);
+
+        // Upwinding flags
+        if (upw_hi != 0.) upw_hi = (upw_hi > 0) ? 1. : -1.;
+        if (upw_lo != 0.) upw_lo = (upw_lo > 0) ? 1. : -1.;
+
+        // Interpolate hi
+        val_hi = Evaluate(sp2,sp1,s,sm1,upw_hi);
+
+        // Interpolate lo
+        val_lo = Evaluate(sp1,s,sm1,sm2,upw_lo);
+    }
+
+    AMREX_GPU_DEVICE
+    AMREX_FORCE_INLINE
+    void
+    InterpolateInY(const int& i,
+                   const int& j,
+                   const int& k,
+                   const int& qty_index,
+                   amrex::Real& val_hi,
+                   amrex::Real& val_lo,
+                   amrex::Real upw_hi,
+                   amrex::Real upw_lo) const
+    {
+        // Data to interpolate on
+        amrex::Real sp2 = m_phi(i  , j+2, k  , qty_index);
+        amrex::Real sp1 = m_phi(i  , j+1, k  , qty_index);
+        amrex::Real s   = m_phi(i  , j  , k  , qty_index);
+        amrex::Real sm1 = m_phi(i  , j-1, k  , qty_index);
+        amrex::Real sm2 = m_phi(i  , j-2, k  , qty_index);
+
+        // Upwinding flags
+        if (upw_hi != 0.) upw_hi = (upw_hi > 0) ? 1. : -1.;
+        if (upw_lo != 0.) upw_lo = (upw_lo > 0) ? 1. : -1.;
+
+        // Interpolate hi
+        val_hi = Evaluate(sp2,sp1,s,sm1,upw_hi);
+
+        // Interpolate lo
+        val_lo = Evaluate(sp1,s,sm1,sm2,upw_lo);
+    }
+
+    AMREX_GPU_DEVICE
+    AMREX_FORCE_INLINE
+    void
+    InterpolateInZ_lo(const int& i,
+                      const int& j,
+                      const int& k,
+                      const int& qty_index,
+                      amrex::Real& val_lo,
+                      amrex::Real upw_lo) const
+    {
+        // Data to interpolate on
+        amrex::Real sp1 = m_phi(i  , j  , k+1, qty_index);
+        amrex::Real s   = m_phi(i  , j  , k  , qty_index);
+        amrex::Real sm1 = m_phi(i  , j  , k-1, qty_index);
[…]
