+++ Source/Advection/AdvectionSrcForState.cpp
@@ -1,8 +1,8 @@
 #include <IndexDefines.H>
 #include <TerrainMetrics.H>
 #include <Advection.H>
-#include <Interpolation.H>
-#include <Interpolation_WENO.H>
+#include <AdvectionSrcForState_N.H>
+#include <AdvectionSrcForState_T.H>
 
 using namespace amrex;
 
@@ -51,8 +51,6 @@ AdvectionSrcForRhoAndTheta (const Box& bx, const Box& valid_bx,
                             const Array4<const Real>& mf_m,
                             const Array4<const Real>& mf_u,
                             const Array4<const Real>& mf_v,
-                            const bool all_use_WENO,
-                            const int  spatial_order_WENO,
                             const int horiz_spatial_order,
                             const int vert_spatial_order,
                             const int use_terrain)
@@ -61,254 +59,149 @@ AdvectionSrcForRhoAndTheta (const Box& bx, const Box& valid_bx,
     auto dxInv = cellSizeInv[0], dyInv = cellSizeInv[1], dzInv = cellSizeInv[2];
 
     // We note that valid_bx is the actual grid, while bx may be a tile within that grid
-    const auto& vbx_hi = amrex::ubound(valid_bx);
+    const auto& vbx_hi = ubound(valid_bx);
 
-    if ( use_terrain && (std::max(horiz_spatial_order,vert_spatial_order) == 2) && !all_use_WENO) {
-        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
-        {
-            Real invdetJ = 1./ detJ(i,j,k);
+    if (!use_terrain) {
+        // Inline with 2nd order for efficiency
+        if (std::max(horiz_spatial_order,vert_spatial_order) == 2) {
+            ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
+            {
+                Real xflux_lo = rho_u(i  ,j,k) / mf_u(i  ,j  ,0);
+                Real xflux_hi = rho_u(i+1,j,k) / mf_u(i+1,j  ,0);
+                Real yflux_lo = rho_v(i,j  ,k) / mf_v(i  ,j  ,0);
+                Real yflux_hi = rho_v(i,j+1,k) / mf_v(i  ,j+1,0);
+                Real zflux_lo = Omega(i,j,k  );
+                Real zflux_hi = Omega(i,j,k+1);
 
-            Real xflux_lo = rho_u(i  ,j,k) / mf_u(i  ,j  ,0);
-            Real xflux_hi = rho_u(i+1,j,k) / mf_u(i+1,j  ,0);
-            Real yflux_lo = rho_v(i,j  ,k) / mf_v(i  ,j  ,0);
-            Real yflux_hi = rho_v(i,j+1,k) / mf_v(i  ,j+1,0);
-            Real zflux_lo = Omega(i,j,k  );
-            Real zflux_hi = Omega(i,j,k+1);
+                avg_xmom(i  ,j,k) += fac*xflux_lo;
+                if (i == vbx_hi.x)
+                    avg_xmom(i+1,j,k) += fac*xflux_hi;
+                avg_ymom(i,j  ,k) += fac*yflux_lo;
+                if (j == vbx_hi.y)
+                    avg_ymom(i,j+1,k) += fac*yflux_hi;
+                avg_zmom(i,j,k  ) += fac*zflux_lo;
+                if (k == vbx_hi.z)
[…]
