### 2023-05-15 8257eb9
Read moisture vars for init real.
