+++ Source/Utils/Interpolation_WENO_Z.H
@@ -161,7 +161,7 @@ struct WENO_Z3
     }
 
 private:
-    const amrex::Array4<const amrex::Real>& m_phi;
+    amrex::Array4<const amrex::Real> m_phi;   // Quantity to interpolate
     const amrex::Real eps=1.0e-6;
     const amrex::Real tol=1.0e-12;
     static constexpr amrex::Real g1=(1.0/3.0);
@@ -342,7 +342,7 @@ struct WENO_Z5
     }
 
 private:
-    const amrex::Array4<const amrex::Real>& m_phi;
+    amrex::Array4<const amrex::Real> m_phi;   // Quantity to interpolate
     const amrex::Real eps=1.0e-6;
     const amrex::Real tol=1.0e-12;
     static constexpr amrex::Real c1=(13.0/12.0);
