### 2023-05-18 #5202 @kthyng
@adam-m-jcbs It does look like you dove in deep working to understand author information, beyond what we expected. I'm sorry you've had a poor experience reviewing here — if you end up wanting to, I'd invite you to follow up to share your experience to see if we can improve things for reviewers, or at least to better communicate the tasks we're requesting. I'll also share your comment with the lead editorial team.

I'll remove you as a reviewer and from our list of reviewers, as requested.
