### 2023-05-22 9eed856
Provides a plugin for WENO_Z. However, WENO_Z causes oscillations with ScalarAdvection tests.
