### 2023-05-22 1fddfc3
Provides a plugin for WENO_Z. However, WENO_Z causes oscillations with ScalarAdvection tests. (#1097)
