+++ Source/Advection/Advection.H
@@ -43,6 +43,8 @@ void AdvectionSrcForScalars (const amrex::Box& bx,
                              const amrex::Array4<const amrex::Real>& mf_m,
                              const bool all_use_WENO,
                              const bool moist_use_WENO,
+                             const bool all_use_WENO_Z,
+                             const bool moist_use_WENO_Z,
                              const int  spatial_order_WENO,
                              const int horiz_spatial_order, const int vert_spatial_order, const int use_terrain);
 
