### 2023-05-24 952255b
WIP initialization from met_em files
