+++ Source/Initialization/ERF_init_from_metgrid.cpp
@@ -14,56 +14,95 @@ void
 read_from_metgrid(int lev, const std::string& fname,
                   FArrayBox& NC_xvel_fab, FArrayBox& NC_yvel_fab,
                   FArrayBox& NC_temp_fab, FArrayBox& NC_rhum_fab,
-                  FArrayBox& NC_pres_fab, FArrayBox& NC_hgt_fab,
+                  FArrayBox& NC_pres_fab, FArrayBox& NC_ght_fab,
+                  FArrayBox& NC_hgt_fab,  FArrayBox& NC_psfc_fab, 
                   FArrayBox& NC_msfu_fab, FArrayBox& NC_msfv_fab,
                   FArrayBox& NC_msfm_fab);
+
 void
 interpolate_column(int i, int j, int src_comp, int dest_comp,
                    const Array4<Real const>& orig_z, const Array4<Real const>& orig_data,
                    const Array4<Real const>&  new_z, const Array4<Real>&  new_data);
 
 void
-init_terrain_from_metgrid(int lev, FArrayBox& z_phys_nd_fab,
+interpolate_column_metgrid(int i, int j, char stag, int src_comp, int dest_comp,
+                           const Array4<Real const>& orig_z, const Array4<Real const>& orig_data,
+                           const Array4<Real const>&  new_z, const Array4<Real>&  new_data);
+
+void
+init_terrain_from_metgrid(FArrayBox& z_phys_nd_fab,
                           const Vector<FArrayBox>& NC_hgt_fab);
 
 void
-init_state_from_metgrid(int lev, FArrayBox& state_fab,
+init_state_from_metgrid(FArrayBox& state_fab,
                         FArrayBox& x_vel_fab, FArrayBox& y_vel_fab,
                         FArrayBox& z_vel_fab, FArrayBox& z_phys_nd_fab,
                         const Vector<FArrayBox>& NC_hgt_fab,
+                        const Vector<FArrayBox>& NC_ght_fab,
                         const Vector<FArrayBox>& NC_xvel_fab,
                         const Vector<FArrayBox>& NC_yvel_fab,
                         const Vector<FArrayBox>& NC_zvel_fab,
-                        const Vector<FArrayBox>& NC_rho_fab,
-                        const Vector<FArrayBox>& NC_rhotheta_fab);
+                        const Vector<FArrayBox>& NC_temp_fab,
+                        const Vector<FArrayBox>& NC_rhum_fab,
+                              Vector<FArrayBox>& mxrat_fab);
+
 void
-init_msfs_from_metgrid(int lev, FArrayBox& msfu_fab,
-                       FArrayBox& msfv_fab, FArrayBox& msfm_fab,
+init_msfs_from_metgrid(FArrayBox& msfu_fab,
+                       FArrayBox& msfv_fab,
+                       FArrayBox& msfm_fab,
                        const Vector<FArrayBox>& NC_MSFU_fab,
                        const Vector<FArrayBox>& NC_MSFV_fab,
                        const Vector<FArrayBox>& NC_MSFM_fab);
+
 void
-init_base_state_from_metgrid(int lev, const Box& valid_bx, const Real l_rdOcp,
-                             FArrayBox& p_hse, FArrayBox& pi_hse, FArrayBox& r_hse,
-                             const Vector<FArrayBox>& NC_ALB_fab,
-                             const Vector<FArrayBox>& NC_PB_fab);
+init_base_state_from_metgrid(const Real l_rdOcp,
[…]
