### 2023-05-24 be4c72e
Fix logic to use WENO_Z. (#1098)
