### 2023-05-25 bfddebc
Add horizontal and vertical spatial orders back with modular advection.
