+++ Source/Advection/AdvectionSrcForState.cpp
@@ -101,26 +101,36 @@ AdvectionSrcForRhoAndTheta (const Box& bx, const Box& valid_bx,
             });
         // Template higher order methods
         } else {
-            if (std::max(horiz_spatial_order,vert_spatial_order) == 3) {
-                AdvectionSrcForRhoThetaWrapper_N<UPWIND3>(bx, vbx_hi, fac, advectionSrc,
-                                                          cell_prim, rho_u, rho_v, Omega,
-                                                          avg_xmom, avg_ymom, avg_zmom,
-                                                          cellSizeInv, mf_m, mf_u, mf_v);
-            } else if (std::max(horiz_spatial_order,vert_spatial_order) == 4) {
-                AdvectionSrcForRhoThetaWrapper_N<UPWIND4>(bx, vbx_hi, fac, advectionSrc,
-                                                          cell_prim, rho_u, rho_v, Omega,
-                                                          avg_xmom, avg_ymom, avg_zmom,
-                                                          cellSizeInv, mf_m, mf_u, mf_v);
-            } else if (std::max(horiz_spatial_order,vert_spatial_order) == 5) {
-                AdvectionSrcForRhoThetaWrapper_N<UPWIND5>(bx, vbx_hi, fac, advectionSrc,
-                                                          cell_prim, rho_u, rho_v, Omega,
-                                                          avg_xmom, avg_ymom, avg_zmom,
-                                                          cellSizeInv, mf_m, mf_u, mf_v);
-            } else if (std::max(horiz_spatial_order,vert_spatial_order) == 6) {
-                AdvectionSrcForRhoThetaWrapper_N<UPWIND6>(bx, vbx_hi, fac, advectionSrc,
-                                                          cell_prim, rho_u, rho_v, Omega,
-                                                          avg_xmom, avg_ymom, avg_zmom,
-                                                          cellSizeInv, mf_m, mf_u, mf_v);
+            if (horiz_spatial_order == 2) {
+                AdvectionSrcForRhoThetaVert_N<UPWIND2>(bx, vbx_hi, fac, advectionSrc,
+                                                       cell_prim, rho_u, rho_v, Omega,
+                                                       avg_xmom, avg_ymom, avg_zmom,
+                                                       cellSizeInv, mf_m, mf_u, mf_v,
+                                                       vert_spatial_order);
+            } else if (horiz_spatial_order == 3) {
+                AdvectionSrcForRhoThetaVert_N<UPWIND3>(bx, vbx_hi, fac, advectionSrc,
+                                                       cell_prim, rho_u, rho_v, Omega,
+                                                       avg_xmom, avg_ymom, avg_zmom,
+                                                       cellSizeInv, mf_m, mf_u, mf_v,
+                                                       vert_spatial_order);
+            } else if (horiz_spatial_order == 4) {
+                AdvectionSrcForRhoThetaVert_N<UPWIND4>(bx, vbx_hi, fac, advectionSrc,
+                                                       cell_prim, rho_u, rho_v, Omega,
+                                                       avg_xmom, avg_ymom, avg_zmom,
+                                                       cellSizeInv, mf_m, mf_u, mf_v,
+                                                       vert_spatial_order);
+            } else if (horiz_spatial_order == 5) {
+                AdvectionSrcForRhoThetaVert_N<UPWIND5>(bx, vbx_hi, fac, advectionSrc,
+                                                       cell_prim, rho_u, rho_v, Omega,
+                                                       avg_xmom, avg_ymom, avg_zmom,
+                                                       cellSizeInv, mf_m, mf_u, mf_v,
+                                                       vert_spatial_order);
[…]
