### 2023-05-26 cea554c
add option to do incompressible flow
