+++ Source/Advection/AdvectionSrcForState_N.H
@@ -4,7 +4,7 @@
 /**
  * Wrapper function for computing the advective tendency w/ spatial order > 2.
  */
-template<typename InterpType>
+template<typename InterpType_H, typename InterpType_V>
 void
 AdvectionSrcForRhoThetaWrapper_N(const amrex::Box& bx,
                                  const amrex::Dim3& vbx_hi,
@@ -23,7 +23,9 @@ AdvectionSrcForRhoThetaWrapper_N(const amrex::Box& bx,
                                  const amrex::Array4<const amrex::Real>& mf_v)
 {
     // Instantiate struct
-    InterpType interp_prim(cell_prim);
+    InterpType_H interp_prim_h(cell_prim);
+    InterpType_V interp_prim_v(cell_prim);
+
     auto dxInv = cellSizeInv[0], dyInv = cellSizeInv[1], dzInv = cellSizeInv[2];
     amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) noexcept
     {
@@ -56,11 +58,11 @@ AdvectionSrcForRhoThetaWrapper_N(const amrex::Box& bx,
         amrex::Real interpy_hi(0.), interpy_lo(0.);
         amrex::Real interpz_hi(0.), interpz_lo(0.);
 
-        interp_prim.InterpolateInX(i,j,k,prim_index,interpx_hi,interpx_lo,rho_u(i+1,j  ,k  ),rho_u(i  ,j  ,k  ));
-        interp_prim.InterpolateInY(i,j,k,prim_index,interpy_hi,interpy_lo,rho_v(i  ,j+1,k  ),rho_v(i  ,j  ,k  ));
+        interp_prim_h.InterpolateInX(i,j,k,prim_index,interpx_hi,interpx_lo,rho_u(i+1,j  ,k  ),rho_u(i  ,j  ,k  ));
+        interp_prim_h.InterpolateInY(i,j,k,prim_index,interpy_hi,interpy_lo,rho_v(i  ,j+1,k  ),rho_v(i  ,j  ,k  ));
 
-        interp_prim.InterpolateInZ_hi(i,j,k,prim_index,interpz_hi,rho_w(i  ,j  ,k+1));
-        interp_prim.InterpolateInZ_lo(i,j,k,prim_index,interpz_lo,rho_w(i  ,j  ,k  ));
+        interp_prim_v.InterpolateInZ_hi(i,j,k,prim_index,interpz_hi,rho_w(i  ,j  ,k+1));
+        interp_prim_v.InterpolateInZ_lo(i,j,k,prim_index,interpz_lo,rho_w(i  ,j  ,k  ));
 
         advectionSrc(i,j,k,1) = -(
                                   ( xflux_hi * interpx_hi - xflux_lo * interpx_lo ) * dxInv * mfsq +
@@ -70,9 +72,61 @@ AdvectionSrcForRhoThetaWrapper_N(const amrex::Box& bx,
 }
 
 /**
+ * Wrapper function for templating the vertical advective tendency w/ spatial order > 2.
+ */
+template<typename InterpType_H>
+void
+AdvectionSrcForRhoThetaVert_N(const amrex::Box& bx,
+                              const amrex::Dim3& vbx_hi,
+                              const amrex::Real fac,
+                              const amrex::Array4<      amrex::Real>& advectionSrc,
+                              const amrex::Array4<const amrex::Real>& cell_prim,
+                              const amrex::Array4<const amrex::Real>& rho_u,
+                              const amrex::Array4<const amrex::Real>& rho_v,
+                              const amrex::Array4<const amrex::Real>& rho_w,
+                              const amrex::Array4<      amrex::Real>& avg_xmom,
+                              const amrex::Array4<      amrex::Real>& avg_ymom,
[…]
