### 2023-05-30 2380ff4
Fix wps regtest (#1101)
* Fix WPS regtest.

* Remove accidentally added log files.
