+++ Source/Advection/AdvectionSrcForState.cpp
@@ -282,34 +282,26 @@ AdvectionSrcForScalars (const Box& bx, const int &icomp, const int &ncomp,
     // Running with WENO for moisture but not for other vars
     if(moist_use_WENO && ((icomp+ncomp)==NVAR) ) {
         ncomp_end -= 2;
-        if (moist_use_WENO && spatial_order_WENO==3) {
-            AdvectionSrcForScalarsVert_N<WENO3>(bx, 2, moist_off,
-                                                use_terrain, advectionSrc, cell_prim,
-                                                avg_xmom, avg_ymom, avg_zmom, detJ,
-                                                cellSizeInv, mf_m, all_use_WENO,
-                                                all_use_WENO_Z, spatial_order_WENO,
-                                                vert_spatial_order);
-        } else if (moist_use_WENO && spatial_order_WENO==5) {
-            AdvectionSrcForScalarsVert_N<WENO5>(bx, 2, moist_off,
-                                                use_terrain, advectionSrc, cell_prim,
-                                                avg_xmom, avg_ymom, avg_zmom, detJ,
-                                                cellSizeInv, mf_m, all_use_WENO,
-                                                all_use_WENO_Z, spatial_order_WENO,
-                                                vert_spatial_order);
-        } else if (moist_use_WENO_Z && spatial_order_WENO==3) {
-            AdvectionSrcForScalarsVert_N<WENO_Z3>(bx, 2, moist_off,
-                                                  use_terrain, advectionSrc, cell_prim,
-                                                  avg_xmom, avg_ymom, avg_zmom, detJ,
-                                                  cellSizeInv, mf_m, all_use_WENO,
-                                                  all_use_WENO_Z, spatial_order_WENO,
-                                                  vert_spatial_order);
-        } else if (moist_use_WENO_Z && spatial_order_WENO==5) {
-            AdvectionSrcForScalarsVert_N<WENO_Z5>(bx, 2, moist_off,
-                                                  use_terrain, advectionSrc, cell_prim,
-                                                  avg_xmom, avg_ymom, avg_zmom, detJ,
-                                                  cellSizeInv, mf_m, all_use_WENO,
-                                                  all_use_WENO_Z, spatial_order_WENO,
-                                                  vert_spatial_order);
+        if (spatial_order_WENO==3) {
+            AdvectionSrcForScalarsWrapper_N<WENO3,WENO3>(bx, 2, moist_off,
+                                                         use_terrain, advectionSrc, cell_prim,
+                                                         avg_xmom, avg_ymom, avg_zmom, detJ,
+                                                         cellSizeInv, mf_m);
+        } else if (spatial_order_WENO==5) {
+            AdvectionSrcForScalarsWrapper_N<WENO5,WENO5>(bx, 2, moist_off,
+                                                         use_terrain, advectionSrc, cell_prim,
+                                                         avg_xmom, avg_ymom, avg_zmom, detJ,
+                                                         cellSizeInv, mf_m);
+        } else if (spatial_order_WENO==3) {
+            AdvectionSrcForScalarsWrapper_N<WENO_Z3,WENO_Z3>(bx, 2, moist_off,
+                                                             use_terrain, advectionSrc, cell_prim,
+                                                             avg_xmom, avg_ymom, avg_zmom, detJ,
+                                                             cellSizeInv, mf_m);
+        } else if (spatial_order_WENO==5) {
+            AdvectionSrcForScalarsWrapper_N<WENO_Z5,WENO_Z5>(bx, 2, moist_off,
+                                                             use_terrain, advectionSrc, cell_prim,
+                                                             avg_xmom, avg_ymom, avg_zmom, detJ,
+                                                             cellSizeInv, mf_m);
         } else {
[…]
