+++ Source/Initialization/InputSoundingData.H
@@ -94,9 +94,8 @@ public:
             }
 
             // At this point, we have an input_sounding from z=0 AGL up to
-            // z_inp_sound_tmp[N-1] >= ztop. Now, interpolate to grid level 0
-            // heights.
-            const int Ninp = z_inp_sound_tmp.size();
+            // z_inp_sound_tmp[N-1] >= ztop. Now, interpolate to grid level 0 heights
+            const int Ninp = z_inp_sound_tmp.size()-1;
             z_inp_sound[0]     = zbot;
             theta_inp_sound[0] = theta_inp_sound_tmp[0];
             qv_inp_sound[0]    = qv_inp_sound_tmp[0];
