### 2023-05-31 7439665
Generalize to have arbitrary start time (default 0)
