### 2023-05-31 900767c
Generalize to have arbitrary start time (default 0) (#1103)
* Generalize to have arbitrary start time (default 0)

* update docs to include start_time

* update inputs file to start at non-zero time
