+++ Source/IO/Plotfile.cpp
@@ -548,52 +548,52 @@ ERF::WritePlotFile (int which, Vector<std::string> plot_var_names)
         calculate_derived("qt",          derived::erf_derQt);
         calculate_derived("qp",          derived::erf_derQp);
 
-        MultiFab qv_fab(qv[lev], make_alias, 0, 1);
-        MultiFab qc_fab(qc[lev], make_alias, 0, 1);
-        MultiFab qi_fab(qi[lev], make_alias, 0, 1);
-        MultiFab qrain_fab(qrain[lev], make_alias, 0, 1);
-        MultiFab qsnow_fab(qsnow[lev], make_alias, 0, 1);
-        MultiFab qgraup_fab(qgraup[lev], make_alias, 0, 1);
+        MultiFab qv_mf(qv[lev], make_alias, 0, 1);
+        MultiFab qc_mf(qc[lev], make_alias, 0, 1);
+        MultiFab qi_mf(qi[lev], make_alias, 0, 1);
+        MultiFab qrain_mf(qrain[lev], make_alias, 0, 1);
+        MultiFab qsnow_mf(qsnow[lev], make_alias, 0, 1);
+        MultiFab qgraup_mf(qgraup[lev], make_alias, 0, 1);
 
         if (containerHasElement(plot_var_names, "qv"))
         {
             // r_0 is first component of base_state
-            MultiFab::Copy(mf[lev],qv_fab,0,mf_comp,1,0);
+            MultiFab::Copy(mf[lev],qv_mf,0,mf_comp,1,0);
             mf_comp += 1;
         }
 
         if (containerHasElement(plot_var_names, "qc"))
         {
             // r_0 is first component of base_state
-            MultiFab::Copy(mf[lev],qc_fab,0,mf_comp,1,0);
+            MultiFab::Copy(mf[lev],qc_mf,0,mf_comp,1,0);
             mf_comp += 1;
         }
 
         if (containerHasElement(plot_var_names, "qi"))
         {
             // r_0 is first component of base_state
-            MultiFab::Copy(mf[lev],qi_fab,0,mf_comp,1,0);
+            MultiFab::Copy(mf[lev],qi_mf,0,mf_comp,1,0);
             mf_comp += 1;
         }
 
         if (containerHasElement(plot_var_names, "qrain"))
         {
             // r_0 is first component of base_state
-            MultiFab::Copy(mf[lev],qrain_fab,0,mf_comp,1,0);
+            MultiFab::Copy(mf[lev],qrain_mf,0,mf_comp,1,0);
             mf_comp += 1;
         }
 
         if (containerHasElement(plot_var_names, "qsnow"))
         {
             // r_0 is first component of base_state
-            MultiFab::Copy(mf[lev],qsnow_fab,0,mf_comp,1,0);
+            MultiFab::Copy(mf[lev],qsnow_mf,0,mf_comp,1,0);
             mf_comp += 1;
         }
 
         if (containerHasElement(plot_var_names, "qgraup"))
         {
             // r_0 is first component of base_state
-            MultiFab::Copy(mf[lev],qgraup_fab,0,mf_comp,1,0);
+            MultiFab::Copy(mf[lev],qgraup_mf,0,mf_comp,1,0);
             mf_comp += 1;
         }
 #elif defined(ERF_USE_WARM_NO_PRECIP)
