### 2023-05-31 f887509
Consolidate moisture variables to make bc filling easier
