### 2023-06-01 2ad0acd
Consolidate moisture variables to make bc filling easier (#1105)
