+++ Source/Microphysics/Update.cpp
@@ -7,62 +7,54 @@
  * Updates conserved and microphysics variables in the provided MultiFabs from
  * the internal MultiFabs that store Microphysics module data.
  *
- * @param[out] cons_in Conserved variables
- * @param[out] qv_in Water vapor component
- * @param[out] qc_in Cloud liquid component
- * @param[out] qi_in Cloud ice component
- * @param[out] qrain_in Precipitation liquid component
- * @param[out] qsnow_in Precipitation ice component
- * @param[out] qgraup_in Graup component
+ * @param[out] cons Conserved variables
+ * @param[out] qmoist: qv, qc, qi, qr, qs, qg
  */
-void Microphysics::Update(amrex::MultiFab& cons_in,
-                          amrex::MultiFab& qv_in,
-                          amrex::MultiFab& qc_in,
-                          amrex::MultiFab& qi_in,
-                          amrex::MultiFab& qrain_in,
-                          amrex::MultiFab& qsnow_in,
-                          amrex::MultiFab& qgraup_in)
+void Microphysics::Update(amrex::MultiFab& cons,
+                          amrex::MultiFab& qmoist)
 {
   // copy multifab data to qc, qv, and qi
-  amrex::MultiFab::Copy(qv_in    , *mic_fab_vars[MicVar::qv],  0, 0, 1, mic_fab_vars[MicVar::qv]->nGrowVect());
-  amrex::MultiFab::Copy(qc_in    , *mic_fab_vars[MicVar::qcl], 0, 0, 1, mic_fab_vars[MicVar::qcl]->nGrowVect());
-  amrex::MultiFab::Copy(qi_in    , *mic_fab_vars[MicVar::qci], 0, 0, 1, mic_fab_vars[MicVar::qci]->nGrowVect());
-  amrex::MultiFab::Copy(qrain_in , *mic_fab_vars[MicVar::qpl], 0, 0, 1, mic_fab_vars[MicVar::qpl]->nGrowVect());
-  amrex::MultiFab::Copy(qsnow_in , *mic_fab_vars[MicVar::qpi], 0, 0, 1, mic_fab_vars[MicVar::qpi]->nGrowVect());
-  amrex::MultiFab::Copy(qgraup_in, *mic_fab_vars[MicVar::qpi], 0, 0, 1, mic_fab_vars[MicVar::qci]->nGrowVect());
+  amrex::MultiFab::Copy(qmoist, *mic_fab_vars[MicVar::qv],  0, 0, 1, mic_fab_vars[MicVar::qv]->nGrowVect());  // vapor
+  amrex::MultiFab::Copy(qmoist, *mic_fab_vars[MicVar::qcl], 0, 1, 1, mic_fab_vars[MicVar::qcl]->nGrowVect()); // cloud water
+  amrex::MultiFab::Copy(qmoist, *mic_fab_vars[MicVar::qci], 0, 2, 1, mic_fab_vars[MicVar::qci]->nGrowVect()); // cloud ice
+  amrex::MultiFab::Copy(qmoist, *mic_fab_vars[MicVar::qpl], 0, 3, 1, mic_fab_vars[MicVar::qpl]->nGrowVect()); // rain
+  amrex::MultiFab::Copy(qmoist, *mic_fab_vars[MicVar::qpi], 0, 4, 1, mic_fab_vars[MicVar::qpi]->nGrowVect()); // snow
 
-  // get the temperature, density, theta, qt and qp from input
-  for ( amrex::MFIter mfi(cons_in,amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-     auto states_array = cons_in.array(mfi);
[…]
