+++ Source/TimeIntegration/TI_headers.H
@@ -23,7 +23,7 @@ void erf_slow_rhs_pre(int level, int nrk,
                       const amrex::MultiFab& yvel,
                       const amrex::MultiFab& zvel,
 #if defined(ERF_USE_MOISTURE)
-                     const amrex::MultiFab& qvapor,
+                     const amrex::MultiFab& qmoist,
 #endif
                       std::unique_ptr<amrex::MultiFab>& z_t,
                             amrex::MultiFab& Omega,
@@ -196,9 +196,7 @@ void make_buoyancy(amrex::BoxArray& grids_to_evolve,
                    const           amrex::MultiFab & S_prim,
                          amrex::MultiFab& buoyancy,
 #if defined(ERF_USE_MOISTURE)
-                   const amrex::MultiFab& qvapor,
-                   const amrex::MultiFab& qcloud,
-                   const amrex::MultiFab& qice,
+                   const amrex::MultiFab& qmoist,
                    const amrex::Gpu::DeviceVector<amrex::Real> qv_d,
                    const amrex::Gpu::DeviceVector<amrex::Real> qc_d,
                    const amrex::Gpu::DeviceVector<amrex::Real> qi_d,
