+++ Source/TimeIntegration/ERF_advance_dycore.cpp
@@ -37,9 +37,7 @@ using namespace amrex;
  * @param[in] zmom_crse old-time z-component of momentum at coarser level
  * @param[in] source source term for conserved variables
  * @param[in] buoyancy buoyancy source term for z-component of momentum
- * @param[in] qvapor water vapor
- * @param[in] qcloud cloud water
- * @param[in] qice cloud ice
+ * @param[in] qmoist {water vapor, cloud water, cloud ice, rain, snow, graupel}
  * @param[in] fine_geom container for geometry information at current level
  * @param[in] dt_advance time step for this time advance
  * @param[in] old_time old time for this time advance
@@ -55,7 +53,7 @@ void ERF::advance_dycore(int level,
                          MultiFab& xmom_crse, MultiFab& ymom_crse, MultiFab& zmom_crse,
                          MultiFab& source, MultiFab& buoyancy,
 #if defined(ERF_USE_MOISTURE)
-                         const MultiFab& qvapor, const MultiFab& qcloud, const MultiFab& qice,
+                         const MultiFab& qmoist,
 #endif
                          const amrex::Geometry fine_geom,
                          const amrex::Real dt_advance, const amrex::Real old_time,
@@ -69,6 +67,12 @@ void ERF::advance_dycore(int level,
     MultiFab r_hse (base_state[level], make_alias, 0, 1); // r_0 is first  component
     MultiFab p_hse (base_state[level], make_alias, 1, 1); // p_0 is second component
     MultiFab pi_hse(base_state[level], make_alias, 2, 1); // pi_0 is second component
+
+#if defined(ERF_USE_MOISTURE)
+    MultiFab qvapor (qmoist, make_alias, 0, 1);
+    MultiFab qcloud (qmoist, make_alias, 1, 1);
+    MultiFab qice   (qmoist, make_alias, 2, 1);
+#endif
 
     MultiFab* r0  = &r_hse;
     MultiFab* p0  = &p_hse;
