### 2023-06-01 983a7bf
Add read and write of mapfactors.
