+++ Source/IO/Checkpoint.cpp
@@ -139,29 +139,51 @@ ERF::WriteCheckpointFile () const
            MultiFab::Copy(z_height,*z_phys_nd[lev],0,0,1,ngvect);
            VisMF::Write(z_height, amrex::MultiFabFileFullPrefix(lev, checkpointname, "Level_", "Z_Phys_nd"));
        }
+
+       // Note that we also write the ghost cells of the mapfactors (2D)
+       BoxList bl2d = grids[lev].boxList();
+       for (auto& b : bl2d) {
+           b.setRange(2,0);
+       }
+       BoxArray ba2d(std::move(bl2d));
+
+       IntVect ngvect_mf = mapfac_m[lev]->nGrowVect();
+       MultiFab mf_m(ba2d,dmap[lev],1,ngvect_mf);
+       MultiFab::Copy(mf_m,*mapfac_m[lev],0,0,1,ngvect_mf);
+       VisMF::Write(mf_m, amrex::MultiFabFileFullPrefix(lev, checkpointname, "Level_", "MapFactor_m"));
+
+       ngvect_mf = mapfac_u[lev]->nGrowVect();
+       MultiFab mf_u(convert(ba2d,IntVect(1,0,0)),dmap[lev],1,ngvect_mf);
+       MultiFab::Copy(mf_u,*mapfac_u[lev],0,0,1,ngvect_mf);
+       VisMF::Write(mf_u, amrex::MultiFabFileFullPrefix(lev, checkpointname, "Level_", "MapFactor_u"));
+
+       ngvect_mf = mapfac_v[lev]->nGrowVect();
+       MultiFab mf_v(convert(ba2d,IntVect(0,1,0)),dmap[lev],1,ngvect_mf);
+       MultiFab::Copy(mf_v,*mapfac_v[lev],0,0,1,ngvect_mf);
+       VisMF::Write(mf_v, amrex::MultiFabFileFullPrefix(lev, checkpointname, "Level_", "MapFactor_v"));
    }
 
    // Write bdy_data files
    if (ParallelDescriptor::IOProcessor() && (init_type == "real")) {
+
      // Vector dimensions
      int num_time = bdy_data_xlo.size();
      int num_var  = bdy_data_xlo[0].size();
 
      // Open header file and write to it
      std::ofstream bdy_h_file(amrex::MultiFabFileFullPrefix(0, checkpointname, "Level_", "bdy_H"));
+     bdy_h_file << std::setprecision(1) << std::fixed;
      bdy_h_file << num_time << "\n";
      bdy_h_file << num_var  << "\n";
      bdy_h_file << start_bdy_time << "\n";
      bdy_h_file << bdy_time_interval << "\n";
      bdy_h_file << wrfbdy_width << "\n";
-     bdy_h_file << "\n";
      for (int ivar(0); ivar<num_var; ++ivar) {
        bdy_h_file << bdy_data_xlo[0][ivar].box() << "\n";
        bdy_h_file << bdy_data_xhi[0][ivar].box() << "\n";
        bdy_h_file << bdy_data_ylo[0][ivar].box() << "\n";
        bdy_h_file << bdy_data_yhi[0][ivar].box() << "\n";
      }
-     bdy_h_file << "\n";
 
      // Open data file and write to it
      std::ofstream bdy_d_file(amrex::MultiFabFileFullPrefix(0, checkpointname, "Level_", "bdy_D"));
@@ -304,74 +326,98 @@ ERF::ReadCheckpointFile ()
            MultiFab z_height(convert(grids[lev],IntVect(1,1,1)),dmap[lev],1,ngvect);
[…]
