### 2023-06-01 5d3fb12
ifdef netcdf for bdy write.
