### 2023-06-02 5ec9c45
Worked on Ghidorah with 4 procs.
