### 2023-06-02 2bac841
Wps res fix (#1109)
* Check this on another machine.

* Worked on Ghidorah with 4 procs.

* restore inputs

---------

Co-authored-by: Aaron Lattanzi <amlattanzi@ghidorah.dhcp.lbl.gov>
