+++ Source/IO/Checkpoint.cpp
@@ -355,63 +355,90 @@ ERF::ReadCheckpointFile ()
 
 #ifdef ERF_USE_NETCDF
     // Read bdy_data files
-    if (ParallelDescriptor::IOProcessor() && (init_type == "real")) {
+    if (init_type == "real") {
         int ioproc = ParallelDescriptor::IOProcessorNumber();  // I/O rank
-
-        // Vector dimensions
         int num_time;
         int num_var;
         Vector<Box> bx_v;
+        if (ParallelDescriptor::IOProcessor()) {
+            // Open header file and read from it
+            std::ifstream bdy_h_file(amrex::MultiFabFileFullPrefix(0, restart_chkfile, "Level_", "bdy_H"));
+            bdy_h_file >> num_time;
+            bdy_h_file >> num_var;
+            bdy_h_file >> start_bdy_time;
+            bdy_h_file >> bdy_time_interval;
+            bdy_h_file >> wrfbdy_width;
+            bx_v.resize(4*num_var);
+            for (int ivar(0); ivar<num_var; ++ivar) {
+                bdy_h_file >> bx_v[4*ivar  ];
+                bdy_h_file >> bx_v[4*ivar+1];
+                bdy_h_file >> bx_v[4*ivar+2];
+                bdy_h_file >> bx_v[4*ivar+3];
+            }
 
-        // Open header file and read from it
-        std::ifstream bdy_h_file(amrex::MultiFabFileFullPrefix(0, restart_chkfile, "Level_", "bdy_H"));
-        bdy_h_file >> num_time;
-        bdy_h_file >> num_var;
-        bdy_h_file >> start_bdy_time;
-        bdy_h_file >> bdy_time_interval;
-        bdy_h_file >> wrfbdy_width;
+            // IO size the FABs
+            bdy_data_xlo.resize(num_time);
+            bdy_data_xhi.resize(num_time);
+            bdy_data_ylo.resize(num_time);
+            bdy_data_yhi.resize(num_time);
+            for (int itime(0); itime<num_time; ++itime) {
+                bdy_data_xlo[itime].resize(num_var);
+                bdy_data_xhi[itime].resize(num_var);
+                bdy_data_ylo[itime].resize(num_var);
+                bdy_data_yhi[itime].resize(num_var);
+                for (int ivar(0); ivar<num_var; ++ivar) {
+                    bdy_data_xlo[itime][ivar].resize(bx_v[4*ivar  ]);
+                    bdy_data_xhi[itime][ivar].resize(bx_v[4*ivar+1]);
+                    bdy_data_ylo[itime][ivar].resize(bx_v[4*ivar+2]);
+                    bdy_data_yhi[itime][ivar].resize(bx_v[4*ivar+3]);
+                }
+            }
+
+            // Open data file and read from it
+            std::ifstream bdy_d_file(amrex::MultiFabFileFullPrefix(0, restart_chkfile, "Level_", "bdy_D"));
+            for (int itime(0); itime<num_time; ++itime) {
+                for (int ivar(0); ivar<num_var; ++ivar) {
+                    bdy_data_xlo[itime][ivar].readFrom(bdy_d_file);
+                    bdy_data_xhi[itime][ivar].readFrom(bdy_d_file);
+                    bdy_data_ylo[itime][ivar].readFrom(bdy_d_file);
+                    bdy_data_yhi[itime][ivar].readFrom(bdy_d_file);
+                }
+            }
+        } // IO
+
+        // Broadcast the data
+        ParallelDescriptor::Barrier();
[…]
