+++ Source/BoundaryConditions/ERF_PhysBCFunct.H
@@ -56,25 +56,25 @@ public:
     */
     void operator() (const amrex::Vector<amrex::MultiFab*>& mfs, int icomp, int ncomp,
                      amrex::IntVect const& nghost_cons, amrex::IntVect const& nghost_vels,
-                     amrex::Real time, std::string& init_type, bool cons_only);
+                     amrex::Real time, std::string& init_type, bool cons_only, int bcoff=0);
 
     void impose_lateral_xvel_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                   const amrex::Box& bx, const amrex::Box& domain,
-                                  amrex::Real time, int bccomp);
+                                  amrex::Real time, int bccomp, int bcoff=0);
     void impose_vertical_xvel_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                   const amrex::Box& bx, const amrex::Box& domain,
                                    const amrex::Array4<amrex::Real const>& z_nd,
                                    const amrex::GpuArray<amrex::Real,AMREX_SPACEDIM> dxInv,
-                                   amrex::Real time, int bccomp);
+                                   amrex::Real time, int bccomp, int bcoff=0);
 
     void impose_lateral_yvel_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                   const amrex::Box& bx, const amrex::Box& domain,
-                                  amrex::Real time, int bccomp);
+                                  amrex::Real time, int bccomp, int bcoff=0);
     void impose_vertical_yvel_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                    const amrex::Box& bx, const amrex::Box& domain,
                                    const amrex::Array4<amrex::Real const>& z_nd,
                                    const amrex::GpuArray<amrex::Real,AMREX_SPACEDIM> dxInv,
-                                   amrex::Real time, int bccomp);
+                                   amrex::Real time, int bccomp, int bcoff=0);
 
 
     void impose_lateral_zvel_bcs (const amrex::Array4<amrex::Real>& dest_arr,
@@ -83,7 +83,7 @@ public:
                                   const amrex::Box& bx, const amrex::Box& domain,
                                   const amrex::Array4<amrex::Real const>& z_nd,
                                   const amrex::GpuArray<amrex::Real,AMREX_SPACEDIM> dxInv,
-                                  amrex::Real time, int bccomp_w);
+                                  amrex::Real time, int bccomp_w, int bcoff=0);
     void impose_vertical_zvel_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                    const amrex::Array4<amrex::Real const>& xvel_arr,
                                    const amrex::Array4<amrex::Real const>& yvel_arr,
@@ -91,16 +91,16 @@ public:
                                    const amrex::Array4<amrex::Real const>& z_nd,
                                    const amrex::GpuArray<amrex::Real,AMREX_SPACEDIM> dxInv,
                                    amrex::Real time, int bccomp_u, int bccomp_v, int bccomp_w,
-                                   int terrain_type);
+                                   int terrain_type, int bcoff=0);
 
     void impose_lateral_cons_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                   const amrex::Box& bx, const amrex::Box& domain,
-                                  int icomp, int ncomp, amrex::Real time, int bccomp);
+                                  int icomp, int ncomp, amrex::Real time, int bccomp, int bcoff=0);
     void impose_vertical_cons_bcs (const amrex::Array4<amrex::Real>& dest_arr,
                                    const amrex::Box& bx, const amrex::Box& domain,
                                    const amrex::Array4<amrex::Real const>& z_nd,
[…]
