### 2023-06-03 8599679
Modify bccomp for moist FP.
