+++ Source/BoundaryConditions/ERF_FillPatch.cpp
@@ -89,7 +89,7 @@ ERF::FillPatch (int lev, Real time, const Vector<MultiFab*>& mfs)
     if (m_r2d) fill_from_bndryregs(mfs,time);
 
     // We call this even if init_type == real because this routine will fill the vertical bcs
-    (*physbcs[lev])(mfs,icomp_cons,ncomp_cons,ngvect_cons,ngvect_vels,time,init_type,cons_only);
+    (*physbcs[lev])(mfs,icomp_cons,ncomp_cons,ngvect_cons,ngvect_vels,time,init_type,cons_only,BCVars::cons_bc);
 }
 
 /*
@@ -125,7 +125,7 @@ ERF::FillPatchMoistVars (int lev, Real time, MultiFab& mf)
     IntVect ngvect_vels = {0,0,0};
 
     if (init_type != "real") {
-       (*physbcs[lev])({&mf},icomp_cons,ncomp_cons,ngvect_cons,ngvect_vels,time,init_type,cons_only);
+        (*physbcs[lev])({&mf},icomp_cons,ncomp_cons,ngvect_cons,ngvect_vels,time,init_type,cons_only,BCVars::RhoQt_bc_comp);
     }
 
     mf.FillBoundary(geom[lev].periodicity());
@@ -236,7 +236,7 @@ ERF::FillIntermediatePatch (int lev, Real time,
     if (m_r2d) fill_from_bndryregs(mfs,time);
 
     // We call this even if init_type == real because this routine will fill the vertical bcs
-    (*physbcs[lev])(mfs,icomp_cons,ncomp_cons,ngvect_cons,ngvect_vels,time,init_type,cons_only);
+    (*physbcs[lev])(mfs,icomp_cons,ncomp_cons,ngvect_cons,ngvect_vels,time,init_type,cons_only,BCVars::cons_bc);
     // ***************************************************************************
 
     //
@@ -308,7 +308,7 @@ ERF::FillCoarsePatch (int lev, Real time, const Vector<MultiFab*>& mfs)
     IntVect ngvect_vels = mfs[Vars::xvel]->nGrowVect();
     bool cons_only = false;
 
-    (*physbcs[lev])(mfs,0,mfs[Vars::cons]->nComp(),ngvect_cons,ngvect_vels,time,init_type,cons_only);
+    (*physbcs[lev])(mfs,0,mfs[Vars::cons]->nComp(),ngvect_cons,ngvect_vels,time,init_type,cons_only,BCVars::cons_bc);
 
     // ***************************************************************************
     // Since lev > 0 here we don't worry about m_r2d or wrfbdy data
