### 2023-06-03 4a26009
fix WPS restart and clean up some calls
