+++ Source/BoundaryConditions/BoundaryConditions_zvel.cpp
@@ -11,7 +11,6 @@ using namespace amrex;
  * @param[in] bx        box associated with this data
  * @param[in] domain    computational domain
  * @param[in] z_phys_nd height coordinate at nodes
- * @param[in] time      time at which the data should be filled
  * @param[in] bccomp    index into m_domain_bcs_type
  */
 void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr,
@@ -20,7 +19,7 @@ void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr,
                                               const Box& bx, const Box& domain,
                                               const Array4<Real const>& z_phys_nd,
                                               const GpuArray<Real,AMREX_SPACEDIM> dxInv,
-                                              Real /*time*/, int bccomp)
+                                              int bccomp)
 {
     BL_PROFILE_VAR("impose_lateral_zvel_bcs()",impose_lateral_zvel_bcs);
     const auto& dom_lo = amrex::lbound(domain);
@@ -147,7 +146,6 @@ void ERFPhysBCFunct::impose_lateral_zvel_bcs (const Array4<Real>& dest_arr,
  * @param[in] domain    computational domain
  * @param[in] z_phys_nd height coordinate at nodes
  * @param[in] dxInv     inverse cell size array
- * @param[in] time      time at which the data should be filled
  * @param[in] bccomp_u  index into m_domain_bcs_type corresponding to u
  * @param[in] bccomp_v  index into m_domain_bcs_type corresponding to v
  * @param[in] bccomp_w  index into m_domain_bcs_type corresponding to w
@@ -160,8 +158,7 @@ void ERFPhysBCFunct::impose_vertical_zvel_bcs (const Array4<Real>& dest_arr,
                                                const Box& bx, const Box& domain,
                                                const Array4<Real const>& z_phys_nd,
                                                const GpuArray<Real,AMREX_SPACEDIM> dxInv,
-                                               Real /*time*/, int bccomp_u,
-                                               int bccomp_v, int bccomp_w,
+                                               int bccomp_u, int bccomp_v, int bccomp_w,
                                                int terrain_type)
 {
     BL_PROFILE_VAR("impose_vertical_zvel_bcs()",impose_vertical_zvel_bcs);
