+++ Source/BoundaryConditions/ERF_PhysBCFunct.cpp
@@ -12,13 +12,12 @@ using namespace amrex;
  * @param[in] ncomp_cons  number of components for conserved variables
  * @param[in] nghost_cons number of ghost cells to be filled for conserved variables
  * @param[in] nghost_vels number of ghost cells to be filled for velocity components
- * @param[in] time        time at which the data should be filled
  * @param[in] init_type   if "real" then we fill boundary conditions for interior locations
  * @param[in] cons_only   if 1 then only fill conserved variables
  */
 
 void ERFPhysBCFunct::operator() (const Vector<MultiFab*>& mfs, int icomp_cons, int ncomp_cons,
-                                 IntVect const& nghost_cons, IntVect const& nghost_vels, Real time,
+                                 IntVect const& nghost_cons, IntVect const& nghost_vels,
                                  std::string& init_type, bool cons_only, int bccomp_cons)
 {
     BL_PROFILE("ERFPhysBCFunct::()");
@@ -90,7 +89,7 @@ void ERFPhysBCFunct::operator() (const Vector<MultiFab*>& mfs, int icomp_cons, i
                 //! we need to fill them here
                 if (!gdomain.contains(cbx))
                 {
-                    impose_lateral_cons_bcs(cons_arr,cbx,domain,icomp_cons,ncomp_cons,time,bccomp_cons);
+                    impose_lateral_cons_bcs(cons_arr,cbx,domain,icomp_cons,ncomp_cons,bccomp_cons);
                 }
 
                 if (!cons_only)
@@ -100,15 +99,15 @@ void ERFPhysBCFunct::operator() (const Vector<MultiFab*>& mfs, int icomp_cons, i
                     const Array4<      Real> velz_arr = mfs[Vars::zvel]->array(mfi);;
                     if (!gdomainx.contains(xbx))
                     {
-                        impose_lateral_xvel_bcs(velx_arr,xbx,domain,time,BCVars::xvel_bc);
+                        impose_lateral_xvel_bcs(velx_arr,xbx,domain,BCVars::xvel_bc);
                     }
 
                     if (!gdomainy.contains(ybx))
                     {
-                        impose_lateral_yvel_bcs(vely_arr,ybx,domain,time,BCVars::yvel_bc);
+                        impose_lateral_yvel_bcs(vely_arr,ybx,domain,BCVars::yvel_bc);
                     }
 
-                    impose_lateral_zvel_bcs(velz_arr,velx_arr,vely_arr,zbx,domain,z_nd_arr,dxInv,time,BCVars::zvel_bc);
+                    impose_lateral_zvel_bcs(velz_arr,velx_arr,vely_arr,zbx,domain,z_nd_arr,dxInv,BCVars::zvel_bc);
                 } // !cons_only
             } // init_type != "real"
 
@@ -116,7 +115,7 @@ void ERFPhysBCFunct::operator() (const Vector<MultiFab*>& mfs, int icomp_cons, i
             //       for every box -- and we need to call these even if init_type == real
             {
             impose_vertical_cons_bcs(cons_arr,cbx,domain,z_nd_arr,dxInv,
-                                     icomp_cons,ncomp_cons,time,bccomp_cons);
+                                     icomp_cons,ncomp_cons,bccomp_cons);
             }
 
             if (!cons_only) {
@@ -124,10 +123,10 @@ void ERFPhysBCFunct::operator() (const Vector<MultiFab*>& mfs, int icomp_cons, i
                 const Array4<      Real> vely_arr = mfs[Vars::yvel]->array(mfi);;
                 const Array4<      Real> velz_arr = mfs[Vars::zvel]->array(mfi);;
                 impose_vertical_xvel_bcs(velx_arr,xbx,domain,z_nd_arr,dxInv,
[…]
