### 2023-06-03 6f1c465
fix WPS restart and clean up some calls (#1113)
