+++ Source/Initialization/ERF_init_from_metgrid.cpp
@@ -57,9 +57,9 @@ init_msfs_from_metgrid(FArrayBox& msfu_fab,
 void
 init_base_state_from_metgrid(const Real l_rdOcp,
                              FArrayBox& state,
-                             amrex::Array4<amrex::Real> const &r_hse_arr,
-                             amrex::Array4<amrex::Real> const &p_hse_arr,
-                             amrex::Array4<amrex::Real> const &pi_hse_arr,
+                             FArrayBox& r_hse_fab,
+                             FArrayBox& p_hse_fab,
+                             FArrayBox& pi_hse_fab,
                              FArrayBox& z_phys_nd_fab,
                              const Vector<FArrayBox>& NC_ght_fab,
                              const Vector<FArrayBox>& NC_psfc_fab);
@@ -151,18 +151,25 @@ ERF::init_from_metgrid(int lev)
     // This defines all the z(i,j,k) values given z(i,j,0) from above.
     init_terrain_grid(geom[lev], *z_phys);
 
+#ifndef AMREX_USE_GPU
     const amrex::Box& domain = geom[lev].Domain();
-//    auto dx = geom[lev].CellSizeArray();
-//    amrex::Print() << " geom.CellSizeArray(): " << dx << std::endl;
-    amrex::Print() << " geom[" << lev << "].Domain.smallEnd(0): " << domain.smallEnd(0) << "    geom[" << lev << "Domain.bigEnd(0): " << domain.bigEnd(0) << std::endl;
-    amrex::Print() << " geom[" << lev << "].Domain.smallEnd(1): " << domain.smallEnd(1) << "    geom[" << lev << "Domain.bigEnd(1): " << domain.bigEnd(1) << std::endl;
-    amrex::Print() << " geom[" << lev << "].Domain.smallEnd(2): " << domain.smallEnd(2) << "    geom[" << lev << "Domain.bigEnd(2): " << domain.bigEnd(2) << std::endl;
+    auto dx = geom[lev].CellSizeArray();
+    for (int i = 0; i < dx.size(); i++)
+    {
+        amrex::Print() << " geom.CellSizeArray()[" << i << "]: " << dx[i] << std::endl;
+    }
+    amrex::Print() << " geom[" << lev << "].Domain.smallEnd(0): " << domain.smallEnd(0) << "    geom[" << lev << "].Domain.bigEnd(0): " << domain.bigEnd(0) << std::endl;
+    amrex::Print() << " geom[" << lev << "].Domain.smallEnd(1): " << domain.smallEnd(1) << "    geom[" << lev << "].Domain.bigEnd(1): " << domain.bigEnd(1) << std::endl;
+    amrex::Print() << " geom[" << lev << "].Domain.smallEnd(2): " << domain.smallEnd(2) << "    geom[" << lev << "].Domain.bigEnd(2): " << domain.bigEnd(2) << std::endl;
     amrex::Print() << " geom[" << lev << "].Domain.length(0): " << domain.length(0) << std::endl;
     amrex::Print() << " geom[" << lev << "].Domain.length(1): " << domain.length(1) << std::endl;
     amrex::Print() << " geom[" << lev << "].Domain.length(2): " << domain.length(2) << std::endl;
+#endif
+
+    // TODO: Verify that DX and DY from met_em file matches that in geom (from ERF inputs file).
 
     // This makes the Jacobian.
-    make_J  (geom[lev],*z_phys,  *detJ_cc[lev]);
+    make_J(geom[lev],*z_phys,  *detJ_cc[lev]);
 
[…]
