### 2023-06-12 39326a3
WIP metgrid initialization now checks met_em global attr
