+++ Source/Initialization/ERF_init_from_metgrid.cpp
@@ -12,6 +12,10 @@ using namespace amrex;
 
 void
 read_from_metgrid(int lev, const std::string& fname,
+                  std::string& NC_datetime,
+                  int& flag_psfc, int& flag_msfu, int& flag_msfv, int& flag_msfm,
+                  int& flag_hgt,  int& NC_nx,     int& NC_ny,
+                  Real& NC_dx,    Real& NC_dy,
                   FArrayBox& NC_xvel_fab, FArrayBox& NC_yvel_fab,
                   FArrayBox& NC_temp_fab, FArrayBox& NC_rhum_fab,
                   FArrayBox& NC_pres_fab, FArrayBox& NC_ght_fab,
@@ -50,12 +54,16 @@ void
 init_msfs_from_metgrid(FArrayBox& msfu_fab,
                        FArrayBox& msfv_fab,
                        FArrayBox& msfm_fab,
+                       const int& flag_msfu,
+                       const int& flag_msfv,
+                       const int& flag_msfm,
                        const Vector<FArrayBox>& NC_MSFU_fab,
                        const Vector<FArrayBox>& NC_MSFV_fab,
                        const Vector<FArrayBox>& NC_MSFM_fab);
 
 void
 init_base_state_from_metgrid(const Real l_rdOcp,
+                             const int& flag_psfc,
                              FArrayBox& state,
                              FArrayBox& r_hse_fab,
                              FArrayBox& p_hse_fab,
@@ -73,6 +81,7 @@ rh_to_mxrat(int i, int j, int k,
 
 void
 calc_rho_p(int i, int j,
+           const int& flag_psfc,
            const Array4<Real const>& psfc,
            const Array4<Real const>& new_z,
            const Array4<Real>& new_data,
@@ -89,12 +98,14 @@ void
 ERF::init_from_metgrid(int lev)
 {
 #if defined(ERF_USE_MOISTURE)
+    amrex::Print() << "Init with met_em with ERF_USE_MOISTURE" << std::endl;
 #elif defined(ERF_USE_WARM_NO_PRECIP)
+    amrex::Print() << "Init with met_em with ERF_USE_WARM_NO_PRECIP" << std::endl;
 #else
-    amrex::Error("Init with met_em requires either ERF_USE_MOISTURE or ERF_USE_WARM_NO_PRECIP");
+    amrex::Print() << "Init with met_em without moisture" << std::endl;
 #endif
 
-    // *** FArrayBox's at this level for holding the INITIAL data
+    // *** FArrayBox's at this level for holding the metgrid data
     Vector<FArrayBox> NC_xvel_fab;  NC_xvel_fab.resize(num_boxes_at_level[lev]);
     Vector<FArrayBox> NC_yvel_fab;  NC_yvel_fab.resize(num_boxes_at_level[lev]);
     Vector<FArrayBox> NC_temp_fab;  NC_temp_fab.resize(num_boxes_at_level[lev]);
@@ -117,12 +128,40 @@ ERF::init_from_metgrid(int lev)
     if (nc_init_file[lev].empty())
         amrex::Error("NetCDF initialization file name must be provided via input");
 
+    // *** Variables at this level for holding metgrid file global attributes
+    Vector<int> flag_psfc;           flag_psfc.resize(num_boxes_at_level[lev]);
+    Vector<int> flag_msfu;           flag_msfu.resize(num_boxes_at_level[lev]);
+    Vector<int> flag_msfv;           flag_msfv.resize(num_boxes_at_level[lev]);
+    Vector<int> flag_msfm;           flag_msfm.resize(num_boxes_at_level[lev]);
[…]
