### 2023-06-14 ddf9957
include GPU-aware MPI information for running on Perlmutter and update performance plot
