+++ Docs/sphinx_doc/index.rst
@@ -92,7 +92,6 @@ In addition to this documentation, there is API documentation for ERF generated 
 
    Performance.rst
    RegressionTests.rst
-   Performance.rst
 
 .. toctree::
    :caption: GOALS
