### 2023-06-23 e95eb6c
WIP started writing procedures for setting lateral boundary conditions from met_em files.
