+++ Source/Initialization/ERF_init_from_metgrid.cpp
@@ -12,7 +12,7 @@ using namespace amrex;
 
 void
 read_from_metgrid(int lev, const std::string& fname,
-                  std::string& NC_datetime,
+                  std::string& NC_dateTime, Real& NC_epochTime,
                   int& flag_psfc, int& flag_msfu, int& flag_msfv, int& flag_msfm,
                   int& flag_hgt,  int& NC_nx,     int& NC_ny,
                   Real& NC_dx,    Real& NC_dy,
@@ -38,7 +38,8 @@ init_terrain_from_metgrid(FArrayBox& z_phys_nd_fab,
                           const Vector<FArrayBox>& NC_hgt_fab);
 
 void
-init_state_from_metgrid(FArrayBox& state_fab,
+init_state_from_metgrid(const Real l_rdOcp,
+                        FArrayBox& state_fab,
                         FArrayBox& x_vel_fab, FArrayBox& y_vel_fab,
                         FArrayBox& z_vel_fab, FArrayBox& z_phys_nd_fab,
                         const Vector<FArrayBox>& NC_hgt_fab,
@@ -48,6 +49,7 @@ init_state_from_metgrid(FArrayBox& state_fab,
                         const Vector<FArrayBox>& NC_zvel_fab,
                         const Vector<FArrayBox>& NC_temp_fab,
                         const Vector<FArrayBox>& NC_rhum_fab,
+                              Vector<FArrayBox>& theta_fab,
                               Vector<FArrayBox>& mxrat_fab);
 
 void
@@ -63,7 +65,8 @@ init_msfs_from_metgrid(FArrayBox& msfu_fab,
 
 void
 init_base_state_from_metgrid(const Real l_rdOcp,
-                             const int& flag_psfc,
+                             const Box& valid_bx,
+                             const Vector<int>& flag_psfc,
                              FArrayBox& state,
                              FArrayBox& r_hse_fab,
                              FArrayBox& p_hse_fab,
@@ -105,22 +108,8 @@ ERF::init_from_metgrid(int lev)
     amrex::Print() << "Init with met_em without moisture" << std::endl;
 #endif
 
-    // *** FArrayBox's at this level for holding the metgrid data
-    Vector<FArrayBox> NC_xvel_fab;  NC_xvel_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_yvel_fab;  NC_yvel_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_temp_fab;  NC_temp_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_rhum_fab;  NC_rhum_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_pres_fab;  NC_pres_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_ght_fab;   NC_ght_fab.resize(num_boxes_at_level[lev]);
-
-    Vector<FArrayBox> NC_hgt_fab;   NC_hgt_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_psfc_fab;  NC_psfc_fab.resize(num_boxes_at_level[lev]);
-
-    Vector<FArrayBox> NC_MSFU_fab;  NC_MSFU_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_MSFV_fab;  NC_MSFV_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_MSFM_fab;  NC_MSFM_fab.resize(num_boxes_at_level[lev]);
-
[…]
