### 2023-06-23 #5202 @kthyng
@pratikvn Could you summarize if you still have any outstanding questions/comments for this JOSS submission?

@rafmudaf Great to see you back!
