### 2023-06-23 erf-model/ERF#1120 @rafmudaf
@asalmgren I've added some additional comments to the original post above, just heads up!
