### 2023-06-23 d1f6dc6
Write out what AMR stands for (#1121)
