+++ Docs/sphinx_doc/index.rst
@@ -32,6 +32,7 @@ In addition to this documentation, there is API documentation for ERF generated 
    :hidden:
 
    coc
+   Doxygen
 
 .. toctree::
    :caption: USER GUIDE
