+++ Docs/sphinx_doc/Doxygen.rst
@@ -1,0 +1,11 @@
+
+ .. role:: cpp(code)
+    :language: c++
+
+.. _doxygen_link:
+
+Doxygen Documentation
+=====================
+
+For details about the ERF code structure and APIs,
+see the `Doxygen documentation <https://erf-model.github.io/docs/index.html>`_
