### 2023-06-24 431b3cf
Add reference for WRF technote to JOSS paper. (#1123)
