+++ paper/paper.bib
@@ -117,3 +117,14 @@
          year={2017},
          publisher={Wiley Online Library}
 }
+
+@techreport{WRF:skamarock,
+	doi = {10.5065/1dfh-6p97},
+	url = {https://dx.doi.org/10.5065/1dfh-6p97},
+	title = {A Description of the Advanced Research WRF Model Version 4.3 (No. NCAR/TN-556+STR)},
+	author = {Skamarock, W.~C. and Klemp, J.~B. and Dudhia, J. and Gill, D.~O. and Liu, Z. and Berner, J. and Wang, W. and Powers, J.~G. and Duda, M.~G. and Barker, D.~M. and Huang, X.},
+	institution = {National Center for Atmospheric Research},
+	month = {7},
+	year = {2021},
+	number = {NCAR/TN-556+STR}
+}
