+++ paper/paper.md
@@ -126,7 +126,7 @@ Physical forcings include Coriolis and geostrophic forcing as well as
 Rayleigh damping in the upper regions of the domain.  Lateral boundary
 conditions can be specified as periodic, inflow/outflow, or time-varying
 values read in from external files in netcdf format generated by the WRF
-Preprocessing System (WPS). The surface boundary condition may be
+Preprocessing System (WPS) [@WRF:Skamarock]. The surface boundary condition may be
 specified either as a simple wall or by using Monin-Obukhov similarity theory (MOST)
 [@MOST:monin; @MOST:van] to model the surface layer. The initial data can
 be read from WPS-generated files, reconstructed from 1-d input sounding
