### 2023-06-25 9dfd7e0
Add discussion of energy applications
