### 2023-06-25 4d052c3
Revert "update cmake for CI"
This reverts commit a74d7ea3aa56ab2a1a19262a6232d00f5792387e.
