+++ CMake/SetAmrexCompileFlags.cmake
@@ -1,0 +1,8 @@
+# Disable loop not vectorized warnings on Clang. This generates a lot of
+# diagnostic messages when compiling AMReX that we can't do anything about
+# within amr-wind
+if (${CMAKE_CXX_COMPILER_ID} STREQUAL "Clang" OR
+    ${CMAKE_CXX_COMPILER_ID} STREQUAL "AppleClang")
+  target_compile_options(
+    amrex_3d PUBLIC $<$<COMPILE_LANGUAGE:CXX>:-Wno-pass-failed>)
+endif()
