+++ Source/Initialization/ERF_init_from_metgrid.cpp
@@ -50,7 +50,8 @@ init_state_from_metgrid(const Real l_rdOcp,
                         const Vector<FArrayBox>& NC_temp_fab,
                         const Vector<FArrayBox>& NC_rhum_fab,
                               Vector<FArrayBox>& theta_fab,
-                              Vector<FArrayBox>& mxrat_fab);
+                              Vector<FArrayBox>& mxrat_fab,
+                        Vector<Vector<FArrayBox>>& fabs_for_bcs);
 
 void
 init_msfs_from_metgrid(FArrayBox& msfu_fab,
@@ -73,7 +74,8 @@ init_base_state_from_metgrid(const Real l_rdOcp,
                              FArrayBox& pi_hse_fab,
                              FArrayBox& z_phys_nd_fab,
                              const Vector<FArrayBox>& NC_ght_fab,
-                             const Vector<FArrayBox>& NC_psfc_fab);
+                             const Vector<FArrayBox>& NC_psfc_fab,
+                             Vector<Vector<FArrayBox>>& fabs_for_bcs);
 
 void
 rh_to_mxrat(int i, int j, int k,
@@ -85,6 +87,9 @@ rh_to_mxrat(int i, int j, int k,
 void
 calc_rho_p(int i, int j,
            const int& flag_psfc,
+#if defined(ERF_USE_MOISTURE) || defined(ERF_USE_WARM_NO_PRECIP)
+           const int& RhoQ_comp,
+#endif
            const Array4<Real const>& psfc,
            const Array4<Real const>& new_z,
            const Array4<Real>& new_data,
@@ -186,19 +191,31 @@ ERF::init_from_metgrid(int lev)
     // Start at the earliest time in nc_init_file[lev].
     start_bdy_time = NC_epochTime[0];
 
+    // Determine the spacing between met_em files.
+    bdy_time_interval = NC_epochTime[1]-NC_epochTime[0];
+
+    // Verify that met_em files have even spacing in time.
+    for (int it = 1; it < ntimes; it++) {
+        Real NC_dt = NC_epochTime[it]-NC_epochTime[it-1];
+#ifndef AMREX_USE_GPU
+        amrex::Print() << " " << nc_init_file[lev][it-1] << " / " << nc_init_file[lev][it] << " are " << NC_dt << " seconds apart" << std::endl;
+#endif
+        if (NC_dt != bdy_time_interval) amrex::Error("Time interval between consecutive met_em files must be consistent.");
+    }
+
     // Set up a FAB for mixing ratio and another for potential temperature.
     // Our input data only has relative humidity and temperature.
     // TODO: add alternate pathways for other origin model cases where different variables may be present.
     Vector<FArrayBox> mxrat_fab; mxrat_fab.resize(ntimes);
     Vector<FArrayBox> theta_fab; theta_fab.resize(ntimes);
     for (int it = 0; it < ntimes; it++) {
-        Box my_box = NC_rhum_fab[it].box();
+        Box NC_box_unstag = NC_rhum_fab[it].box();
 #ifdef AMREX_USE_GPU
-        mxrat_fab[it].resize(my_box, 1, The_Pinned_Arena());
-        theta_fab[it].resize(my_box, 1, The_Pinned_Arena());
+        mxrat_fab[it].resize(NC_box_unstag, 1, The_Pinned_Arena());
+        theta_fab[it].resize(NC_box_unstag, 1, The_Pinned_Arena());
 #else
-        mxrat_fab[it].resize(my_box, 1);
-        theta_fab[it].resize(my_box, 1);
+        mxrat_fab[it].resize(NC_box_unstag, 1);
[…]
