+++ Exec/CMakeLists.txt
@@ -15,7 +15,7 @@ else ()
   add_subdirectory(EkmanSpiral_ideal)
   add_subdirectory(EkmanSpiral_input_sounding)
   add_subdirectory(IsentropicVortex)
-  add_subdirectory(MovingTerrain)
+  add_subdirectory(DevTests/MovingTerrain)
   add_subdirectory(PoiseuilleFlow)
   add_subdirectory(ScalarAdvDiff)
   add_subdirectory(TaylorGreenVortex)
