### 2023-06-27 45202df
Move test dirs into Exec/RegTests and update README
