+++ README.rst
@@ -12,74 +12,27 @@ Regression Tests    |regtests|
 =================  =============
 
 .. |regtests| image:: https://github.com/erf-model/ERF/actions/workflows/ci.yml/badge.svg?branch=development
-.. |unittests| image:: https://github.com/erf-model/ERF/actions/workflows/ci.yml/badge.svg?branch=development
-
 
 Getting Started
 ~~~~~~~~~~~~~~~
 
-* To compile and run the `ERF` suite of codes, one needs a C++ compiler that supports the C++14 standard.  A hierarchical strategy for parallelism is supported, based on MPI + OpenMP or MPI + CUDA.  The codes work with all major MPI and OpenMP implementations.  The codes should build and run with no modifications to the `make` system if using a Linux system with the GNU compilers, version 4.9.3 and above.
-
-To clone the source code of `ERF`:
-
-1. One can have ERF use the default submodule for ``AMReX`` in its own repo by simply performing: ::
-
-    export ERF_HOME=<location for ERF code>
-    git clone --recursive git@github.com:ERF-model/ERF.git ${ERF_HOME}
-
-Alternatively, one can set environment variable ``AMREX_HOME`` to use ``AMReX`` repo from external locations independent of ``ERF_HOME``: ::
-
-1. Set the environment variable ``AMREX_HOME`` and clone a copy of ``AMReX`` there: ::
-
-    export AMREX_HOME=<location for AMReX>  # Can be anywhere
-    git clone git@github.com:AMReX-Codes/amrex.git ${AMREX_HOME}
-
-2. Set the environment variable ``ERF_HOME`` and clone a copy of ``ERF`` there. You should be placed in the ``development`` branch: ::
-
-    export ERF_HOME=<location for ERF code>
-    git clone git@github.com:ERF-model/ERF.git ${ERF_HOME}
-
-Note that cloning using the format ``git@github.com:ERF-model/ERF.git`` requires that your public rsa keys are set in your github profile. See `here <https://docs.github.com/en/free-pro-team@latest/github/authenticating-to-github/adding-a-new-ssh-key-to-your-github-account>`_ how to do this.
-
-If one doesn't have the rsa keys setup in github profile, one can clone using ``https://github.com/erf-model/ERF.git`` format.
-
-To build the code and run a sample problem:
-
-1. Move to an example build folder, build an executable, and run a test case: ::
-
-    cd ${ERF_HOME}/Exec/ScalarAdvDiff
-    make
-    ./ERF3d.xxx.yyy.ex inputs_ex
-
-   In order to make on multiple processors use ``make -j N`` instead of ``make``, where ``N`` is the number of processors to use for building and is typically 8 or 16. The executable name is of the format ``ERF3d.xxx.yyy.ex``.
-
-* Notes:
-
-   A. In the exec line above, ``xxx.yyy`` is a tag identifying your compiler and various build options, for example, ``gnu.MPI`` and will vary across platform.  (Note that GNU compilers must be at least 4.8.4, and MPI should be at least version 3).
-   B. In addition to informative output to the terminal, periodic plot files are written in the run folder.  These may be viewed with CCSE's `Amrvis <https://ccse.lbl.gov/Downloads/downloadAmrvis.html>`_ or `Vis-It <http://vis.lbl.gov/NERSC/Software/visit/>`_ (!! FIX the URLs !!):
-
-      1. In Vis-It, direct the File->Open dialogue to select the file named "Header" that is inside each plotfile folder.
-      2. With Amrvis, ``amrvis3d plt00030``, for example.
-
+See `Getting Started <https://erf.readthedocs.io/en/latest/GettingStarted>`_ for instructions as to how to clone the ERF
+and AMReX codes, and for how to build and run an ERF example.  Mimimum requirements for system software are also given there.
[…]
