+++ Tests/CTestList.cmake
@@ -93,31 +93,31 @@ endfunction(add_test_u)
 # Regression tests
 #=============================================================================
 #add_test_r(Bubble_DensityCurrent             "Bubble/bubble" "plt00010")
-add_test_r(CouetteFlow                       "CouetteFlow/erf_couette_flow" "plt00050")
+add_test_r(CouetteFlow                       "RegTests/CouetteFlow/erf_couette_flow" "plt00050")
 add_test_r(DensityCurrent                    "DensityCurrent/density_current" "plt00010")
 add_test_r(DensityCurrent_detJ2              "DensityCurrent/density_current" "plt00010")
 add_test_r(DensityCurrent_detJ2_nosub        "DensityCurrent/density_current" "plt00020")
 add_test_r(DensityCurrent_detJ2_MT           "DensityCurrent/density_current" "plt00010")
-add_test_r(EkmanSpiral                       "EkmanSpiral_custom/ekman_spiral_custom" "plt00010")
-add_test_r(IsentropicVortexStationary       "IsentropicVortex/erf_isentropic_vortex" "plt00010")
-add_test_r(IsentropicVortexAdvecting        "IsentropicVortex/erf_isentropic_vortex" "plt00010")
+add_test_r(EkmanSpiral                       "RegTests/EkmanSpiral_custom/ekman_spiral_custom" "plt00010")
+add_test_r(IsentropicVortexStationary        "RegTests/IsentropicVortex/erf_isentropic_vortex" "plt00010")
+add_test_r(IsentropicVortexAdvecting         "RegTests/IsentropicVortex/erf_isentropic_vortex" "plt00010")
 add_test_r(MovingTerrain_nosub               "DevTests/MovingTerrain/moving_terrain"   "plt00020")
 add_test_r(MovingTerrain_sub                 "DevTests/MovingTerrain/moving_terrain"   "plt00010")
-add_test_r(PoiseuilleFlow                    "PoiseuilleFlow/erf_poiseuille_flow" "plt00010")
-add_test_r(RayleighDamping                  "ScalarAdvDiff/erf_scalar_advdiff" "plt00100")
-add_test_r(ScalarAdvectionUniformU          "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarAdvectionShearedU          "ScalarAdvDiff/erf_scalar_advdiff" "plt00080")
-add_test_r(ScalarAdvDiff_order2             "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarAdvDiff_order3             "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarAdvDiff_order4             "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarAdvDiff_order5             "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarAdvDiff_order6             "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarDiffusionGaussian          "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(ScalarDiffusionSine              "ScalarAdvDiff/erf_scalar_advdiff" "plt00020")
-add_test_r(TaylorGreenAdvecting             "TaylorGreenVortex/taylor_green" "plt00010")
-add_test_r(TaylorGreenAdvectingDiffusing    "TaylorGreenVortex/taylor_green" "plt00010")
[…]
