### 2023-06-27 965fbe5
Reorg exec (#1127)
* update cmake for CI

* Revert "update cmake for CI"

This reverts commit a74d7ea3aa56ab2a1a19262a6232d00f5792387e.

* Move test dirs into Exec/RegTests and update README

* update GettingStarted docs
