+++ Exec/CMakeLists.txt
@@ -6,19 +6,19 @@ if (ERF_ENABLE_MULTIBLOCK)
   add_subdirectory(MultiBlock)
 else ()
   add_subdirectory(ABL)
-  add_subdirectory(Bubble)
-  add_subdirectory(ChannelFlow)
-  add_subdirectory(CouetteFlow)
+  add_subdirectory(RegTests/Bubble)
+  add_subdirectory(RegTests/ChannelFlow)
+  add_subdirectory(RegTests/CouetteFlow)
   add_subdirectory(DensityCurrent)
   add_subdirectory(SuperCell)
-  add_subdirectory(EkmanSpiral_custom)
-  add_subdirectory(EkmanSpiral_ideal)
-  add_subdirectory(EkmanSpiral_input_sounding)
-  add_subdirectory(IsentropicVortex)
+  add_subdirectory(RegTests/EkmanSpiral_custom)
+  add_subdirectory(RegTests/EkmanSpiral_ideal)
+  add_subdirectory(RegTests/EkmanSpiral_input_sounding)
+  add_subdirectory(RegTests/IsentropicVortex)
   add_subdirectory(DevTests/MovingTerrain)
-  add_subdirectory(PoiseuilleFlow)
-  add_subdirectory(ScalarAdvDiff)
-  add_subdirectory(TaylorGreenVortex)
-  add_subdirectory(WitchOfAgnesi)
-  add_subdirectory(WPS_Test)
+  add_subdirectory(RegTests/PoiseuilleFlow)
+  add_subdirectory(RegTests/ScalarAdvDiff)
+  add_subdirectory(RegTests/TaylorGreenVortex)
+  add_subdirectory(RegTests/WitchOfAgnesi)
+  add_subdirectory(RegTests/WPS_Test)
 endif()
