### 2023-06-27 2db28d2
Add README's and fix path for compilation with gmake
