### 2023-06-27 faa4661
Tweak (#1129)
* update cmake for CI

* Revert "update cmake for CI"

This reverts commit a74d7ea3aa56ab2a1a19262a6232d00f5792387e.

* Add README's and fix path for compilation with gmake
