+++ Exec/CMakeLists.txt
@@ -7,7 +7,6 @@ if (ERF_ENABLE_MULTIBLOCK)
 else ()
   add_subdirectory(ABL)
   add_subdirectory(RegTests/Bubble)
-  add_subdirectory(RegTests/ChannelFlow)
   add_subdirectory(RegTests/CouetteFlow)
   add_subdirectory(DensityCurrent)
   add_subdirectory(SuperCell)
