### 2023-06-27 erf-model/ERF#1120 @asalmgren
Thanks @rafmudaf .  I believe we have now addressed all of the items above, including the suggestions as well as requirements :-)   For the example usage we have added a short section in the Building subsection about the paradigm for how the Exec directory is laid out and how to build with gmake vs cmake.  In addition, we have restructured Exec to make it more clear which are "science" tests vs "regression tests" vs tests of features under development.
