### 2023-06-27 #5202 @asalmgren
@pratikvn -- We have restructured the Exec directory to make it clear which types of tests are which, and added README files to each Exec/* subdirectory.  In addition, we have added more information under Getting Started / Building to describe how to build and run tests.  We have also removed duplicative test from the README.md file and instead we point to Getting Started.  We believe this should show the interested user how to build and run the code.
