### 2023-06-27 erf-model/ERF#1120 @rafmudaf
Great work @asalmgren. Thanks for addressing my comments, I'm happy with the state of the JOSS submission and will update the review accordingly. Feel free to close this issue.
