### 2023-06-30 #5202 @pratikvn
@asalmgren, I think that approach looks good. To make sure people understand that you do have examples, I would maybe add a small section under Getting Started/Testing and Verification that says that the tests themselves are the examples.
