### 2023-06-30 erf-model/ERF#1010 @pratikvn
As I mentioned in the other issue, adding a section that the tests can be used as examples would be helpful for the users. Otherwise, all points were addressed and I will close this issue now.
