### 2023-06-30 c95d8bc
initialize arrays to 1.234e20 to avoid uninitialized values
