### 2023-06-30 eb9663d
Fix wps restart (#1135)
* update cmake for CI

* Revert "update cmake for CI"

This reverts commit a74d7ea3aa56ab2a1a19262a6232d00f5792387e.

* initialize arrays to 1.234e20 to avoid uninitialized values

* fix tab
