### 2023-06-30 #5202 @kthyng
@asalmgren Next can you make an archive of your software at its present state (or the state of the JOSS review if it has changed a lot) at a place like Zenodo, and report the DOI back here. Also please make sure the title and author list match your JOSS paper (unless you have a reason you'd prefer not to).
