### 2023-06-30 b4a28ec
Quiet warnings (#1136)
* update cmake for CI

* Revert "update cmake for CI"

This reverts commit a74d7ea3aa56ab2a1a19262a6232d00f5792387e.

* quiet HIP compiler warnings
