### 2023-06-30 ab0f77d
Fix oops in bib (#1138)
* update cmake for CI

* Revert "update cmake for CI"

This reverts commit a74d7ea3aa56ab2a1a19262a6232d00f5792387e.

* fix bib file
