+++ paper/paper.bib
@@ -118,7 +118,7 @@
          publisher={Wiley Online Library}
 }
 
-@techreport{WRF:skamarock,
+@techreport{WRF:Skamarock,
 	doi = {10.5065/1dfh-6p97},
 	url = {https://dx.doi.org/10.5065/1dfh-6p97},
 	title = {A Description of the Advanced Research WRF Model Version 4.3 (No. NCAR/TN-556+STR)},
