### 2023-07-01 #5202 @asalmgren
> @asalmgren what version of your package would you like associated with the JOSS submission?

We have created the tag "JOSS_paper" to indicate the version of ERF associated with the JOSS submission.
