### 2023-07-01 #5202 @asalmgren
> @asalmgren The paper looks good except for the reference on line 54, page 2 which has a "?" after it — can you see what is wrong there?

Thank you for catching that!  We have fixed the paper.bib file.
