### 2023-07-01 #5202 @asalmgren
> @asalmgren Next can you make an archive of your software at its present state (or the state of the JOSS review if it has changed a lot) at a place like Zenodo, and report the DOI back here. Also please make sure the title and author list match your JOSS paper (unless you have a reason you'd prefer not to).

@kthyng The code has been archived at zenodo and the DOI is 10.5281/zenodo.8102984
